style: new illustrations
New illustrations depicting:
Cloud Connect
Cloud Connect Error
Prohibited search term
