+++ illustrations/src/illustrations/chunk-0.tsx
@@ -13,6 +13,10 @@ export const BluetoothPermissionsDark = (/*#__PURE__*/createFluentIllustration('
 export const BluetoothPermissionsLight = (/*#__PURE__*/createFluentIllustration('BluetoothPermissionsLight', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.414 144.8c34.267 0 62.048-27.801 62.048-62.215 0-34.415-27.903-62.215-62.048-62.215-34.267 0-62.048 27.8-62.048 62.215 0 34.414 27.78 62.215 62.048 62.215z" fill="#ECEBEA"/><path d="M173.423 41.802a5.02 5.02 0 005.018-5.021 5.02 5.02 0 00-5.018-5.022 5.019 5.019 0 00-5.017 5.022 5.019 5.019 0 005.017 5.021zM180.766 22.207a3.428 3.428 0 003.427-3.43 3.428 3.428 0 10-6.853 0 3.428 3.428 0 003.426 3.43zM56.67 41.68a3.428 3.428 0 003.427-3.43 3.428 3.428 0 10-6.853 0 3.428 3.428 0 003.426 3.43zM35.743 107.814a6.367 6.367 0 006.363-6.369 6.366 6.366 0 00-6.363-6.368 6.366 6.366 0 00-6.364 6.368 6.367 6.367 0 006.364 6.369z" fill="#E8DFD3"/><mask id="a" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="53" y="8" width="125" height="137"><path d="M169.171 8H63.846v40.251C57.345 58.085 53.56 69.885 53.56 82.585c0 34.414 27.8 62.215 62.093 62.215 34.291 0 62.092-27.801 62.092-62.215 0-11.52-3.128-22.298-8.573-31.541V8z" fill="#EBEBEB"/></mask><g filter="url(#filter0_d_3831_4590)" mask="url(#a)"><path d="M72.616 147.616a4.5 4.5 0 01-4.5-4.461l-1.064-124.88c0-3.032 2.48-5.376 5.511-5.376h86.528c3.031 0 5.511 2.481 5.511 5.377v123.963c0 3.033-2.48 5.377-5.511 5.377H72.616z" fill="url(#paint0_linear_3831_4590)"/></g><ellipse cx="115.552" cy="82.325" rx="38.778" ry="46.318" fill="url(#paint1_linear_3831_4590)"/><path fill-rule="evenodd" clip-rule="evenodd" d="M115.465 51.575a1.8 1.8 0 011.965.404l14.486 14.726a1.8 1.8 0 010 2.525l-12.351 12.556 12.351 12.556a1.8 1.8 0 010 2.525l-14.486 14.726a1.8 1.8 0 01-3.083-1.262v-23.24l-12.592 12.802a1.8 1.8 0 01-2.567-2.525l15.159-15.411v-.342L99.188 66.204a1.8 1.8 0 012.567-2.525l12.592 12.802v-23.24a1.8 1.8 0 011.118-1.666zm2.482 33.7l10.16 10.33-10.16 10.329v-20.66zm0-6.978v-20.66l10.16 10.33-10.16 10.33z" fill="#fff"/><defs><linearGradient id="paint0_linear_3831_4590" x1="115.859" y1="150.733" x2="115.859" y2="11.446" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_3831_4590" x1="114.833" y1="21.944" x2="116.643" y2="91.192" gradientUnits="userSpaceOnUse"><stop stop-color="#E1DAD0"/><stop offset="1" stop-color="#CCC1B1"/></linearGradient><filter id="filter0_d_3831_4590" x="56.252" y="8.399" width="119.15" height="156.318" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="6.3"/><feGaussianBlur stdDeviation="5.4"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.17 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3831_4590"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3831_4590" result="shape"/></filter></defs></svg>`));
 export const BluetoothPermissonsErrorDark = (/*#__PURE__*/createFluentIllustration('BluetoothPermissonsErrorDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M116.413 144.8c34.312 0 62.13-27.801 62.13-62.215s-27.94-62.215-62.13-62.215c-34.312 0-62.13 27.8-62.13 62.215 0 34.414 27.818 62.215 62.13 62.215z" fill="#242424"/><path d="M174.499 41.802a5.022 5.022 0 100-10.043 5.022 5.022 0 100 10.043zM181.852 22.207a3.43 3.43 0 10-3.431-3.43 3.43 3.43 0 003.431 3.43zM57.592 41.68a3.43 3.43 0 10-3.43-3.43 3.43 3.43 0 003.43 3.43zM36.636 107.814a6.37 6.37 0 006.373-6.369 6.37 6.37 0 00-6.373-6.368 6.37 6.37 0 00-6.372 6.368 6.37 6.37 0 006.372 6.369z" fill="#5C5C5C"/><mask id="a" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="54" y="8" width="125" height="137"><path d="M170.241 8H64.778v40.251c-6.51 9.834-10.301 21.634-10.301 34.334 0 34.414 27.837 62.215 62.174 62.215s62.175-27.801 62.175-62.215c0-11.52-3.133-22.298-8.585-31.541V8z" fill="#D9D9D9"/></mask><g mask="url(#a)"><path d="M69.091 147.616l-1.103-129.34c0-3.034 2.483-5.377 5.518-5.377h86.641c3.036 0 5.519 2.481 5.519 5.377v123.963c0 3.033-2.483 5.377-5.519 5.377H69.092z" fill="url(#paint0_linear_3869_1833)"/></g><ellipse cx="116.551" cy="82.325" rx="38.829" ry="46.318" fill="#3A3A3A"/><path d="M101.451 64.941l30.201 30.663-14.505 14.727v-57.09l14.505 14.726-30.201 30.663" stroke="#F5F2EC" stroke-width="3.6" stroke-linecap="round" stroke-linejoin="round"/><path d="M153.134 85.665l-22.396 35.652c-2.704 4.305.394 9.898 5.482 9.898h44.792c5.088 0 8.186-5.593 5.482-9.898l-22.396-35.652c-2.536-4.038-8.427-4.038-10.964 0z" fill="url(#paint1_linear_3869_1833)"/><path d="M158.274 98.482a2.736 2.736 0 00-2.732 2.889l.679 12.122a2.055 2.055 0 004.105 0l.679-12.122a2.736 2.736 0 00-2.731-2.89zm-.009 25.041c1.468 0 2.615-1.186 2.615-2.672 0-1.486-1.147-2.653-2.615-2.653-1.451 0-2.597 1.167-2.597 2.653s1.146 2.672 2.597 2.672z" fill="#242424"/><defs><linearGradient id="paint0_linear_3869_1833" x1="116.859" y1="150.733" x2="116.859" y2="11.446" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint1_linear_3869_1833" x1="136.505" y1="143.184" x2="180.669" y2="98.962" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient></defs></svg>`));
 export const BluetoothPermissonsErrorLight = (/*#__PURE__*/createFluentIllustration('BluetoothPermissonsErrorLight', `<svg viewBox="0 0 231 161" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M116.3 144.8c34.267 0 62.048-27.801 62.048-62.215 0-34.415-27.904-62.215-62.048-62.215-34.268 0-62.048 27.8-62.048 62.215 0 34.414 27.78 62.215 62.048 62.215z" fill="#ECEBEA"/><path d="M174.309 41.802a5.019 5.019 0 005.017-5.021 5.019 5.019 0 00-5.017-5.021 5.02 5.02 0 00-5.018 5.02 5.02 5.02 0 005.018 5.022zM181.652 22.207a3.429 3.429 0 000-6.859 3.428 3.428 0 00-3.427 3.43 3.428 3.428 0 003.427 3.429zM57.556 41.68a3.428 3.428 0 003.426-3.43 3.428 3.428 0 00-3.426-3.429 3.428 3.428 0 00-3.427 3.43 3.428 3.428 0 003.427 3.429zM36.628 107.814a6.367 6.367 0 006.364-6.369 6.366 6.366 0 00-6.364-6.368 6.366 6.366 0 00-6.364 6.368 6.367 6.367 0 006.364 6.369z" fill="#E8DFD3"/><mask id="a" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="54" y="8" width="125" height="137"><path d="M170.057 8H64.732v40.251c-6.502 9.834-10.288 21.634-10.288 34.334 0 34.414 27.801 62.215 62.093 62.215s62.093-27.801 62.093-62.215c0-11.52-3.129-22.298-8.573-31.541V8z" fill="#EBEBEB"/></mask><g filter="url(#filter0_d_3869_1817)" mask="url(#a)"><path d="M73.502 147.617a4.5 4.5 0 01-4.5-4.462L67.938 18.276c0-3.033 2.48-5.377 5.511-5.377h86.527c3.031 0 5.511 2.482 5.511 5.377v123.963c0 3.034-2.48 5.378-5.511 5.378H73.502z" fill="url(#paint0_linear_3869_1817)"/></g><ellipse cx="116.438" cy="82.325" rx="38.778" ry="46.318" fill="url(#paint1_linear_3869_1817)"/><path d="M101.357 64.941l30.161 30.663-14.485 14.727v-57.09l14.485 14.726-30.161 30.663" stroke="#fff" stroke-width="3.6" stroke-linecap="round" stroke-linejoin="round"/><g filter="url(#filter1_d_3869_1817)"><path d="M153.188 85.32l-22.828 36.388c-2.594 4.134.378 9.507 5.259 9.507h45.656c4.881 0 7.852-5.373 5.259-9.507l-22.829-36.387c-2.433-3.879-8.084-3.879-10.517 0z" fill="url(#paint2_linear_3869_1817)"/></g><path d="M158.105 98.482a2.732 2.732 0 00-2.728 2.885l.678 12.129a2.053 2.053 0 004.099 0l.679-12.129a2.733 2.733 0 00-2.728-2.885zm-.009 25.041c1.466 0 2.611-1.186 2.611-2.672 0-1.487-1.145-2.653-2.611-2.653-1.449 0-2.594 1.166-2.594 2.653 0 1.486 1.145 2.672 2.594 2.672z" fill="#fff"/><defs><linearGradient id="paint0_linear_3869_1817" x1="116.744" y1="150.733" x2="116.744" y2="11.446" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_3869_1817" x1="115.719" y1="21.944" x2="117.529" y2="91.192" gradientUnits="userSpaceOnUse"><stop stop-color="#E1DAD0"/><stop offset="1" stop-color="#CCC1B1"/></linearGradient><linearGradient id="paint2_linear_3869_1817" x1="185.188" y1="82.412" x2="131.173" y2="82.636" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><filter id="filter0_d_3869_1817" x="57.138" y="8.399" width="119.15" height="156.318" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="6.3"/><feGaussianBlur stdDeviation="5.4"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.17 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_1817"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_1817" result="shape"/></filter><filter id="filter1_d_3869_1817" x="109.601" y="72.512" width="97.692" height="88.403" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="9.9"/><feGaussianBlur stdDeviation="9.9"/><feColorMatrix values="0 0 0 0 0.482733 0 0 0 0 0.482733 0 0 0 0 0.482733 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_1817"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_1817" result="shape"/></filter></defs></svg>`));
+export const CloudConnectDark = (/*#__PURE__*/createFluentIllustration('CloudConnectDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M114.977 135c30.306 0 55-24.66 55-54.949-.102-30.39-24.694-55.051-55-55.051-30.409 0-55 24.66-55 54.949 0 30.391 24.591 55.051 55 55.051z" fill="#242424"/><g filter="url(#filter0_d_4963_604)"><path d="M173.383 86.692c0 11.707-9.468 21.174-21.277 21.174H81.049c-14.15.306-25.45-10.994-25.45-24.636 0-13.743 11.401-25.145 25.755-24.432 12.318-38.48 68.716-33.085 73.501 6.719 10.587 1.323 18.528 10.282 18.528 21.175z" fill="#424242"/></g><path d="M117.392 108.07c20.87 0 37.87-16.9 37.87-37.769s-17-37.768-37.87-37.768c-20.869 0-37.87 16.9-37.87 37.768 0 20.87 16.9 37.769 37.87 37.769z" fill="url(#paint0_linear_4963_604)"/><path d="M101.209 84.859c3.971-2.545 8.654-3.97 13.642-3.97 13.641 0 24.737 10.994 24.737 24.636 0 13.641-11.096 24.635-24.737 24.635-13.642 0-24.738-10.994-24.738-24.635-.204-8.654 4.276-16.187 11.096-20.666z" fill="url(#paint1_linear_4963_604)"/><path d="M152.107 107.866c11.707 0 21.276-9.467 21.276-21.174 0-11.708-9.569-21.175-21.276-21.175-11.708 0-21.277 9.467-21.277 21.175 0 11.707 9.569 21.174 21.277 21.174z" fill="url(#paint2_linear_4963_604)"/><path d="M107.942 104.016a1.394 1.394 0 000-1.983c-.552-.551-1.215-.441-2.099-.441 2.541-6.39 10.826-7.822 15.134-2.644 1.657 1.763 4.308-.55 2.762-2.314-4.529-5.288-12.594-5.949-17.896-1.432-1.988 1.653-3.314 3.856-3.977 6.39-1.767 0-2.099 0-2.54.441a1.394 1.394 0 000 1.983c5.081 4.958 3.755 4.958 8.616 0zM129.594 107.211l-3.093-3.195a1.405 1.405 0 00-1.989 0l-3.203 3.195a1.394 1.394 0 000 1.984c.552.55 1.215.44 2.099.44-2.541 6.5-11.047 7.823-15.355 2.314-.663-.882-1.768-.992-2.652-.331-2.761 2.314 3.756 7.382 9.501 7.382 5.634 0 10.826-3.856 12.151-9.475 1.768 0 2.099 0 2.541-.441.552-.33.552-1.212 0-1.873z" fill="#242424"/><path d="M171.041 60.63a4.174 4.174 0 100-8.348 4.174 4.174 0 000 8.348zM177.149 44.342a2.85 2.85 0 10.001-5.7 2.85 2.85 0 00-.001 5.7zM61.707 53.708a2.85 2.85 0 100-5.7 2.85 2.85 0 000 5.7zM55.294 116.417a5.293 5.293 0 100-10.587 5.293 5.293 0 000 10.587z" fill="#5C5C5C"/><defs><linearGradient id="paint0_linear_4963_604" x1="88.752" y1="48.222" x2="101.434" y2="58.024" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint1_linear_4963_604" x1="95.904" y1="142.301" x2="140.078" y2="104.939" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint2_linear_4963_604" x1="128.566" y1="60.958" x2="147.427" y2="81.587" gradientUnits="userSpaceOnUse"><stop stop-color="#E0D6C0"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><filter id="filter0_d_4963_604" x="44.484" y="30.916" width="140.014" height="97.477" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="9.405"/><feGaussianBlur stdDeviation="5.558"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4963_604"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4963_604" result="shape"/></filter></defs></svg>`));
+export const CloudConnectErrorDark = (/*#__PURE__*/createFluentIllustration('CloudConnectErrorDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M114.977 135c30.306 0 55-24.66 55-54.949-.102-30.39-24.694-55.051-55-55.051-30.409 0-55 24.66-55 54.949 0 30.391 24.591 55.051 55 55.051z" fill="#242424"/><g filter="url(#filter0_d_4968_460)"><path d="M173.383 86.692c0 11.707-9.468 21.174-21.277 21.174H81.049c-14.15.306-25.45-10.994-25.45-24.636 0-13.743 11.401-25.145 25.755-24.432 12.318-38.48 68.716-33.085 73.501 6.719 10.587 1.323 18.528 10.282 18.528 21.175z" fill="#424242"/></g><path d="M117.392 108.07c20.87 0 37.87-16.9 37.87-37.769s-17-37.768-37.87-37.768c-20.869 0-37.87 16.9-37.87 37.768 0 20.87 16.9 37.769 37.87 37.769z" fill="url(#paint0_linear_4968_460)"/><g filter="url(#filter1_d_4968_460)"><path d="M90.828 116.093l19.767-31.852c2.152-2.988 6.657-2.988 8.808 0l19.767 31.852c2.085 2.856 0 6.907-3.563 6.907H94.459c-3.63-.066-5.715-4.051-3.63-6.907z" fill="url(#paint1_linear_4968_460)"/></g><path d="M115.079 93.01c-1.761 0-3.131 1.596-3 3.47l.978 11.237c.065 1.11.978 2.012 2.022 2.012 1.043 0 1.956-.902 2.021-2.012l.978-11.238c.131-1.942-1.239-3.468-2.999-3.468zM115.08 111.81c-1.304 0-2.348 1.109-2.348 2.497 0 1.387 1.044 2.497 2.348 2.497 1.304 0 2.347-1.11 2.347-2.497 0-1.388-1.043-2.497-2.347-2.497z" fill="#252525"/><path d="M152.107 107.866c11.707 0 21.276-9.467 21.276-21.174 0-11.708-9.569-21.175-21.276-21.175-11.708 0-21.277 9.467-21.277 21.175 0 11.707 9.569 21.174 21.277 21.174z" fill="url(#paint2_linear_4968_460)"/><path d="M171.041 60.63a4.174 4.174 0 100-8.348 4.174 4.174 0 000 8.348zM177.149 44.342a2.85 2.85 0 10.001-5.7 2.85 2.85 0 00-.001 5.7zM61.707 53.708a2.85 2.85 0 100-5.7 2.85 2.85 0 000 5.7zM55.294 116.417a5.293 5.293 0 100-10.587 5.293 5.293 0 000 10.587z" fill="#5C5C5C"/><defs><linearGradient id="paint0_linear_4968_460" x1="88.752" y1="48.222" x2="101.434" y2="58.024" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint1_linear_4968_460" x1="115" y1="82" x2="115" y2="123" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint2_linear_4968_460" x1="128.566" y1="60.958" x2="147.427" y2="81.587" gradientUnits="userSpaceOnUse"><stop stop-color="#E0D6C0"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><filter id="filter0_d_4968_460" x="44.484" y="30.916" width="140.014" height="97.477" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="9.405"/><feGaussianBlur stdDeviation="5.558"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4968_460"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4968_460" result="shape"/></filter><filter id="filter1_d_4968_460" x="74" y="70" width="82" height="73" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4"/><feGaussianBlur stdDeviation="8"/><feColorMatrix values="0 0 0 0 0.482733 0 0 0 0 0.482733 0 0 0 0 0.482733 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4968_460"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4968_460" result="shape"/></filter></defs></svg>`));
+export const CloudConnectErrorLight = (/*#__PURE__*/createFluentIllustration('CloudConnectErrorLight', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.977 135c30.306 0 55-24.66 55-54.949-.102-30.39-24.694-55.051-55-55.051-30.409 0-55 24.66-55 54.949 0 30.391 24.591 55.051 55 55.051z" fill="#EBEBEB"/><g filter="url(#filter0_d_4968_442)"><path d="M174.383 86.692c0 11.707-9.468 21.174-21.277 21.174H82.049c-14.15.306-25.45-10.994-25.45-24.636 0-13.743 11.401-25.145 25.755-24.432 12.318-38.48 68.716-33.085 73.501 6.719 10.587 1.323 18.528 10.282 18.528 21.175z" fill="url(#paint0_linear_4968_442)"/></g><path d="M153.107 107.866c11.707 0 21.276-9.467 21.276-21.174 0-11.708-9.569-21.175-21.276-21.175-11.708 0-21.277 9.467-21.277 21.175 0 11.707 9.569 21.174 21.277 21.174z" fill="url(#paint1_linear_4968_442)"/><path d="M118.392 108.07c20.87 0 37.87-16.9 37.87-37.769s-17-37.768-37.87-37.768c-20.869 0-37.87 16.9-37.87 37.768 0 20.87 16.9 37.769 37.87 37.769z" fill="url(#paint2_linear_4968_442)"/><g filter="url(#filter1_d_4968_442)"><path d="M91.828 116.093l19.767-31.852c2.152-2.988 6.657-2.988 8.808 0l19.767 31.852c2.085 2.856 0 6.907-3.563 6.907H95.459c-3.63-.066-5.715-4.051-3.63-6.907z" fill="url(#paint3_linear_4968_442)"/></g><path d="M116.079 93.01c-1.761 0-3.131 1.596-3 3.47l.978 11.237c.065 1.11.978 2.012 2.022 2.012 1.043 0 1.956-.902 2.021-2.012l.978-11.238c.131-1.942-1.239-3.468-2.999-3.468zM116.08 111.81c-1.304 0-2.348 1.11-2.348 2.497 0 1.387 1.044 2.497 2.348 2.497 1.304 0 2.347-1.11 2.347-2.497 0-1.387-1.043-2.497-2.347-2.497z" fill="#fff"/><path d="M172.041 60.63a4.174 4.174 0 100-8.348 4.174 4.174 0 000 8.348zM178.149 44.342a2.85 2.85 0 10.001-5.7 2.85 2.85 0 00-.001 5.7zM62.707 53.708a2.85 2.85 0 100-5.7 2.85 2.85 0 000 5.7zM56.294 116.417a5.293 5.293 0 100-10.587 5.293 5.293 0 000 10.587z" fill="#EDE5DA"/><defs><linearGradient id="paint0_linear_4968_442" x1="115.491" y1="63.72" x2="115.452" y2="108.683" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#EFEBE3"/></linearGradient><linearGradient id="paint1_linear_4968_442" x1="129.566" y1="60.958" x2="148.427" y2="81.587" gradientUnits="userSpaceOnUse"><stop stop-color="#E0D6C0"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint2_linear_4968_442" x1="76.976" y1="52.971" x2="106.704" y2="64.604" gradientUnits="userSpaceOnUse"><stop stop-color="#F0EBE1"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint3_linear_4968_442" x1="139.017" y1="82" x2="92.525" y2="82.198" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><filter id="filter0_d_4968_442" x="37.789" y="23.22" width="155.404" height="112.867" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="9.405"/><feGaussianBlur stdDeviation="9.405"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4968_442"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4968_442" result="shape"/></filter><filter id="filter1_d_4968_442" x="75" y="70" width="82" height="73" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4"/><feGaussianBlur stdDeviation="8"/><feColorMatrix values="0 0 0 0 0.482733 0 0 0 0 0.482733 0 0 0 0 0.482733 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4968_442"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4968_442" result="shape"/></filter></defs></svg>`));
+export const CloudConnectLight = (/*#__PURE__*/createFluentIllustration('CloudConnectLight', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.977 135c30.306 0 55-24.66 55-54.949-.102-30.39-24.694-55.051-55-55.051-30.409 0-55 24.66-55 54.949 0 30.391 24.591 55.051 55 55.051z" fill="#EBEBEB"/><g filter="url(#filter0_d_4963_551)"><path d="M174.383 86.692c0 11.707-9.468 21.174-21.277 21.174H82.049c-14.15.306-25.45-10.994-25.45-24.636 0-13.743 11.401-25.145 25.755-24.432 12.318-38.48 68.716-33.085 73.501 6.719 10.587 1.323 18.528 10.282 18.528 21.175z" fill="url(#paint0_linear_4963_551)"/></g><path d="M153.107 107.866c11.707 0 21.276-9.467 21.276-21.174 0-11.708-9.569-21.175-21.276-21.175-11.708 0-21.277 9.467-21.277 21.175 0 11.707 9.569 21.174 21.277 21.174z" fill="url(#paint1_linear_4963_551)"/><path d="M118.392 108.07c20.87 0 37.87-16.9 37.87-37.769s-17-37.768-37.87-37.768c-20.869 0-37.87 16.9-37.87 37.768 0 20.87 16.9 37.769 37.87 37.769z" fill="url(#paint2_linear_4963_551)"/><g filter="url(#filter1_d_4963_551)"><path d="M102.209 84.859c3.971-2.545 8.654-3.97 13.642-3.97 13.641 0 24.737 10.994 24.737 24.636 0 13.641-11.096 24.635-24.737 24.635-13.642 0-24.738-10.994-24.738-24.635-.204-8.654 4.276-16.187 11.096-20.666z" fill="url(#paint3_linear_4963_551)"/></g><path d="M108.942 104.016a1.394 1.394 0 000-1.983c-.552-.551-1.215-.441-2.099-.441 2.541-6.39 10.826-7.822 15.134-2.644 1.657 1.763 4.308-.55 2.762-2.314-4.529-5.288-12.594-5.949-17.896-1.432-1.988 1.653-3.314 3.856-3.977 6.39-1.767 0-2.099 0-2.541.441a1.394 1.394 0 000 1.983c5.082 4.958 3.756 4.958 8.617 0zM130.594 107.211l-3.093-3.195a1.405 1.405 0 00-1.989 0l-3.203 3.195a1.394 1.394 0 000 1.984c.552.55 1.215.44 2.099.44-2.541 6.5-11.047 7.823-15.355 2.314-.663-.882-1.768-.992-2.652-.331-2.761 2.314 3.756 7.382 9.501 7.382 5.634 0 10.826-3.856 12.151-9.475 1.768 0 2.099 0 2.541-.441.552-.33.552-1.212 0-1.873z" fill="#fff"/><path d="M172.041 60.63a4.174 4.174 0 100-8.348 4.174 4.174 0 000 8.348zM178.149 44.342a2.85 2.85 0 10.001-5.7 2.85 2.85 0 00-.001 5.7zM62.707 53.708a2.85 2.85 0 100-5.7 2.85 2.85 0 000 5.7zM56.294 116.417a5.293 5.293 0 100-10.587 5.293 5.293 0 000 10.587z" fill="#EDE5DA"/><defs><linearGradient id="paint0_linear_4963_551" x1="115.491" y1="63.72" x2="115.452" y2="108.683" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#EFEBE3"/></linearGradient><linearGradient id="paint1_linear_4963_551" x1="129.566" y1="60.958" x2="148.427" y2="81.587" gradientUnits="userSpaceOnUse"><stop stop-color="#E0D6C0"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint2_linear_4963_551" x1="76.976" y1="52.971" x2="106.704" y2="64.604" gradientUnits="userSpaceOnUse"><stop stop-color="#F0EBE1"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint3_linear_4963_551" x1="138.626" y1="80.889" x2="92.615" y2="81.05" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><filter id="filter0_d_4963_551" x="37.789" y="23.22" width="155.404" height="112.867" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="9.405"/><feGaussianBlur stdDeviation="9.405"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4963_551"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4963_551" result="shape"/></filter><filter id="filter1_d_4963_551" x="71.306" y="70.989" width="89.082" height="88.872" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="9.9"/><feGaussianBlur stdDeviation="9.9"/><feColorMatrix values="0 0 0 0 0.482733 0 0 0 0 0.482733 0 0 0 0 0.482733 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4963_551"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4963_551" result="shape"/></filter></defs></svg>`));
 export const CloudLocationDark = (/*#__PURE__*/createFluentIllustration('CloudLocationDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M117.941 148.8c36.641 0 66.497-29.732 66.497-66.25-.124-36.64-29.856-66.373-66.497-66.373-36.764 0-66.496 29.733-66.496 66.25 0 36.641 29.732 66.373 66.496 66.373z" fill="#242424"/><path d="M186.042 59.357a5.058 5.058 0 100-10.117 5.058 5.058 0 000 10.117zM193.444 39.617a3.454 3.454 0 100-6.907 3.454 3.454 0 000 6.907zM45.77 126.963a6.415 6.415 0 100-12.83 6.415 6.415 0 000 12.83z" fill="#5C5C5C"/><g filter="url(#filter0_d_3840_836)"><path d="M97.066 99.464c0 5.736-4.639 10.375-10.425 10.375H51.825c-6.933.149-12.47-5.387-12.47-12.071 0-6.734 5.587-12.32 12.62-11.971 6.035-18.855 33.669-16.211 36.013 3.292 5.187.648 9.078 5.037 9.078 10.375z" fill="#424242"/></g><path d="M86.641 109.839c5.736 0 10.425-4.639 10.425-10.375s-4.689-10.375-10.425-10.375c-5.736 0-10.425 4.639-10.425 10.375s4.689 10.375 10.425 10.375z" fill="url(#paint0_linear_3840_836)"/><path d="M69.632 109.938c10.226 0 18.556-8.28 18.556-18.505s-8.33-18.506-18.556-18.506c-10.225 0-18.555 8.28-18.555 18.506 0 10.225 8.28 18.505 18.555 18.505z" fill="url(#paint1_linear_3840_836)"/><g filter="url(#filter1_d_3840_836)"><path d="M113.557 48.607c0 7.914-6.4 14.313-14.382 14.313h-48.03c-9.565.207-17.203-7.431-17.203-16.652 0-9.29 7.707-16.997 17.41-16.515 8.325-26.01 46.447-22.364 49.681 4.542 7.156.894 12.524 6.95 12.524 14.312z" fill="#424242"/></g><path d="M75.71 63.058c14.107 0 25.598-11.423 25.598-25.529S89.817 12 75.71 12c-14.106 0-25.597 11.423-25.597 25.529S61.534 63.058 75.71 63.058z" fill="url(#paint2_linear_3840_836)"/><path d="M99.175 62.92c7.913 0 14.382-6.4 14.382-14.313s-6.469-14.312-14.382-14.312c-7.913 0-14.382 6.4-14.382 14.312 0 7.914 6.469 14.313 14.382 14.313z" fill="url(#paint3_linear_3840_836)"/><path d="M94.94 70.225c0-11.557 9.415-20.953 20.995-20.953 11.579 0 20.993 9.396 20.993 20.953 0 11.556-9.414 20.952-20.993 20.952-11.58 0-20.994-9.302-20.994-20.952zm20.995-39.65c-22.783 0-41.33 18.51-41.33 41.247 0 16.536 21.842 47.824 33.798 59.475 2.636 2.536 6.214 3.57 9.603 2.912.282 0 .47-.094.753-.188.094 0 .188 0 .282-.094 1.601-.469 3.201-1.409 4.519-2.63 9.979-9.772 26.831-35.328 32.103-50.737.753-2.255 1.318-4.51 1.506-6.671.094-.752.094-1.41.094-2.161 0-22.644-18.452-41.153-41.328-41.153z" fill="url(#paint4_linear_3840_836)"/><path d="M115.934 144.075c-17.416 0-35.115-4.604-35.115-13.342 0-2.349 1.318-5.731 7.626-8.644 1.224-.564 2.542 0 3.106 1.128.565 1.221 0 2.536-1.13 3.1-3.67 1.691-4.895 3.383-4.895 4.322 0 3.571 11.862 8.644 30.408 8.644 18.547 0 30.409-5.167 30.409-8.644 0-1.221-1.789-3.194-6.684-5.073-1.224-.47-1.789-1.786-1.318-3.007.47-1.222 1.788-1.785 3.012-1.316 8.002 3.101 9.697 6.859 9.697 9.49 0 8.738-17.699 13.342-35.116 13.342z" fill="#F5F2EC"/><g filter="url(#filter2_d_3840_836)"><path d="M191.1 103.658c0 5.736-4.639 10.375-10.425 10.375h-34.816c-6.934.15-12.47-5.387-12.47-12.071 0-6.733 5.586-12.32 12.619-11.97 6.036-18.855 33.669-16.212 36.014 3.291 5.187.649 9.078 5.038 9.078 10.375z" fill="#424242"/></g><path d="M163.666 114.133c10.225 0 18.555-8.28 18.555-18.505 0-10.226-8.33-18.506-18.555-18.506-10.226 0-18.556 8.28-18.556 18.505 0 10.226 8.28 18.506 18.556 18.506z" fill="url(#paint5_linear_3840_836)"/><path d="M180.675 114.033c5.736 0 10.425-4.639 10.425-10.375s-4.689-10.375-10.425-10.375c-5.736 0-10.425 4.639-10.425 10.375s4.689 10.375 10.425 10.375z" fill="url(#paint6_linear_3840_836)"/><defs><linearGradient id="paint0_linear_3840_836" x1="78.757" y1="93.399" x2="82.243" y2="96.098" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint1_linear_3840_836" x1="55.599" y1="80.615" x2="61.813" y2="85.417" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint2_linear_3840_836" x1="56.351" y1="22.605" x2="64.923" y2="29.23" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint3_linear_3840_836" x1="88.299" y1="40.24" x2="93.107" y2="43.965" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint4_linear_3840_836" x1="115.935" y1="30.575" x2="115.935" y2="134.403" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint5_linear_3840_836" x1="149.633" y1="84.809" x2="155.846" y2="89.612" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint6_linear_3840_836" x1="172.791" y1="97.593" x2="176.276" y2="100.293" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><filter id="filter0_d_3840_836" x="27.655" y="71.173" width="81.111" height="60.269" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="9.9"/><feGaussianBlur stdDeviation="5.85"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_836"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_836" result="shape"/></filter><filter id="filter1_d_3840_836" x="22.242" y="10.262" width="103.015" height="74.262" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="9.9"/><feGaussianBlur stdDeviation="5.85"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_836"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_836" result="shape"/></filter><filter id="filter2_d_3840_836" x="121.689" y="75.367" width="81.111" height="60.269" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="9.9"/><feGaussianBlur stdDeviation="5.85"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_836"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_836" result="shape"/></filter></defs></svg>`));
 export const CloudLocationLight = (/*#__PURE__*/createFluentIllustration('CloudLocationLight', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M117.942 148.8c36.641 0 66.496-29.732 66.496-66.25-.123-36.64-29.855-66.373-66.496-66.373-36.764 0-66.496 29.733-66.496 66.25 0 36.641 29.732 66.373 66.496 66.373z" fill="#EBEBEB"/><path d="M186.042 59.357a5.058 5.058 0 100-10.116 5.058 5.058 0 000 10.116zM193.444 39.618a3.454 3.454 0 100-6.908 3.454 3.454 0 000 6.908zM45.77 126.963a6.415 6.415 0 100-12.83 6.415 6.415 0 000 12.83z" fill="#E8DFD3"/><g filter="url(#filter0_d_3840_818)"><path d="M97.066 99.464c0 5.736-4.639 10.375-10.425 10.375H51.825c-6.933.149-12.47-5.387-12.47-12.071 0-6.734 5.587-12.32 12.62-11.971 6.035-18.855 33.669-16.211 36.013 3.292 5.187.648 9.078 5.038 9.078 10.375z" fill="url(#paint0_linear_3840_818)"/></g><path d="M86.641 109.839c5.736 0 10.425-4.639 10.425-10.375s-4.689-10.375-10.425-10.375c-5.736 0-10.425 4.639-10.425 10.375s4.689 10.375 10.425 10.375z" fill="url(#paint1_linear_3840_818)"/><path d="M69.632 109.939c10.226 0 18.555-8.281 18.555-18.506s-8.33-18.505-18.555-18.505-18.555 8.28-18.555 18.505 8.28 18.506 18.555 18.506z" fill="url(#paint2_linear_3840_818)"/><g filter="url(#filter1_d_3840_818)"><path d="M113.557 48.608c0 7.913-6.4 14.313-14.382 14.313h-48.03c-9.565.206-17.203-7.432-17.203-16.653 0-9.29 7.707-16.996 17.41-16.515 8.325-26.01 46.447-22.363 49.681 4.542 7.156.895 12.524 6.95 12.524 14.313z" fill="url(#paint3_linear_3840_818)"/></g><path d="M99.175 62.92c7.913 0 14.382-6.399 14.382-14.312 0-7.913-6.469-14.313-14.382-14.313-7.913 0-14.382 6.4-14.382 14.313S91.262 62.92 99.175 62.92z" fill="url(#paint4_linear_3840_818)"/><path d="M75.71 63.058c14.106 0 25.598-11.423 25.598-25.529S89.816 12 75.71 12c-14.106 0-25.598 11.423-25.598 25.529S61.535 63.058 75.71 63.058z" fill="url(#paint5_linear_3840_818)"/><path d="M94.941 70.225c0-11.557 9.414-20.953 20.994-20.953s20.994 9.396 20.994 20.953c0 11.556-9.414 20.952-20.994 20.952s-20.994-9.302-20.994-20.952zm20.994-39.65c-22.783 0-41.329 18.51-41.329 41.247 0 16.536 21.841 47.824 33.798 59.475 2.636 2.537 6.213 3.57 9.602 2.912.283 0 .471-.094.753-.188.094 0 .189 0 .283-.094 1.6-.469 3.201-1.409 4.519-2.63 9.979-9.772 26.83-35.328 32.102-50.737.754-2.255 1.318-4.51 1.507-6.671.094-.752.094-1.41.094-2.161 0-22.644-18.452-41.153-41.329-41.153z" fill="url(#paint6_linear_3840_818)"/><path d="M115.934 144.075c-17.416 0-35.115-4.604-35.115-13.342 0-2.349 1.318-5.731 7.626-8.644 1.224-.564 2.542 0 3.106 1.128.565 1.221 0 2.536-1.13 3.1-3.67 1.691-4.895 3.383-4.895 4.322 0 3.571 11.862 8.644 30.408 8.644 18.547 0 30.409-5.167 30.409-8.644 0-1.221-1.789-3.194-6.684-5.073-1.224-.47-1.789-1.786-1.318-3.007.47-1.222 1.788-1.785 3.012-1.316 8.002 3.101 9.697 6.859 9.697 9.49 0 8.738-17.699 13.342-35.116 13.342z" fill="#D9D2C7"/><g filter="url(#filter2_d_3840_818)"><path d="M191.1 103.658c0 5.737-4.639 10.375-10.425 10.375h-34.816c-6.934.15-12.47-5.387-12.47-12.071 0-6.733 5.586-12.32 12.619-11.97 6.036-18.855 33.669-16.212 36.014 3.291 5.187.649 9.078 5.038 9.078 10.375z" fill="url(#paint7_linear_3840_818)"/></g><path d="M180.675 114.034c5.736 0 10.425-4.639 10.425-10.376 0-5.736-4.689-10.375-10.425-10.375-5.736 0-10.425 4.64-10.425 10.375 0 5.737 4.689 10.376 10.425 10.376z" fill="url(#paint8_linear_3840_818)"/><path d="M163.666 114.133c10.225 0 18.555-8.28 18.555-18.505 0-10.226-8.33-18.506-18.555-18.506s-18.555 8.28-18.555 18.506c0 10.225 8.28 18.505 18.555 18.505z" fill="url(#paint9_linear_3840_818)"/><defs><linearGradient id="paint0_linear_3840_818" x1="68.21" y1="88.209" x2="68.192" y2="110.239" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#EFEBE3"/></linearGradient><linearGradient id="paint1_linear_3840_818" x1="75.107" y1="86.855" x2="84.349" y2="96.963" gradientUnits="userSpaceOnUse"><stop stop-color="#E0D6C0"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint2_linear_3840_818" x1="49.339" y1="82.941" x2="63.905" y2="88.641" gradientUnits="userSpaceOnUse"><stop stop-color="#F0EBE1"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint3_linear_3840_818" x1="73.749" y1="33.081" x2="73.723" y2="63.473" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#EFEBE3"/></linearGradient><linearGradient id="paint4_linear_3840_818" x1="83.263" y1="31.213" x2="96.012" y2="45.157" gradientUnits="userSpaceOnUse"><stop stop-color="#E0D6C0"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint5_linear_3840_818" x1="47.715" y1="25.814" x2="67.809" y2="33.678" gradientUnits="userSpaceOnUse"><stop stop-color="#F0EBE1"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint6_linear_3840_818" x1="115.908" y1="28.173" x2="115.908" y2="135.522" gradientUnits="userSpaceOnUse"><stop stop-color="#F4EADD"/><stop offset=".996" stop-color="#D7CDC1"/></linearGradient><linearGradient id="paint7_linear_3840_818" x1="162.244" y1="92.403" x2="162.225" y2="114.434" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#EFEBE3"/></linearGradient><linearGradient id="paint8_linear_3840_818" x1="169.141" y1="91.05" x2="178.382" y2="101.157" gradientUnits="userSpaceOnUse"><stop stop-color="#E0D6C0"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint9_linear_3840_818" x1="143.373" y1="87.136" x2="157.939" y2="92.836" gradientUnits="userSpaceOnUse"><stop stop-color="#F0EBE1"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><filter id="filter0_d_3840_818" x="30.355" y="67.573" width="75.711" height="54.869" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="3.6"/><feGaussianBlur stdDeviation="4.5"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_818"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_818" result="shape"/></filter><filter id="filter1_d_3840_818" x="26.742" y="12.062" width="94.015" height="65.262" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="7.2"/><feGaussianBlur stdDeviation="3.6"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_818"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_818" result="shape"/></filter><filter id="filter2_d_3840_818" x="124.389" y="71.767" width="75.711" height="54.869" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="3.6"/><feGaussianBlur stdDeviation="4.5"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_818"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_818" result="shape"/></filter></defs></svg>`));
 export const CodeErrorDark = (/*#__PURE__*/createFluentIllustration('CodeErrorDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M113.432 147.8c38.957 0 70.538-30.624 70.538-68.4 0-37.776-31.581-68.4-70.538-68.4-38.957 0-70.538 30.624-70.538 68.4 0 37.776 31.581 68.4 70.538 68.4z" fill="#242424"/><g filter="url(#filter0_d_3871_353928)"><path d="M178.142 38.584v81.059c0 2.933-2.491 5.348-5.515 5.348H54.971c-3.025 0-5.515-2.415-5.515-5.348v-81.06c0-2.932 2.49-5.347 5.515-5.347h117.656c3.024 0 5.515 2.415 5.515 5.348z" fill="#5C5C5C"/></g><path fill="#4D4D4D" d="M56.205 76.664h116.006v22.982H56.205z"/><path d="M178.142 38.584v3.881H49.456v-3.881c0-2.933 2.49-5.348 5.515-5.348h117.656c3.024 0 5.515 2.415 5.515 5.348z" fill="#3A3A3A"/><path d="M55.237 39.533c.934 0 1.69-.734 1.69-1.64 0-.904-.756-1.638-1.69-1.638-.933 0-1.69.734-1.69 1.639s.757 1.639 1.69 1.639z" fill="#D1D1D1"/><path opacity=".6" d="M60.485 39.533c.934 0 1.69-.734 1.69-1.64 0-.904-.756-1.638-1.69-1.638-.933 0-1.69.734-1.69 1.639s.757 1.639 1.69 1.639z" fill="#5C5C5C"/><path opacity=".4" d="M65.823 39.533c.933 0 1.69-.734 1.69-1.64 0-.904-.757-1.638-1.69-1.638-.934 0-1.69.734-1.69 1.639s.756 1.639 1.69 1.639z" fill="#5C5C5C"/><path d="M82.346 53.317H63.63c-1.17 0-2.145-1.178-2.145-2.121 0-1.178.975-2.121 2.145-2.121h18.716c1.17 0 2.145 1.178 2.145 2.12 0 1.18-.975 2.122-2.145 2.122z" fill="#707070"/><path d="M77.285 85.199H63.579c-1.142 0-2.093-1.179-2.093-2.121 0-1.179.951-2.121 2.093-2.121h13.706c1.142 0 2.094 1.178 2.094 2.12-.19.943-1.142 2.122-2.094 2.122zM165.364 85.199h-26.271c-1.201 0-2.202-1.179-2.202-2.121 0-1.179 1.001-2.121 2.202-2.121h26.271c1.201 0 2.201 1.178 2.201 2.12-.2.943-1.2 2.122-2.201 2.122z" fill="#D1D1D1"/><path d="M172.121 85.53l-22.724 36.223c-2.744 4.373.4 10.056 5.562 10.056h45.45c5.162 0 8.305-5.683 5.562-10.056L183.246 85.53c-2.574-4.102-8.551-4.102-11.125 0z" fill="url(#paint0_linear_3871_353928)"/><path d="M177.825 97.213a2.775 2.775 0 00-2.771 2.931l.689 12.323a2.086 2.086 0 004.164 0l.69-12.323a2.776 2.776 0 00-2.772-2.931zm-.009 25.441c1.49 0 2.653-1.204 2.653-2.714s-1.163-2.696-2.653-2.696c-1.472 0-2.635 1.186-2.635 2.696 0 1.51 1.163 2.714 2.635 2.714z" fill="#3A3A3A"/><path d="M128.278 85.199H87.993c-1.213 0-2.224-1.179-2.224-2.121 0-1.179 1.011-2.121 2.224-2.121h40.285c1.213 0 2.224 1.178 2.224 2.12-.202.943-1.213 2.122-2.224 2.122zM96.304 94.743H71.399c-1.225 0-2.245-1.178-2.245-2.121 0-1.178 1.02-2.121 2.245-2.121h24.905c1.225 0 2.245 1.178 2.245 2.12-.204 1.18-1.225 2.122-2.245 2.122zM119.392 94.743h-13.566c-1.18 0-2.164-1.178-2.164-2.121 0-1.178.984-2.121 2.164-2.121h13.566c1.18 0 2.163 1.178 2.163 2.12-.196 1.18-1.18 2.122-2.163 2.122z" fill="#D1D1D1"/><path d="M92.473 107.536H63.728c-1.223 0-2.242-1.178-2.242-2.121 0-1.178 1.019-2.121 2.242-2.121h28.745c1.223 0 2.243 1.179 2.243 2.121 0 1.179-.816 2.121-2.243 2.121zM113.007 107.536h-11.02c-1.178 0-2.159-1.178-2.159-2.121 0-1.178.981-2.121 2.159-2.121h11.02c1.177 0 2.158 1.179 2.158 2.121 0 1.179-.785 2.121-2.158 2.121zM75.872 117.08H63.714c-1.215 0-2.228-1.178-2.228-2.12 0-1.179 1.013-2.121 2.228-2.121h12.158c1.215 0 2.228 1.178 2.228 2.121 0 1.178-.81 2.12-2.228 2.12zM86.316 117.08h-2.374c-1.095 0-2.008-1.178-2.008-2.12 0-1.179.913-2.121 2.008-2.121h2.374c1.096 0 2.008 1.178 2.008 2.121 0 1.178-.912 2.12-2.008 2.12zM68.502 73.466h-5.087c-1.052 0-1.93-1.179-1.93-2.121 0-1.179.878-2.121 1.93-2.121h5.087c1.053 0 1.93 1.178 1.93 2.12 0 1.179-.877 2.122-1.93 2.122zM112.957 73.466h-31.37c-1.204 0-2.208-1.179-2.208-2.121 0-1.179 1.004-2.121 2.209-2.121h31.369c1.205 0 2.209 1.178 2.209 2.12 0 1.179-1.004 2.122-2.209 2.122zM92.473 62.861H63.728c-1.223 0-2.242-1.178-2.242-2.12 0-1.18 1.019-2.122 2.242-2.122h28.745c1.223 0 2.243 1.178 2.243 2.121 0 1.178-.816 2.121-2.243 2.121zM130.816 62.861h-28.745c-1.224 0-2.243-1.178-2.243-2.12 0-1.18 1.019-2.122 2.243-2.122h28.745c1.223 0 2.242 1.178 2.242 2.121 0 1.178-.815 2.121-2.242 2.121zM151.349 62.861h-11.02c-1.177 0-2.158-1.178-2.158-2.12 0-1.18.981-2.122 2.158-2.122h11.02c1.178 0 2.159 1.178 2.159 2.121 0 1.178-.785 2.121-2.159 2.121z" fill="#707070"/><defs><linearGradient id="paint0_linear_3871_353928" x1="182.43" y1="124.818" x2="198.846" y2="102.93" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset="1" stop-color="#F5F2EC"/></linearGradient><filter id="filter0_d_3871_353928" x="29.656" y="23.336" width="168.286" height="131.355" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="9.9"/><feGaussianBlur stdDeviation="9.9"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3871_353928"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3871_353928" result="shape"/></filter></defs></svg>`));
@@ -45,7 +49,3 @@ export const LaunchDark = (/*#__PURE__*/createFluentIllustration('LaunchDark', `
 export const LaunchLight = (/*#__PURE__*/createFluentIllustration('LaunchLight', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M117.592 132.19c31.303 0 56.641-25.229 56.641-56.421 0-31.193-25.311-56.449-56.641-56.449-31.192 0-56.614 25.229-56.614 56.421 0 31.193 25.312 56.449 56.614 56.449z" fill="url(#paint0_linear_3269_5314)"/><path d="M179.272 141.34c11.626-12.037 3.188-32.457-13.768-32.457-7.723 0-14.346 4.562-17.397 11.131-2.693.165-5.194 1.017-10.608 4.589l-.879-19.347h-6.651l-.55 12.037c-2.171-1.676-4.259-1.484-5.716-1.374l.357-.055-.577-11.9h-11.075l-.687 14.153c-1.869-1.429-2.446-.742-3.023-.742a6.526 6.526 0 00-3.106.797l-.604-12.944h-6.651l-.467 10.251a20.298 20.298 0 00-10.334-2.831c-4.562 0-8.767 1.512-12.174 4.04-14.813-7.035-31.083 6.624-26.686 22.453l10.636 6.816-.742.495.742-.495c-3.435 3.216 6.76-6.073 9.536-4.809.22.494.44.962.687 1.429l-.467-1.292c.302.247.467.66.467 1.292 0 0 0 .027.028.027a20.115 20.115 0 00-7.695 1.924l8.85-1.594c7.557 1.622 49 7.31 49.083 8.025l-.605-1.127c.247.028.522.055.797.083 1.347 1.978 3.325 3.517 5.634 4.259a11.853 11.853 0 01-4.562-4.149c9.096.879 20.941.274 25.284-.193a15.122 15.122 0 0011.57-2.666 15.498 15.498 0 002.033-1.786c4.205-.027 11.928 1.182 11.818-.687l7.585 1.347a20.835 20.835 0 00-6.926-1.869l6.843-2.831zm-21.958 4.727a.5.5 0 01.137-.22l-.137.22z" fill="#fff"/><path d="M105.592 118.172l-.605-12.916h-6.65l-.468 10.223" fill="#fff"/><path fill-rule="evenodd" clip-rule="evenodd" d="M97.996 105.24a.34.34 0 01.34-.325h6.651c.182 0 .332.143.341.325l.604 12.916a.341.341 0 11-.681.032l-.589-12.592h-6l-.452 9.899a.341.341 0 01-.681-.032l.467-10.223z" fill="#CCC"/><path d="M137.527 125.098l-.907-19.842h-6.651l-.55 12.064" fill="#fff"/><path fill-rule="evenodd" clip-rule="evenodd" d="M129.628 105.24a.34.34 0 01.341-.325h6.65a.34.34 0 01.341.325l.907 19.842a.34.34 0 11-.681.031l-.892-19.517h-6l-.534 11.74a.342.342 0 01-.682-.031l.55-12.065z" fill="#CCC"/><path d="M124.06 115.837l-.604-11.873h-11.048l-.687 13.961.027.22" fill="#fff"/><path fill-rule="evenodd" clip-rule="evenodd" d="M112.067 103.947a.341.341 0 01.34-.324h11.048a.34.34 0 01.341.324l.604 11.872a.341.341 0 01-.681.035l-.588-11.549h-10.399l-.67 13.607.024.191a.34.34 0 01-.296.38.34.34 0 01-.38-.296l-.028-.22a.384.384 0 01-.002-.059l.687-13.961zM67.649 135.314a.341.341 0 01-.305.374 13.361 13.361 0 00-2.974.643.341.341 0 01-.222-.645c.988-.339 2.03-.564 3.127-.676a.341.341 0 01.374.304z" fill="#CCC"/><path d="M122.302 99.732h-9.069l-1.869 1.264a1.396 1.396 0 00-.605 1.154v1.237h14.291v-1.182c0-.495-.247-.934-.659-1.182l-2.089-1.291z" fill="#fff"/><path d="M125.654 103.964c0 .302-.247.55-.549.55h-14.511a.552.552 0 01-.55-.55c0-.302.248-.55.55-.55h14.511c.302 0 .549.248.549.55z" fill="url(#paint1_linear_3269_5314)"/><path d="M94.077 23.084h-.22a.622.622 0 01-.632-.633v-.22a.493.493 0 00-.494-.494.493.493 0 00-.495.495v.22a.622.622 0 01-.632.631h-.22a.493.493 0 00-.495.495c0 .275.22.495.495.495h.22c.357 0 .632.275.632.632v.22c0 .275.22.494.495.494.274 0 .494-.22.494-.494v-.22c0-.357.275-.632.632-.632h.22c.275 0 .495-.22.495-.495a.511.511 0 00-.495-.494zM80.061 14.316h-.192a.522.522 0 01-.522-.522v-.192a.423.423 0 00-.412-.412.423.423 0 00-.413.412v.192c0 .275-.22.522-.522.522h-.192a.423.423 0 00-.412.413c0 .22.192.412.412.412H78c.275 0 .522.22.522.522v.193c0 .22.193.412.413.412.22 0 .412-.193.412-.412v-.193c0-.275.22-.522.522-.522h.192c.22 0 .413-.192.413-.412a.423.423 0 00-.413-.412zM62.995 72.442h-.193a.522.522 0 01-.522-.522v-.193a.423.423 0 00-.412-.412.423.423 0 00-.412.412v.193c0 .275-.22.522-.523.522h-.192a.423.423 0 00-.412.412c0 .22.192.412.412.412h.193c.274 0 .522.22.522.523v.192c0 .22.192.412.412.412.22 0 .412-.192.412-.412v-.192c0-.275.22-.523.522-.523h.193c.22 0 .412-.192.412-.412a.405.405 0 00-.412-.412zM108.23 6.1h-.192a.522.522 0 01-.522-.523v-.165a.423.423 0 00-.413-.412.423.423 0 00-.412.412v.193c0 .274-.22.522-.522.522h-.192a.423.423 0 00-.413.412c0 .22.193.412.413.412h.192c.275 0 .522.22.522.522v.193c0 .22.193.412.412.412.22 0 .413-.192.413-.412v-.248c0-.274.22-.522.522-.522h.192c.22 0 .412-.192.412-.412a.399.399 0 00-.412-.385zM180.427 38.693h-.193a.522.522 0 01-.522-.522v-.192a.422.422 0 00-.412-.413.422.422 0 00-.412.413v.192c0 .275-.22.522-.523.522h-.192a.422.422 0 00-.412.413c0 .22.192.412.412.412h.192c.275 0 .523.22.523.522v.192c0 .22.192.413.412.413.22 0 .412-.193.412-.413v-.192c0-.275.22-.522.522-.522h.193a.423.423 0 00.412-.413.423.423 0 00-.412-.412zM165.174 98.33h-.192a.522.522 0 01-.522-.522v-.192a.423.423 0 00-.413-.413.422.422 0 00-.412.413v.192c0 .275-.22.522-.522.522h-.138a.423.423 0 00-.412.413c0 .22.193.412.412.412h.193c.275 0 .522.22.522.522v.192c0 .22.192.412.412.412.22 0 .412-.192.412-.412v-.192c0-.275.22-.522.523-.522h.192c.22 0 .412-.193.412-.412-.055-.22-.247-.413-.467-.413zM61.18 102.507h-.164a.521.521 0 01-.522-.522v-.192a.422.422 0 00-.413-.412.422.422 0 00-.412.412v.192c0 .275-.22.522-.522.522h-.192a.423.423 0 00-.413.413c0 .22.193.412.413.412h.192c.275 0 .522.22.522.522v.193c0 .219.193.412.412.412.22 0 .413-.193.413-.412v-.193c0-.275.22-.522.522-.522h.192c.22 0 .412-.192.412-.412-.027-.22-.22-.413-.44-.413zM147.64 53.561h-.22a.622.622 0 01-.632-.632v-.22a.493.493 0 00-.494-.494.493.493 0 00-.495.494v.22a.622.622 0 01-.632.632h-.22a.493.493 0 00-.495.495c0 .275.22.495.495.495h.22c.357 0 .632.274.632.632v.22c0 .274.22.494.495.494.274 0 .494-.22.494-.494v-.22c0-.358.275-.632.632-.632h.22c.275 0 .495-.22.495-.495a.493.493 0 00-.495-.495zM153.879 91.405h-.22a.622.622 0 01-.632-.633v-.22a.493.493 0 00-.494-.494.493.493 0 00-.495.495v.22a.622.622 0 01-.632.632h-.248a.493.493 0 00-.494.494c0 .275.22.495.494.495h.22c.358 0 .632.275.632.632v.22c0 .275.22.495.495.495s.495-.22.495-.495v-.22c0-.357.275-.632.632-.632h.22c.274 0 .494-.22.494-.495.028-.275-.192-.494-.467-.494zM177.954 52.764h-.138a.376.376 0 01-.385-.384v-.11a.305.305 0 00-.302-.303.305.305 0 00-.302.303v.137c0 .22-.165.385-.385.385h-.137a.306.306 0 00-.303.302c0 .165.138.303.303.303h.137c.22 0 .385.164.385.384v.138c0 .165.137.302.302.302a.305.305 0 00.302-.302v-.138c0-.22.165-.384.385-.384h.138a.305.305 0 00.302-.303c0-.192-.138-.33-.302-.33zM165.092 59.58h-.11a.305.305 0 01-.302-.302v-.11a.246.246 0 00-.248-.248.246.246 0 00-.247.248v.11a.305.305 0 01-.302.302h-.11a.247.247 0 00-.248.247c0 .138.11.248.248.248h.11c.164 0 .302.137.302.302v.11c0 .137.11.247.247.247.138 0 .248-.11.248-.247v-.11c0-.165.137-.302.302-.302h.11c.137 0 .247-.11.247-.248a.246.246 0 00-.247-.247zM138.764 24.265h-.11a.306.306 0 01-.303-.302v-.11a.246.246 0 00-.247-.247.246.246 0 00-.247.247v.11a.306.306 0 01-.303.302h-.11a.246.246 0 00-.247.248c0 .137.11.247.247.247h.11c.165 0 .303.137.303.302v.11c0 .138.11.248.247.248s.247-.11.247-.248v-.11c0-.165.138-.302.303-.302h.11c.137 0 .247-.11.247-.247-.028-.138-.11-.248-.247-.248zM165.999 23.358h-.11a.305.305 0 01-.302-.302v-.11a.246.246 0 00-.248-.248.246.246 0 00-.247.248v.11a.305.305 0 01-.302.302h-.11a.246.246 0 00-.248.247c0 .138.11.248.248.248h.11c.165 0 .302.137.302.302v.11c0 .137.11.247.247.247.138 0 .248-.11.248-.247v-.11c0-.165.137-.302.302-.302h.11c.137 0 .247-.11.247-.248a.246.246 0 00-.247-.247zM55.684 46.224h-.137a.423.423 0 01-.412-.413v-.137a.317.317 0 00-.33-.33.317.317 0 00-.33.33v.137c0 .22-.192.413-.412.413h-.138a.318.318 0 00-.33.33c0 .192.138.33.33.33h.138c.22 0 .412.191.412.411v.138c0 .192.137.33.33.33.192 0 .33-.138.33-.33v-.138c0-.22.192-.412.412-.412h.137c.193 0 .33-.137.33-.33a.355.355 0 00-.33-.33zM82.425 85.77h-.138a.423.423 0 01-.412-.411v-.138a.317.317 0 00-.33-.33.317.317 0 00-.33.33v.138c0 .22-.192.412-.412.412h-.137a.318.318 0 00-.33.33c0 .192.137.33.33.33h.137c.22 0 .413.192.413.412v.137c0 .192.137.33.33.33.192 0 .329-.138.329-.33v-.138c0-.22.192-.412.412-.412h.138c.192 0 .33-.137.33-.33a.334.334 0 00-.33-.33zM183.285 99.457h-.137a.422.422 0 01-.412-.412v-.138a.318.318 0 00-.33-.33.317.317 0 00-.33.33v.138c0 .22-.192.412-.412.412h-.138a.317.317 0 00-.329.33c0 .192.137.33.329.33h.138c.22 0 .412.192.412.412v.137c0 .193.137.33.33.33.192 0 .33-.137.33-.33v-.137c0-.22.192-.412.412-.412h.137c.193 0 .33-.138.33-.33a.355.355 0 00-.33-.33zM52.964 28.717h-.11a.317.317 0 01-.33-.33v-.11a.275.275 0 00-.55 0v.11c0 .193-.137.33-.33.33h-.11a.275.275 0 000 .55h.138c.192 0 .33.137.33.33v.11a.275.275 0 00.55 0v-.11c0-.193.137-.33.33-.33h.11c.137 0 .274-.11.274-.275a.32.32 0 00-.302-.275zM63.325 16.323h-.11a.318.318 0 01-.33-.33v-.11a.275.275 0 00-.55 0v.11c0 .192-.137.33-.33.33h-.11a.275.275 0 000 .55h.11c.193 0 .33.137.33.33v.109a.275.275 0 00.55 0v-.11c0-.192.137-.33.33-.33h.11c.137 0 .274-.11.274-.275 0-.164-.11-.274-.274-.274zM48.676 70.216h-.11a.318.318 0 01-.33-.33v-.11a.275.275 0 00-.549 0v.11c0 .192-.137.33-.33.33h-.082a.275.275 0 000 .55h.11c.192 0 .33.137.33.33v.109a.275.275 0 00.55 0v-.11c0-.192.137-.33.329-.33h.11c.137 0 .275-.11.275-.275-.028-.164-.138-.274-.303-.274zM89.02 37.182h-.11a.318.318 0 01-.33-.33v-.11a.275.275 0 00-.549 0v.11c0 .193-.137.33-.33.33h-.11a.275.275 0 000 .55h.11c.193 0 .33.137.33.33v.11a.275.275 0 00.55 0v-.11c0-.193.137-.33.33-.33h.11c.137 0 .275-.11.275-.275a.275.275 0 00-.275-.275zM81.628 69.226h-.11a.318.318 0 01-.33-.33v-.11a.275.275 0 00-.55 0v.11c0 .193-.137.33-.33.33h-.11a.275.275 0 000 .55h.11c.193 0 .33.137.33.33v.11a.275.275 0 00.55 0v-.11c0-.193.137-.33.33-.33h.11c.137 0 .275-.11.275-.275-.028-.165-.138-.275-.275-.275zM129.062 12.585a1.536 1.536 0 01-1.539-1.539c0-.852.687-1.539 1.539-1.539.852 0 1.539.687 1.539 1.54 0 .851-.714 1.538-1.539 1.538zm0-2.226a.71.71 0 00-.714.715c0 .412.302.714.714.714a.71.71 0 00.715-.714.71.71 0 00-.715-.715zM75.362 29.514a1.536 1.536 0 01-1.539-1.539c0-.852.687-1.539 1.54-1.539.851 0 1.538.687 1.538 1.54 0 .851-.714 1.538-1.539 1.538zm0-2.226a.71.71 0 00-.715.715c0 .412.303.714.715.714a.693.693 0 00.715-.714.71.71 0 00-.715-.715zM141.621 17.367a1.132 1.132 0 01-1.126-1.127c0-.632.494-1.126 1.126-1.126.605 0 1.127.494 1.127 1.126 0 .632-.522 1.127-1.127 1.127zm0-1.649a.522.522 0 00-.522.522c0 .303.22.522.522.522a.522.522 0 100-1.044zM149.866 106.712a1.132 1.132 0 01-1.127-1.127c0-.604.495-1.126 1.127-1.126.605 0 1.127.494 1.127 1.126-.027.633-.522 1.127-1.127 1.127zm0-1.621a.522.522 0 100 1.043.522.522 0 000-1.043z" fill="#D4D4D4"/><path d="M105.977 105.283h-8.822a.317.317 0 01-.33-.33c0-.192.138-.329.33-.329h8.822c.193 0 .33.137.33.329 0 .193-.165.33-.33.33z" fill="url(#paint2_linear_3269_5314)"/><path d="M105.537 103.167c.248.165.413.44.413.714v.715h-8.685v-.742c0-.275.138-.55.357-.715l1.127-.769h.028v-3.106h5.524v3.161l1.236.742z" fill="url(#paint3_linear_3269_5314)"/><path opacity=".3" d="M105.922 103.909v.714h-2.555v-.714c0-.302-.055-.577-.165-.715l-.495-.769v-3.16h1.594v3.16l1.237.769c.247.138.384.413.384.715z" fill="#D9D2C7"/><g filter="url(#filter0_d_3269_5314)"><path d="M102.129 29.267a.882.882 0 00-1.181 0c-1.622 1.374-4.15 6.238-4.15 11.405v60.709c0 .577.467 1.016 1.017 1.016h7.447c.578 0 1.017-.467 1.017-1.016V40.672c.028-5.167-2.528-10.004-4.15-11.405z" fill="url(#paint4_linear_3269_5314)"/></g><path d="M106.307 40.672v60.709c0 .577-.467 1.044-1.044 1.044h-2.638c.164 0 .302-.467.302-1.044V40.672c0-5.166-.742-10.003-1.209-11.405-.055-.137-.11-.22-.165-.22.22 0 .412.083.577.22 1.621 1.402 4.177 6.238 4.177 11.405z" fill="url(#paint5_linear_3269_5314)"/><g filter="url(#filter1_d_3269_5314)"><path d="M118.839 13.602c-.797-.385-1.979-.385-2.776 0-3.765 1.896-9.784 10.773-9.784 17.836v67.277h22.316V31.438c0-7.035-5.991-15.912-9.756-17.836z" fill="url(#paint6_linear_3269_5314)"/></g><path d="M128.595 31.466v67.277h-6.513V31.466c0-7.063-2.501-15.94-4.067-17.836-.165-.193-.385-.303-.577-.303.494 0 .989.11 1.374.303 3.792 1.896 9.783 10.773 9.783 17.836z" fill="url(#paint7_linear_3269_5314)"/><path opacity=".4" d="M117.437 13.327c-.164 0-.329.11-.439.303-1.237 1.896-3.161 10.773-3.161 17.836v67.277h-4.617V31.466c0-7.063 4.425-15.94 7.228-17.836.275-.22.632-.303.989-.303z" fill="url(#paint8_linear_3269_5314)"/><path opacity=".4" fill-rule="evenodd" clip-rule="evenodd" d="M106.507 36.035h22.162v1.023h-22.162v-1.023zM106.166 47.968h22.503v1.023h-22.503v-1.023z" fill="#979087"/><path fill-rule="evenodd" clip-rule="evenodd" d="M96.779 62.961h6.478v1.023h-6.478v-1.023zM96.779 52.392h6.478v1.023h-6.478v-1.023zM96.779 41.14h6.478v1.024h-6.478V41.14z" fill="#E8E4DC"/><path d="M137.775 105.283h-8.822a.317.317 0 01-.33-.33c0-.192.137-.329.33-.329h8.822c.192 0 .329.137.329.329 0 .193-.164.33-.329.33z" fill="url(#paint9_linear_3269_5314)"/><path d="M137.334 103.167c.248.165.412.44.412.714v.715h-8.684v-.742c0-.275.137-.55.357-.715l1.127-.769h.028v-3.106h5.524v3.161l1.236.742z" fill="url(#paint10_linear_3269_5314)"/><path opacity=".3" d="M137.719 103.909v.714h-2.281v-.714c0-.302-.082-.577-.192-.715l-.578-.769v-3.16h1.43v3.16l1.236.769c.248.138.385.413.385.715z" fill="#D9D2C7"/><g filter="url(#filter2_d_3269_5314)"><path d="M133.871 29.434a.9.9 0 00-1.194 0c-1.637 1.373-4.19 6.234-4.19 11.397v60.665a1.02 1.02 0 001.027 1.016h7.492a1.02 1.02 0 001.027-1.016V40.831c.056-5.163-2.525-9.996-4.162-11.397z" fill="url(#paint11_linear_3269_5314)"/></g><path d="M138.105 40.672v60.709c0 .577-.468 1.044-1.045 1.044h-2.638c.165 0 .302-.467.302-1.044V40.672c0-5.166-.742-10.003-1.209-11.405-.055-.137-.11-.22-.165-.22.22 0 .412.083.577.22 1.622 1.402 4.178 6.238 4.178 11.405z" fill="url(#paint12_linear_3269_5314)"/><path fill-rule="evenodd" clip-rule="evenodd" d="M128.487 62.961h6.478v1.023h-6.478v-1.023zM128.487 53.756h6.478v1.023h-6.478v-1.023zM128.487 41.14h6.478v1.024h-6.478V41.14z" fill="#EEEAE4"/><g filter="url(#filter3_d_3269_5314)"><path d="M111.034 49.604l.467-1.759 1.155-.275 1.923-.44.22-.054.165 2.528 2.858 41.718-6.183 8.437-2.419-.412-16.819-2.885a2.129 2.129 0 01-1.786-2.117v-7.942c0-.934.385-1.869 1.044-2.529l11.323-11.487a14.93 14.93 0 003.82-6.706l2.418-9.124 1.814-6.953z" fill="url(#paint13_linear_3269_5314)"/></g><path d="M105.995 84.62a2.727 2.727 0 11-5.455.001 2.727 2.727 0 015.455 0z" fill="#BCB9B3"/><path d="M112.628 47.57v11.515l-1.237 1.374-.11-.742a.736.736 0 00-.714-.632c-.33 0-.605.22-.715.523l-3.298 12.174a5.035 5.035 0 01-1.237 2.171L92.95 86.348a1.293 1.293 0 00-.384.907v2.693a1.29 1.29 0 001.291 1.292h15.803l.357 2.556c0 .164-15.995.027-19.43 0v-7.448c0-.934.357-1.869 1.044-2.528l11.323-11.488a14.93 14.93 0 003.82-6.706l4.7-17.808 1.154-.248z" fill="url(#paint14_linear_3269_5314)"/><path d="M117.822 91.322l-6.183 8.41-2.446-.412V56.53l1.841-6.926h3.93l2.858 41.718z" fill="url(#paint15_linear_3269_5314)"/><g filter="url(#filter4_d_3269_5314)"><path d="M145.112 86.375v7.943a2.128 2.128 0 01-1.787 2.116l-12.614 2.171-6.623 1.127-6.184-8.41 3.023-44.246.22.055 1.924.467 1.154.275 4.205 15.857.522 1.951a14.93 14.93 0 003.82 6.706l11.323 11.487a3.533 3.533 0 011.017 2.501z" fill="url(#paint16_linear_3269_5314)"/></g><path d="M134.641 82.72h-5.222v3.71h5.222v-3.71z" fill="#BCB9B3"/><path d="M132.14 82.72h-2.721v2.034h2.721V82.72z" fill="#D9D2C7"/><path d="M123.071 47.57v11.515l1.237 1.374.11-.742a.736.736 0 01.714-.632c.33 0 .605.22.715.523l3.325 12.174c.22.825.66 1.567 1.237 2.171l12.367 12.395c.248.247.385.577.385.907v2.693a1.29 1.29 0 01-1.292 1.292h-15.775l-.357 2.556c0 .164 15.995.027 19.43 0v-7.448c0-.934-.357-1.869-1.044-2.528l-11.378-11.46a14.931 14.931 0 01-3.82-6.706l-4.699-17.809-1.155-.275z" fill="url(#paint17_linear_3269_5314)"/><path opacity=".4" d="M130.711 90.938v7.667l-6.623 1.127-6.184-8.41 3.023-44.246.22.055 1.924.467 1.154.275 4.205 15.857v23.937c1.319.467 2.281 1.759 2.281 3.27z" fill="url(#paint18_linear_3269_5314)"/><g filter="url(#filter5_d_3269_5314)"><path d="M124.308 45.592v53.15h-13.109v-53.15c0-2.309.495-4.59 1.484-6.679.193-.384.385-.797.605-1.181v-.028l2.748-4.562c.77-1.292 2.638-1.292 3.435 0l2.749 4.562v.027c.22.385.412.77.604 1.155a15.784 15.784 0 011.484 6.706z" fill="url(#paint19_linear_3269_5314)"/></g><path d="M124.309 45.592v53.15h-3.93v-53.15c0-2.309-.193-4.59-.605-6.679-.083-.384-.137-.77-.247-1.154v-.027l-1.1-4.563c-.165-.66-.412-.961-.687-.961.66 0 1.319.33 1.704.962l2.748 4.562v.027c.22.385.413.77.605 1.154 1.017 2.089 1.512 4.37 1.512 6.679z" fill="url(#paint20_linear_3269_5314)"/><path d="M122.824 38.913a2.575 2.575 0 00-2.143-1.181h-5.799c-.879 0-1.704.467-2.171 1.181.192-.384.385-.797.605-1.181v-.028l2.748-4.562c.769-1.292 2.638-1.292 3.435 0l2.748 4.562v.027c.193.385.413.77.577 1.182z" fill="url(#paint21_linear_3269_5314)"/><path d="M120.708 44.74c0 .22-.248.302-.385.137-.302-.385-.77-.605-1.319-.605h-2.529c-.494 0-.961.22-1.264.578l-.055.054c-.137.165-.385.055-.385-.137v-.66c0-.934.77-1.676 1.704-1.676h2.694c.247.028.467.11.687.22.494.302.852.852.852 1.456v.633z" fill="#77797F"/><path opacity=".5" fill-rule="evenodd" clip-rule="evenodd" d="M111.119 60.058h13.359v1.023h-13.359v-1.023zM111.119 74.733h13.359v1.023h-13.359v-1.023z" fill="#B7B4AF"/><g filter="url(#filter6_d_3269_5314)"><path d="M114.332 87.42h-1.594a3.507 3.507 0 00-3.518 3.517v8.85h8.63v-8.85c-.028-1.95-1.594-3.517-3.518-3.517z" fill="url(#paint22_linear_3269_5314)"/></g><path d="M117.823 90.937v8.85h-3.436V88.299a.868.868 0 00-.879-.88h.797a3.507 3.507 0 013.518 3.518z" fill="url(#paint23_linear_3269_5314)"/><g filter="url(#filter7_d_3269_5314)"><path d="M122.879 87.42h-1.594a3.507 3.507 0 00-3.518 3.517v8.85h8.63v-8.85a3.507 3.507 0 00-3.518-3.517z" fill="url(#paint24_linear_3269_5314)"/></g><path d="M126.397 90.937v8.85h-2.996V88.574c0-.632-.522-1.154-1.154-1.154h.632a3.507 3.507 0 013.518 3.517z" fill="url(#paint25_linear_3269_5314)"/><path d="M117.74 90.937v8.85h2.996V88.574c0-.632.522-1.154 1.154-1.154h-.632a3.525 3.525 0 00-3.518 3.517z" fill="url(#paint26_linear_3269_5314)"/><path d="M122.301 99.732h-9.069l-1.869 1.264a1.398 1.398 0 00-.604 1.154v1.237h14.291v-1.182c0-.495-.248-.934-.66-1.182l-2.089-1.291z" fill="#7B746A"/><path opacity=".3" d="M125.05 102.233v1.181h-4.287v-1.181c0-.495-.11-.935-.275-1.21l-.824-1.291h2.638l2.089 1.291c.412.248.659.715.659 1.21z" fill="#D9D2C7"/><path d="M116.888 52.187h-.604a.551.551 0 01-.55-.55v-.604c0-.302.247-.55.55-.55h.604c.302 0 .55.248.55.55v.605c0 .302-.248.55-.55.55zM119.444 52.187h-.605a.55.55 0 01-.549-.55v-.604c0-.302.247-.55.549-.55h.605c.302 0 .549.248.549.55v.605a.55.55 0 01-.549.55z" fill="#77797F"/><g filter="url(#filter8_d_3269_5314)"><path d="M118.647 105.613h-1.787V85.276c0-.495.413-.88.88-.88.494 0 .879.413.879.88v20.337h.028z" fill="url(#paint27_linear_3269_5314)"/></g><path fill-rule="evenodd" clip-rule="evenodd" d="M108.698 117.716a6.072 6.072 0 00-3.878 1.379.341.341 0 01-.434-.526 6.758 6.758 0 014.312-1.535 6.737 6.737 0 016.05 3.763.341.341 0 01-.612.302 6.051 6.051 0 00-5.438-3.383z" fill="#CCC"/><path opacity=".3" d="M118.097 42.403l-.99 1.87h-.632c-.495 0-.962.219-1.264.576l1.292-2.446h1.594zM119.856 42.623l-.852 1.65h-.824l.989-1.87c.247.028.467.11.687.22z" fill="#fff"/><path fill-rule="evenodd" clip-rule="evenodd" d="M87.509 113.017c-11.052 0-19.996 8.944-19.996 19.996 0 .754.054 1.51.135 2.272a.341.341 0 11-.678.073c-.083-.777-.139-1.56-.139-2.345 0-11.429 9.25-20.678 20.678-20.678 9.529 0 17.548 6.427 19.951 15.201a.34.34 0 11-.657.181c-2.324-8.485-10.08-14.7-19.294-14.7zM67.539 136.912a.342.342 0 01.406.26 20.636 20.636 0 001.92 5.244.342.342 0 01-.141.462.342.342 0 01-.462-.141 21.325 21.325 0 01-1.983-5.418.342.342 0 01.26-.407z" fill="#CCC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M114.332 121.262c-3.556 0-6.486 2.799-6.695 6.302a.341.341 0 01-.681-.04c.231-3.862 3.457-6.944 7.376-6.944 1.47 0 2.824.432 3.973 1.151a.34.34 0 11-.361.578 6.779 6.779 0 00-3.612-1.047zM125.105 116.178a6.668 6.668 0 00-3.765 1.149.342.342 0 01-.385-.563 7.357 7.357 0 014.15-1.268c2.45 0 4.583 1.183 5.935 2.995a.341.341 0 01-.547.408c-1.231-1.651-3.166-2.721-5.388-2.721zM162.442 128.199a.341.341 0 01.456.158 15.155 15.155 0 011.518 6.662c0 1.235-.14 2.444-.422 3.599a.341.341 0 01-.662-.162c.268-1.099.402-2.253.402-3.437 0-2.284-.51-4.432-1.45-6.364a.342.342 0 01.158-.456zM149.015 120.327c-4.995 0-9.4 2.497-12.06 6.313a.341.341 0 01-.559-.39c2.781-3.99 7.391-6.605 12.619-6.605a15.35 15.35 0 0112.226 6.043.342.342 0 01-.542.414 14.668 14.668 0 00-11.684-5.775z" fill="#CCC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M163.734 133.975c0-.188.153-.341.341-.341h.412a6.967 6.967 0 016.964 6.964 6.815 6.815 0 01-1.099 3.731.34.34 0 11-.572-.371 6.133 6.133 0 00.989-3.36 6.285 6.285 0 00-6.282-6.282h-.412a.341.341 0 01-.341-.341z" fill="#CCC"/><path opacity=".3" d="M164.075 135.047c0 10.498-10.498 17.753-20.309 14.098 9.399 2.061 18.303-5.084 18.303-14.703a15.016 15.016 0 00-9.811-14.098c6.76 1.484 11.817 7.502 11.817 14.703z" fill="#EDE5DA"/><path fill-rule="evenodd" clip-rule="evenodd" d="M156.073 148.32a.34.34 0 01-.132.464c-.261.144-.52.26-.764.368l-.005.002a.34.34 0 11-.277-.623c.249-.11.483-.215.715-.344a.34.34 0 01.463.133zM163.766 138.04a.34.34 0 01.256.408 15.456 15.456 0 01-5.929 9.021.342.342 0 01-.403-.55 14.78 14.78 0 005.668-8.623.34.34 0 01.408-.256zM90.58 137.907a.34.34 0 01.39-.284 4.721 4.721 0 013.851 3.554.341.341 0 01-.662.161 4.042 4.042 0 00-3.294-3.042.34.34 0 01-.285-.389z" fill="#CCC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M99.381 135.223c-2.68 0-5.048 1.419-6.39 3.525a.34.34 0 11-.576-.367c1.461-2.291 4.04-3.84 6.966-3.84 1.662 0 3.211.487 4.504 1.349a.34.34 0 11-.378.567 7.406 7.406 0 00-4.126-1.234z" fill="#CCC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M110.209 127.335a7.547 7.547 0 00-7.546 7.547c0 .196 0 .364.024.529a.34.34 0 01-.289.385.34.34 0 01-.386-.289c-.031-.216-.031-.429-.031-.615v-.01c0-4.559 3.699-8.229 8.228-8.229 4.531 0 8.229 3.698 8.229 8.229 0 .599-.057 1.171-.172 1.743a.341.341 0 11-.668-.134 8.145 8.145 0 00.158-1.609c0-4.154-3.393-7.547-7.547-7.547z" fill="#CCC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M117.685 136.323a.341.341 0 01.482-.011 7.149 7.149 0 011.558 2.136.341.341 0 01-.618.287 6.46 6.46 0 00-1.411-1.93.341.341 0 01-.011-.482z" fill="#CCC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M129.09 133.162a10.621 10.621 0 00-10.225 7.695.341.341 0 01-.656-.188 11.304 11.304 0 0110.881-8.189c5.384 0 9.89 3.768 11.023 8.785a.34.34 0 11-.665.15 10.626 10.626 0 00-10.358-8.253z" fill="#CCC"/><path opacity=".3" d="M125.627 154.174c-4.369-1.456-7.53-5.551-7.53-10.415 0-5.085 3.463-9.344 8.135-10.609-8.987 3.93-9.344 16.6-.605 21.024z" fill="#EDE5DA"/><path fill-rule="evenodd" clip-rule="evenodd" d="M118.152 142.373c.188 0 .341.153.341.341 0 .094-.007.185-.014.265l-.001.013a2.97 2.97 0 00-.012.244.34.34 0 11-.682 0c0-.111.008-.213.015-.298v-.003c.007-.088.012-.155.012-.221 0-.188.153-.341.341-.341zM141.622 144.127a5.154 5.154 0 00-5.156 5.156c0 .233.021.445.047.696l.006.067a.341.341 0 11-.678.067 7.292 7.292 0 01-.057-.83 5.836 5.836 0 015.838-5.838 5.836 5.836 0 015.837 5.838.341.341 0 01-.682 0 5.154 5.154 0 00-5.155-5.156zM180.263 134.398c-.647 0-1.228.347-1.554.859a.34.34 0 11-.575-.366 2.535 2.535 0 012.129-1.175 2.487 2.487 0 012.484 2.485.34.34 0 11-.681 0c0-.994-.81-1.803-1.803-1.803zM177.878 136.289a.341.341 0 01.482 0c.128.129.127.28.127.32v.004a.34.34 0 01-.675.066.34.34 0 01.066-.39zm-.073.317v.004-.004zM178.281 137.158a.341.341 0 01.475.084 1.78 1.78 0 001.479.761c.563 0 1.046-.233 1.365-.598a.341.341 0 01.513.449 2.474 2.474 0 01-1.878.831 2.464 2.464 0 01-2.038-1.052.34.34 0 01.084-.475zM100.343 141.736c-.647 0-1.228.347-1.554.859a.34.34 0 01-.575-.366 2.535 2.535 0 012.129-1.175 2.487 2.487 0 012.485 2.485.34.34 0 01-.682 0c0-.994-.81-1.803-1.803-1.803zM97.958 143.655a.34.34 0 01.482 0c.129.128.128.279.128.319v.004a.341.341 0 01-.676.066.34.34 0 01.066-.389zm-.072.316v.004-.004zM98.36 144.496a.34.34 0 01.476.084c.323.461.852.761 1.479.761.563 0 1.046-.233 1.365-.598a.34.34 0 11.513.449 2.472 2.472 0 01-1.878.831 2.463 2.463 0 01-2.038-1.052.341.341 0 01.084-.475zM70.635 142.918c-3.053 0-5.973.702-8.596 1.919a.34.34 0 11-.287-.618 21.101 21.101 0 018.883-1.983c.587 0 1.176.028 1.737.084a.341.341 0 01-.067.678 17.159 17.159 0 00-1.67-.08zM60.377 145.319a.341.341 0 01-.125.466c-.51.295-.994.618-1.482.971a.34.34 0 11-.4-.553 18.092 18.092 0 011.54-1.008.341.341 0 01.467.124zM168.851 143.859c.561-.056 1.15-.084 1.737-.084 3.158 0 6.175.727 8.883 1.983a.341.341 0 11-.287.618 20.415 20.415 0 00-8.596-1.919c-.566 0-1.132.027-1.67.081a.342.342 0 01-.067-.679z" fill="#CCC"/><path opacity=".3" fill-rule="evenodd" clip-rule="evenodd" d="M165.504 109.224c-7.585 0-14.117 4.481-17.087 10.932a.341.341 0 11-.619-.285c3.077-6.685 9.846-11.329 17.706-11.329 10.769 0 19.496 8.728 19.496 19.497 0 5.257-2.07 10.041-5.455 13.538a.34.34 0 11-.49-.474 18.723 18.723 0 005.263-13.064c0-10.393-8.422-18.815-18.814-18.815zM75.216 117.024c-6.48-3.078-14.443-2.295-20.273 2.645-5.776 4.885-7.855 12.523-5.938 19.381a.341.341 0 11-.657.183c-1.986-7.104.168-15.021 6.154-20.084 6.043-5.119 14.294-5.93 21.006-2.741a.34.34 0 11-.292.616z" fill="#CCC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M54.42 93.89c-1.018 0-1.83.833-1.83 1.83a.341.341 0 11-.682 0 2.517 2.517 0 012.512-2.513c.274 0 .537.03.807.132a.341.341 0 01-.24.638 1.56 1.56 0 00-.567-.088zM59.85 94.455a.34.34 0 01.341-.34c.897 0 1.636.635 1.794 1.483h2.906a.341.341 0 010 .682h-3.215a.34.34 0 01-.341-.34c0-.637-.507-1.144-1.144-1.144a.341.341 0 01-.34-.34z" fill="#CCC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M57.553 92.24c-.998 0-1.835.667-2.122 1.551a.34.34 0 01-.648-.21 2.924 2.924 0 012.77-2.023 2.91 2.91 0 012.897 2.897.341.341 0 01-.682 0 2.228 2.228 0 00-2.215-2.215zM77.075 102.678a.344.344 0 01-.197.277.343.343 0 01-.364-.051 1.854 1.854 0 00-1.207-.468c-.994 0-1.803.809-1.803 1.803v.137a.341.341 0 01-.34.341h-2.282a.341.341 0 010-.682h1.949a2.488 2.488 0 012.476-2.281c.465 0 .893.135 1.259.351.727-1.682 2.395-2.88 4.347-2.88a4.751 4.751 0 014.738 4.739c0 .131-.008.268-.02.401h.861a.341.341 0 010 .682h-1.181a.34.34 0 01-.242-.1l-.055-.055a.34.34 0 01-.097-.283 5.48 5.48 0 00.052-.645 4.07 4.07 0 00-4.056-4.056c-1.779 0-3.29 1.167-3.838 2.77zM151.241 71.519a3.646 3.646 0 00-3.644 3.644.341.341 0 01-.682 0 4.329 4.329 0 014.326-4.326c.482 0 .927.09 1.352.204a.34.34 0 11-.176.659c-.4-.107-.779-.181-1.176-.181zM173.49 75.575c0-.189.153-.341.341-.341h7.036a.341.341 0 010 .682h-7.036a.341.341 0 01-.341-.341zM161.59 72.826c0-.188.153-.34.341-.34a3.05 3.05 0 013.044 2.72h5.613a.34.34 0 110 .682h-5.936a.341.341 0 01-.341-.34 2.368 2.368 0 00-2.38-2.38.341.341 0 01-.341-.342z" fill="#CCC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M157.039 68.44a4.379 4.379 0 00-4.182 3.034.34.34 0 11-.65-.208 5.062 5.062 0 014.832-3.508 5.068 5.068 0 015.068 5.069.341.341 0 01-.682 0 4.385 4.385 0 00-4.386-4.387zM85.753 78.77c.21-.07.468-.1.685-.1a2.212 2.212 0 012.21 2.209.341.341 0 01-.682 0 1.53 1.53 0 00-1.528-1.528 1.62 1.62 0 00-.47.065.34.34 0 11-.215-.647zM72.273 81.043a.34.34 0 01.34-.34h3.299a.341.341 0 110 .681h-3.298a.341.341 0 01-.341-.34zM79.894 80.703a1.6 1.6 0 011.57-1.264.341.341 0 110 .681.92.92 0 00-.924.924c0 .188-.153.34-.34.34h-2.777a.341.341 0 110-.681h2.471z" fill="#CCC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M81.205 79.752a2.546 2.546 0 012.54-2.54c1.115 0 2.096.761 2.413 1.779a.341.341 0 11-.65.203 1.88 1.88 0 00-1.764-1.3c-1.02 0-1.857.837-1.857 1.858a.341.341 0 01-.682 0zM160.216 88.959a4.546 4.546 0 014.546-4.546 4.554 4.554 0 014.182 2.744 2.375 2.375 0 011.204-.326c1.247 0 2.264.952 2.366 2.171h1.839a.341.341 0 010 .682h-2.171a.341.341 0 01-.341-.34v-.138c0-.939-.754-1.693-1.693-1.693-.433 0-.823.158-1.152.44a.34.34 0 01-.561-.226 3.877 3.877 0 00-3.673-2.632 3.864 3.864 0 00-3.864 3.864c0 .195.024.393.052.617a.342.342 0 01-.097.283l-.055.055a.342.342 0 01-.241.1h-1.127a.341.341 0 010-.682h.806a3.814 3.814 0 01-.02-.373zM63.835 27.59h.682v.138h-.682v-.137z" fill="#CCC"/><path d="M164.783 41.41a8.107 8.107 0 10-11.466-11.465 8.107 8.107 0 0011.466 11.465z" fill="url(#paint28_linear_3269_5314)"/><path d="M160.749 43.613a144.313 144.313 0 01-9.619-9.62c1.32-6.238 9.097-8.629 13.687-4.066 4.534 4.562 2.171 12.367-4.068 13.686z" fill="#F1F3FA"/><path fill-rule="evenodd" clip-rule="evenodd" d="M151.502 33.895a143.875 143.875 0 009.346 9.346c5.848-1.367 8.029-8.745 3.728-13.073-4.355-4.328-11.707-2.123-13.074 3.727zm-.705.028c1.374-6.5 9.477-8.993 14.26-4.239l.001.002c4.726 4.754 2.264 12.885-4.238 14.26l-.171.036-.129-.117a144.604 144.604 0 01-9.642-9.642l-.117-.13.036-.17z" fill="#CCC"/><g filter="url(#filter9_d_3269_5314)"><path d="M154.154 29.212c-5.607-4.974-10.059-7.942-10.938-7.063-.907.907 2.418 5.826 7.887 11.845 9.014 9.976 20.035 18.963 21.491 17.506 1.512-1.511-8.272-13.301-18.44-22.288zm8.739 13.604c-9.811-7.448-18.111-19.87-9.206-13.247 6.266 4.645 14.345 14.016 13.191 15.17-.412.413-1.924-.357-3.985-1.923z" fill="#D9D2C7"/></g><g opacity=".3" filter="url(#filter10_d_3269_5314)"><path d="M146.981 25.364c-1.017 1.017 3.16 6.843 9.344 13.027 6.183 6.184 12.009 10.36 13.026 9.344 2.391-2.391-19.952-24.79-22.37-22.37zm2.995 2.474c.742-.742 5.139 2.418 9.811 7.09s7.833 9.042 7.091 9.811c-1.787 1.787-18.688-15.115-16.902-16.901z" fill="#EDE5DA"/></g><path d="M162.893 42.816c-3.6-2.721-8.272-7.393-10.993-10.993 2.501-4.672 8.959-5.827 12.889-1.897 3.903 3.903 2.803 10.361-1.896 12.89z" fill="url(#paint29_linear_3269_5314)"/><path opacity=".3" d="M164.762 41.442a8.85 8.85 0 01-1.869 1.374c-3.6-2.721-8.272-7.393-10.993-10.993.33-.632.797-1.182 1.319-1.732-.577 2.254 9.949 12.56 11.543 11.35z" fill="#867966"/><path opacity=".4" d="M164.597 38.144c.622 0 1.126-.825 1.126-1.842 0-1.017-.504-1.841-1.126-1.841-.623 0-1.127.824-1.127 1.841s.504 1.842 1.127 1.842zM161.794 32.565c.501 0 .907-.492.907-1.1 0-.607-.406-1.099-.907-1.099s-.907.492-.907 1.1c0 .607.406 1.099.907 1.099zM156.207 31.706c.464-.075.791-.44.731-.815-.061-.374-.487-.617-.951-.542-.464.075-.792.44-.731.815.06.374.486.617.951.542zM159.223 34.751c.749-.121 1.274-.73 1.172-1.359-.102-.63-.791-1.04-1.541-.92-.749.122-1.273.73-1.172 1.36.102.629.792 1.04 1.541.92z" fill="#85827B"/><g filter="url(#filter11_d_3269_5314)"><path d="M85.915 49.577c0 8.766-8.85 14.813-17.012 11.597a12.484 12.484 0 01-5.359-4.067c-6.183-8.135-.44-20.008 9.922-20.008 1.429 0 2.775.248 4.04.66 4.919 1.731 8.41 6.404 8.41 11.818z" fill="url(#paint30_linear_3269_5314)"/></g><path d="M71.9 50.758a3.19 3.19 0 01-3.189 3.188c-.494 0-.962-.11-1.401-.33a3.152 3.152 0 01-1.787-2.858 3.19 3.19 0 015.524-2.17c.523.549.852 1.318.852 2.17z" fill="url(#paint31_linear_3269_5314)"/><path opacity=".4" d="M71.899 50.758a3.19 3.19 0 01-3.188 3.188c-.495 0-.962-.11-1.401-.33a3.2 3.2 0 01-.853-2.17 3.19 3.19 0 013.189-3.188c.494 0 .961.11 1.401.33.522.549.852 1.319.852 2.17z" fill="#BCB9B3"/><path d="M76.71 45.4c0 1.209-.962 2.17-2.172 2.17-.22 0-.412-.027-.604-.082-.907-.247-1.567-1.1-1.567-2.089 0-1.209.962-2.17 2.171-2.17.605 0 1.127.247 1.512.604.412.412.66.962.66 1.566z" fill="url(#paint32_linear_3269_5314)"/><path opacity=".4" d="M76.709 45.4c0 1.209-.962 2.17-2.171 2.17-.22 0-.413-.027-.605-.082a2.094 2.094 0 01-.66-1.539c0-1.21.962-2.171 2.172-2.171.22 0 .412.027.604.082.412.385.66.935.66 1.54z" fill="#BCB9B3"/><path d="M78.496 54.304c0 1.209-.962 2.17-2.172 2.17-.11 0-.247 0-.357-.027a2.16 2.16 0 01-1.814-2.143c0-1.21.962-2.172 2.171-2.172.44 0 .852.138 1.21.358.577.384.962 1.044.962 1.814z" fill="url(#paint33_linear_3269_5314)"/><path opacity=".4" d="M78.495 54.304c0 1.209-.962 2.17-2.171 2.17-.11 0-.247 0-.357-.027a2.164 2.164 0 01-.962-1.814c0-1.209.962-2.17 2.171-2.17.11 0 .22 0 .357.027.577.385.962 1.044.962 1.814z" fill="#BCB9B3"/><path d="M80.914 49.302a1.42 1.42 0 01-1.43 1.429c-.082 0-.164 0-.246-.028a1.448 1.448 0 01-1.21-1.429c0-.797.66-1.429 1.43-1.429.33 0 .632.11.879.302.357.303.577.715.577 1.155z" fill="url(#paint34_linear_3269_5314)"/><path opacity=".4" d="M80.914 49.302a1.42 1.42 0 01-1.43 1.429c-.082 0-.164 0-.247-.028a1.49 1.49 0 01-.55-1.126c0-.797.633-1.43 1.43-1.43a.76.76 0 01.247.028c.33.275.55.687.55 1.127z" fill="#BCB9B3"/><path opacity=".1" d="M85.916 49.577c0 8.767-8.85 14.813-17.012 11.597.193-.165.44-.275.66-.385 4.26-2.088 8.19-.384 9.866-3.352.77-1.374.33-3.16 1.044-4.562.33-.688.907-1.237 1.237-1.897 1.182-2.418-1.237-7.64-.77-9.096.193-.578.77-.935 1.375-1.045a12.387 12.387 0 013.6 8.74z" fill="#3C3B3B"/><path opacity=".1" d="M77.808 39.683c-.687 1.731-3.545 1.209-4.81 1.236-2.555.028-2.913 1.32-3.435 2.007-.824 1.072-2.336 1.511-3.6 1.072 1.1 5.276-1.182 4.809-1.594 6.485-.467 1.98 3.05 3.188 2.391 5.14-.385 1.126-2.034 1.401-3.243 1.484-6.184-8.135-.44-20.008 9.921-20.008 1.43 0 2.776.248 4.04.66.385.605.577 1.32.33 1.924z" fill="#fff"/><path opacity=".3" d="M67.86 45.18a.88.88 0 100-1.76.88.88 0 000 1.76z" fill="#8A8987"/><defs><linearGradient id="paint0_linear_3269_5314" x1="117.606" y1="19.32" x2="117.606" y2="132.19" gradientUnits="userSpaceOnUse"><stop stop-color="#F0F0F0"/><stop offset="1" stop-color="#E6E6E6"/></linearGradient><linearGradient id="paint1_linear_3269_5314" x1="125.035" y1="103.414" x2="110.556" y2="104.13" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint2_linear_3269_5314" x1="105.931" y1="104.624" x2="97.137" y2="105.064" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint3_linear_3269_5314" x1="105.605" y1="99.264" x2="97.53" y2="99.31" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint4_linear_3269_5314" x1="96.733" y1="74.249" x2="103.516" y2="74.168" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F6F6F4"/></linearGradient><linearGradient id="paint5_linear_3269_5314" x1="107.611" y1="62.97" x2="101.474" y2="62.8" gradientUnits="userSpaceOnUse"><stop stop-color="#998E7B"/><stop offset="1" stop-color="#E9DCCA" stop-opacity=".1"/></linearGradient><linearGradient id="paint6_linear_3269_5314" x1="117.43" y1="11.338" x2="117.43" y2="99.636" gradientUnits="userSpaceOnUse"><stop stop-color="#E9DCC9"/><stop offset=".996" stop-color="#D7CDC1"/></linearGradient><linearGradient id="paint7_linear_3269_5314" x1="130.544" y1="54.617" x2="123.043" y2="54.617" gradientUnits="userSpaceOnUse"><stop stop-color="#998E7B"/><stop offset="1" stop-color="#998E7B" stop-opacity="0"/></linearGradient><linearGradient id="paint8_linear_3269_5314" x1="123.554" y1="35.865" x2="103.097" y2="28.534" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset="1" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint9_linear_3269_5314" x1="137.728" y1="104.624" x2="128.934" y2="105.064" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint10_linear_3269_5314" x1="137.402" y1="99.264" x2="129.327" y2="99.31" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint11_linear_3269_5314" x1="128.422" y1="74.384" x2="135.252" y2="74.302" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F6F6F4"/></linearGradient><linearGradient id="paint12_linear_3269_5314" x1="139.409" y1="62.97" x2="133.272" y2="62.8" gradientUnits="userSpaceOnUse"><stop stop-color="#998E7B"/><stop offset="1" stop-color="#E9DCCA" stop-opacity=".1"/></linearGradient><linearGradient id="paint13_linear_3269_5314" x1="104.21" y1="45.857" x2="104.21" y2="100.327" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint14_linear_3269_5314" x1="111.753" y1="47.57" x2="91.259" y2="47.604" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint15_linear_3269_5314" x1="120.19" y1="72.779" x2="109.102" y2="71.96" gradientUnits="userSpaceOnUse"><stop stop-color="#E9DCCA"/><stop offset="1" stop-color="#E9DCCA" stop-opacity="0"/></linearGradient><linearGradient id="paint16_linear_3269_5314" x1="131.499" y1="45.858" x2="131.499" y2="100.3" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint17_linear_3269_5314" x1="132.749" y1="44.38" x2="115.712" y2="51.553" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint18_linear_3269_5314" x1="124.406" y1="79.847" x2="128.46" y2="80.24" gradientUnits="userSpaceOnUse"><stop stop-color="#E9DCCA"/><stop offset="1" stop-color="#E9DCCA" stop-opacity="0"/></linearGradient><linearGradient id="paint19_linear_3269_5314" x1="111.11" y1="73.199" x2="120.486" y2="73.028" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F6F6F4"/></linearGradient><linearGradient id="paint20_linear_3269_5314" x1="126.11" y1="62.967" x2="120.987" y2="62.962" gradientUnits="userSpaceOnUse"><stop stop-color="#B6B1AA"/><stop offset="1" stop-color="#EDE5D9" stop-opacity=".1"/></linearGradient><linearGradient id="paint21_linear_3269_5314" x1="122.423" y1="32.173" x2="113.019" y2="32.222" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint22_linear_3269_5314" x1="109.161" y1="95.041" x2="115.31" y2="94.645" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F6F6F4"/></linearGradient><linearGradient id="paint23_linear_3269_5314" x1="119.006" y1="93.137" x2="115.641" y2="93.124" gradientUnits="userSpaceOnUse"><stop stop-color="#B6B1AA"/><stop offset="1" stop-color="#EDE5D9" stop-opacity=".1"/></linearGradient><linearGradient id="paint24_linear_3269_5314" x1="117.708" y1="95.041" x2="123.857" y2="94.645" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F6F6F4"/></linearGradient><linearGradient id="paint25_linear_3269_5314" x1="127.535" y1="93.137" x2="124.298" y2="93.125" gradientUnits="userSpaceOnUse"><stop stop-color="#B6B1AA"/><stop offset="1" stop-color="#EDE5D9" stop-opacity=".1"/></linearGradient><linearGradient id="paint26_linear_3269_5314" x1="123.028" y1="93.137" x2="117.776" y2="92.382" gradientUnits="userSpaceOnUse"><stop stop-color="#DBD0C0"/><stop offset="1" stop-color="#E3D3BD" stop-opacity="0"/></linearGradient><linearGradient id="paint27_linear_3269_5314" x1="116.848" y1="97.472" x2="118.126" y2="97.462" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F3F0E9"/></linearGradient><linearGradient id="paint28_linear_3269_5314" x1="153.717" y1="30.342" x2="160.152" y2="36.777" gradientUnits="userSpaceOnUse"><stop stop-color="#CAC2B4"/><stop offset="1" stop-color="#998E7B"/></linearGradient><linearGradient id="paint29_linear_3269_5314" x1="159.533" y1="28.079" x2="159.533" y2="36.648" gradientUnits="userSpaceOnUse"><stop stop-color="#ECE7DE"/><stop offset="1" stop-color="#D2C7B6"/></linearGradient><linearGradient id="paint30_linear_3269_5314" x1="73.447" y1="36.522" x2="73.447" y2="62.318" gradientUnits="userSpaceOnUse"><stop stop-color="#F4EDE5"/><stop offset=".996" stop-color="#DBCFBE"/></linearGradient><linearGradient id="paint31_linear_3269_5314" x1="71.646" y1="47.57" x2="65.718" y2="47.591" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint32_linear_3269_5314" x1="76.537" y1="43.228" x2="72.5" y2="43.242" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint33_linear_3269_5314" x1="78.323" y1="52.133" x2="74.286" y2="52.147" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><linearGradient id="paint34_linear_3269_5314" x1="80.799" y1="47.845" x2="78.116" y2="47.854" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><filter id="filter0_d_3269_5314" x="88.658" y="24.97" width="25.761" height="89.637" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter1_d_3269_5314" x="98.139" y="9.243" width="38.596" height="101.681" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter2_d_3269_5314" x="120.347" y="25.137" width="25.826" height="89.584" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter3_d_3269_5314" x="82.475" y="43.006" width="43.487" height="68.964" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter4_d_3269_5314" x="109.764" y="43.006" width="43.487" height="68.936" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter5_d_3269_5314" x="103.059" y="28.103" width="29.389" height="82.849" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.397708 0 0 0 0 0.47749 0 0 0 0 0.575 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter6_d_3269_5314" x="101.08" y="83.35" width="24.909" height="28.647" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter7_d_3269_5314" x="109.627" y="83.35" width="24.909" height="28.647" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter8_d_3269_5314" x="108.72" y="80.327" width="18.066" height="37.496" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.397708 0 0 0 0 0.47749 0 0 0 0 0.575 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter9_d_3269_5314" x="134.927" y="17.925" width="45.962" height="45.941" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter10_d_3269_5314" x="138.688" y="21.112" width="38.982" height="38.985" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter><filter id="filter11_d_3269_5314" x="52.855" y="33.029" width="41.2" height="41.23" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4.07"/><feGaussianBlur stdDeviation="4.07"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3269_5314"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3269_5314" result="shape"/></filter></defs></svg>`));
 export const LocationDark = (/*#__PURE__*/createFluentIllustration('LocationDark', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><g clip-path="url(#clip0_4173_35227)"><g filter="url(#filter0_d_4173_35227)"><path fill-rule="evenodd" clip-rule="evenodd" d="M115.298 24.026h.203c.695 0 1.424 0 2.158.027 3.889.134 7.674.644 11.326 1.504 26.085 6.118 45.489 29.518 45.489 57.456 0 32.58-26.407 58.987-58.987 58.987S56.5 115.593 56.5 83.013c0-18.318 8.338-34.682 21.446-45.495a1.754 1.754 0 01.179-.136c10.108-8.31 23.069-13.302 37.173-13.356zM78.302 37.281a.033.033 0 01.002-.001h-.002z" fill="url(#paint0_linear_4173_35227)"/></g><path d="M166.955 59.762c-1.855.288-2.639-.313-3.71-2.69-.523.443-1.019.861-1.672 1.436-.366-.81-.653-1.489-.967-2.168-.287-.68-.601-.81-1.254-.288-.444.366-1.28.758-1.646.549-.418-.209-.627-1.045-.679-1.646-.078-.836 0-1.202-1.228-1.228-1.332-.026-2.664-.966-4.153-1.567-.236 1.489-.471 2.952-.732 4.441-.653.157-1.306.287-1.959.47-2.064.549-3.318.157-4.572-1.62-.836-1.175-1.515-2.481-2.456-4.075 1.724-.444 3.213-.914 4.755-1.175 1.411-.236 4.128-2.352 4.493-3.736.026-.131.053-.262.053-.366-.262-1.385-.523-2.743-.81-4.285-.235 0-.732.079-1.176-.026-.836-.156-1.567-.444-1.933-1.384-.157-.392-.679-.81-1.097-.915-2.117-.496-2.587-.914-2.77-2.978-.052-.784-.026-1.567 0-2.351 0-.235.026-.444.026-.784a57.008 57.008 0 00-15.936-6.374c-.496 1.358-1.358 2.116-2.926 2.325-.836.104-1.646.992-2.299 1.672-1.254 1.358-2.351 2.821-3.526 4.284-1.777-.731-2.43-2.194-3.057-3.788-.784-1.985-.549-3.84.418-5.643.078-.13.209-.235.366-.34-.653.079-1.359.079-2.064.079h-.183a.743.743 0 01.235.261c.157.235.052.862-.157 1.071-1.019.915-2.116 1.725-3.187 2.587-1.489-1.803-2.586-2.09-4.938-1.411-.522.157-1.41.261-1.671-.026-.941-.993-1.62-.366-2.352.13-.888.601-1.698 1.28-2.638 1.803-1.411.784-1.542 3.449-.889 4.467.183.288.836.523 1.176.418 1.489-.365 2.822-1.07 3.631-2.43.34-.574.654-.522 1.307-.47.783.053 1.724-.287 2.377-.757 1.437-1.071 1.646-1.15 2.9.104.235.235.47.549.757.732.941.522.915 1.202.418 2.011-.339.55-.574 1.15-.914 1.855.549.13 1.019.235 1.541.34-.078.6-.104 1.202-.261 1.75-.548 1.96-1.907 3.553-4.075 1.777-.732-.601-1.672-.418-2.116.418-.68 1.332-1.907 1.437-3.057 1.384-2.612-.13-4.99.627-7.263 1.803-.26.13-.6.366-.81.313-1.593-.522-2.56.68-3.735 1.307a44.259 44.259 0 01-3.527 1.75c-.758.34-1.202.68-1.254 1.672-.13 1.75-.575 3.475-.915 5.33-1.436-.627-1.698-1.83-1.854-2.822-.392-2.508-1.568-2.926-3.37-3.265-3.318-.627-7.629 1.41-9.014 4.362-.548 1.176-.548 2.717-.418 4.05.105 1.175 1.542 1.776 2.9 1.332.81-.261 1.646-.287 1.829.549.105.548-.444 1.254-.81 1.855-.157.26-.549.391-.81.6-.836.68-.888 1.385 0 2.012.627.444.68.81.497 1.567-.784 3.188.757 5.67 3.945 6.506 1.227.313 2.09-.262 2.873-1.176.523-.627.915-1.541 1.986-1.358.287.052.6.104.836.235 1.515.757 2.978 1.645 4.52 2.325 1.41.627 2.795 1.123 3.761 2.481.392.55 1.176.863 1.829 1.098 2.795 1.07 4.911 2.926 6.427 5.46.992 1.672 1.907 3.396 2.821 5.12.47.862 1.045 1.385 2.064 1.62 2.168.47 4.232 1.359 5.878 2.926 1.594 1.515 1.803 3.004.627 4.885-.47.784-1.254 1.359-1.776 2.117-.627.94-1.333 1.907-1.646 2.978-.758 2.455-2.064 4.598-3.736 6.479-.732.836-1.986 1.463-3.083 1.672-1.907.339-2.952 1.567-3.919 2.978-.81 1.202-1.776 2.325-2.403 3.605-.993 1.986-2.926 2.822-4.572 3.971-2.508 1.724-4.31 4.023-5.12 6.871-.366 1.28.287 2.848.47 4.258.13.993.313 1.986.47 3.057-2.978-1.15-4.86-3.292-6.27-5.826-.548-.993-.6-2.273-.862-3.422-.183-.836-.183-1.725-.496-2.482-.862-2.064-.549-4.154-.235-6.218.182-1.202.209-2.325.078-3.579-.209-1.776.287-3.631.496-5.46.105-.888.314-1.777.392-2.665.079-.836-.235-1.489-.68-2.247-.495-.888-.548-2.116-.495-3.187.052-.679.104-1.019-.523-1.463-2.325-1.62-4.441-3.762-5.355-6.27-1.124-3.135-2.064-6.662.417-9.901 1.15-1.515 2.273-3.135 2.926-4.912.889-2.403.314-3.5-1.41-4.52-1.097-.626-1.803-1.567-1.907-2.847-.105-1.306-.706-2.246-1.672-3.056-.575-.47-.915-.94-.523-1.777.34-.783.026-1.358-.862-1.776-1.75-.81-3.631-1.646-4.05-3.762-.208-1.02-.208-2.325.288-3.187 2.351-4.206 5.33-7.995 8.726-11.39a62.737 62.737 0 013.344-3.11C64.937 49.651 57.02 65.405 57.02 83.013c0 31.716 25.707 57.423 57.423 57.423 12.54 0 24.139-4.024 33.57-10.842l.079-.679c.339-.81.653-1.62.992-2.43.653-1.437 1.28-2.9 1.986-4.285.653-1.254.47-2.534.078-3.709-.993-3.109.261-5.513 2.195-7.707.078-.078.156-.157.209-.261 1.019-1.15 2.142-2.221 2.142-3.867 0-1.306-.078-2.743-.575-3.919-.13-.313-.235-.627-.339-.94-.497-1.855-.131-3.58.679-5.303.026-.079.078-.157.104-.236a7.142 7.142 0 00.523-3.735 5.014 5.014 0 00-.157-.862c-.078-.288-.549-.758-.575-.758-1.515.966-2.56-.601-3.866-.68-1.724-.104-3.475-.391-5.147-.156-1.776.261-3.396-.078-4.702-1.045-2.404-1.777-4.65-3.814-6.897-5.8-.418-.365-.627-1.123-.706-1.698-.261-2.743.445-5.356 1.15-7.968.444-1.62.705-3.161.47-4.911-.209-1.725 1.15-2.979 2.482-3.971.496-.366 1.176-.523 1.724-.836.575-.34 1.306-.627 1.646-1.15 2.064-3.135 5.277-3.553 8.543-3.5 2.56.052 5.199.182 7.628.861 1.725.47 3.266 1.777 4.729 2.9 1.254.94 2.456 1.646 4.102 1.463.757-.078 1.541.105 2.299.21a43.54 43.54 0 00-1.855-4.86zM80.22 57.907c.366-1.489 1.698-2.168 2.665-1.437 1.254.94 2.612 1.333 4.154 1.385.366.026.705.261 1.071.418-.052.183-.078.34-.13.522-2.587-.313-5.147-.6-7.76-.888zm10.973 4.677c-1.646-.157-3.188-.314-4.86-.47 1.646-1.28 3.266-2.064 5.173-.288a38.27 38.27 0 01-.313.758zm36.992-29.077c-.209.522-.47 1.384-.836 1.463-.653.13-1.41-.157-2.063-.444-.654-.262-1.202-.732-1.803-1.124.052-.157.131-.34.183-.496 1.097-.418 2.194-.915 3.318-1.202.418-.104 1.071.21 1.437.549.156.209-.079.862-.236 1.254zm12.044 9.744c-.235-.444-.418-.862-.653-1.306 1.071-.47 1.541-1.202 1.228-2.377-.026-.105.836-.601.914-.523 1.62 1.515 3.71 2.508 4.546 4.886.418 1.175-.131 1.54-.732 2.011-1.306.993-2.769.444-3.109-1.123-.156-.653-.235-1.333-.365-2.195-.68.21-1.228.392-1.829.627z" fill="url(#paint1_linear_4173_35227)"/><path fill-rule="evenodd" clip-rule="evenodd" d="M115.298 24.026h.203c.695 0 1.424 0 2.158.027 3.889.134 7.674.644 11.326 1.504 26.085 6.118 45.489 29.518 45.489 57.456 0 32.58-26.407 58.987-58.987 58.987S56.5 115.593 56.5 83.013c0-18.318 8.338-34.682 21.446-45.495a1.754 1.754 0 01.179-.136c10.108-8.31 23.069-13.302 37.173-13.356zM78.302 37.281a.033.033 0 01.002-.001h-.002zm37.185-10.125h-.18c-13.383.051-25.67 4.795-35.245 12.685a1.66 1.66 0 01-.178.135C67.498 50.219 59.629 65.692 59.629 83.013c0 30.852 25.007 55.858 55.858 55.858 30.851 0 55.858-25.006 55.858-55.858 0-26.453-18.373-48.615-43.075-54.41h-.001a55.02 55.02 0 00-10.72-1.422h-.004a55.642 55.642 0 00-2.058-.026z" fill="#EBE9E6"/><path d="M142.734 17c5.607 0 10.142 4.63 10.142 10.356 0 4.568-6.435 13.61-9.07 17.114-.551.75-1.655.75-2.206 0-2.635-3.504-9.07-12.546-9.07-17.114.061-5.725 4.596-10.356 10.204-10.356z" fill="url(#paint2_linear_4173_35227)"/><path d="M142.477 31.469a4.07 4.07 0 100-8.139 4.07 4.07 0 000 8.139z" fill="#fff"/><path d="M91.883 26.495c8.474 0 15.327 6.722 15.327 15.033 0 6.63-9.724 19.756-13.706 24.843-.834 1.09-2.5 1.09-3.334 0-3.982-5.087-13.706-18.212-13.706-24.843.093-8.311 6.945-15.033 15.42-15.033z" fill="url(#paint3_linear_4173_35227)"/><path d="M91.836 47.746c3.247 0 5.878-2.733 5.878-6.104s-2.631-6.104-5.878-6.104c-3.246 0-5.877 2.733-5.877 6.104s2.631 6.104 5.877 6.104z" fill="#fff"/></g><defs><linearGradient id="paint0_linear_4173_35227" x1="115.449" y1="21.297" x2="115.449" y2="143.272" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_4173_35227" x1="112.879" y1="22.854" x2="112.879" y2="141.674" gradientUnits="userSpaceOnUse"><stop stop-color="#F7F3EE"/><stop offset="1" stop-color="#DBD3CA"/></linearGradient><linearGradient id="paint2_linear_4173_35227" x1="142.705" y1="17.974" x2="142.705" y2="33.709" gradientUnits="userSpaceOnUse"><stop stop-color="#B7B4AF"/><stop offset="1" stop-color="#5C5C5C"/></linearGradient><linearGradient id="paint3_linear_4173_35227" x1="91.839" y1="27.909" x2="91.839" y2="50.749" gradientUnits="userSpaceOnUse"><stop stop-color="#B7B4AF"/><stop offset="1" stop-color="#5C5C5C"/></linearGradient><clipPath id="clip0_4173_35227"><path fill="#fff" d="M0 0h230v160H0z"/></clipPath><filter id="filter0_d_4173_35227" x="40.5" y="16.026" width="149.974" height="149.974" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="8"/><feGaussianBlur stdDeviation="8"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.16 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4173_35227"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4173_35227" result="shape"/></filter></defs></svg>`));
 export const LocationLight = (/*#__PURE__*/createFluentIllustration('LocationLight', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><g clip-path="url(#clip0_4173_35229)"><g filter="url(#filter0_d_4173_35229)"><path fill-rule="evenodd" clip-rule="evenodd" d="M114.359 24.442h.203c.695 0 1.424 0 2.158.027 3.889.134 7.674.644 11.326 1.503 26.085 6.119 45.489 29.518 45.489 57.457 0 32.58-26.407 58.987-58.987 58.987s-58.987-26.407-58.987-58.987c0-18.318 8.338-34.682 21.446-45.496a1.76 1.76 0 01.179-.135c10.108-8.31 23.069-13.303 37.173-13.356zM77.363 37.696c.004-.001.005-.002.002 0h-.002z" fill="#3A3A3A"/></g><path d="M166.016 60.178c-1.855.287-2.639-.314-3.71-2.691-.523.444-1.019.862-1.672 1.437-.366-.81-.653-1.49-.967-2.169-.287-.679-.601-.81-1.254-.287-.444.366-1.28.758-1.645.549-.418-.21-.627-1.045-.68-1.646-.078-.836 0-1.202-1.228-1.228-1.332-.026-2.664-.967-4.153-1.567-.236 1.489-.471 2.952-.732 4.44-.653.157-1.306.288-1.959.471-2.064.549-3.318.157-4.572-1.62-.836-1.175-1.515-2.482-2.456-4.075 1.724-.444 3.213-.915 4.755-1.176 1.411-.235 4.128-2.351 4.493-3.736.026-.13.053-.26.053-.365-.262-1.385-.523-2.743-.81-4.285-.235 0-.732.079-1.176-.026-.836-.157-1.567-.444-1.933-1.385-.157-.392-.679-.81-1.097-.914-2.117-.496-2.587-.914-2.77-2.978-.052-.784-.026-1.568 0-2.351 0-.236.026-.445.026-.784a57.005 57.005 0 00-15.936-6.375c-.496 1.359-1.358 2.116-2.926 2.325-.836.105-1.645.993-2.299 1.672-1.254 1.359-2.351 2.822-3.526 4.285-1.777-.732-2.43-2.195-3.057-3.788-.784-1.986-.549-3.84.418-5.643.078-.13.209-.235.366-.34-.653.079-1.359.079-2.064.079h-.183a.743.743 0 01.235.26c.157.236.052.863-.157 1.072-1.019.914-2.116 1.724-3.187 2.586-1.489-1.802-2.586-2.09-4.937-1.41-.523.156-1.411.26-1.672-.027-.941-.992-1.62-.365-2.352.131-.888.601-1.698 1.28-2.638 1.803-1.411.784-1.542 3.448-.889 4.467.183.287.836.523 1.176.418 1.489-.366 2.822-1.071 3.631-2.43.34-.574.654-.522 1.307-.47.783.053 1.724-.287 2.377-.757 1.437-1.072 1.646-1.15 2.9.104.235.235.47.549.757.732.941.522.915 1.201.418 2.011-.339.549-.574 1.15-.914 1.855.549.13 1.019.235 1.541.34-.078.6-.104 1.201-.261 1.75-.548 1.96-1.907 3.553-4.075 1.776-.732-.6-1.672-.418-2.116.418-.68 1.333-1.907 1.437-3.057 1.385-2.612-.13-4.99.627-7.263 1.803-.26.13-.6.366-.81.313-1.593-.522-2.56.68-3.735 1.306a44.224 44.224 0 01-3.527 1.75c-.758.34-1.202.68-1.254 1.673-.13 1.75-.575 3.474-.915 5.33-1.436-.628-1.698-1.83-1.854-2.822-.392-2.508-1.568-2.926-3.37-3.266-3.318-.627-7.629 1.411-9.014 4.363-.548 1.176-.548 2.717-.418 4.05.105 1.175 1.542 1.776 2.9 1.332.81-.261 1.646-.288 1.829.548.105.549-.444 1.254-.81 1.855-.157.261-.549.392-.81.601-.836.68-.888 1.385 0 2.012.627.444.68.81.497 1.567-.784 3.188.757 5.67 3.944 6.505 1.228.314 2.09-.261 2.874-1.175.523-.627.915-1.542 1.986-1.359.287.052.6.105.836.235 1.515.758 2.978 1.646 4.52 2.325 1.41.627 2.795 1.124 3.761 2.482.392.549 1.176.862 1.829 1.097 2.795 1.072 4.911 2.926 6.427 5.46.992 1.673 1.907 3.397 2.821 5.121.47.862 1.045 1.385 2.064 1.62 2.168.47 4.232 1.358 5.878 2.926 1.594 1.515 1.803 3.004.627 4.885-.47.784-1.254 1.359-1.776 2.116-.627.94-1.333 1.907-1.646 2.978-.758 2.456-2.064 4.598-3.736 6.479-.732.836-1.986 1.463-3.083 1.672-1.907.34-2.952 1.568-3.918 2.979-.81 1.201-1.777 2.325-2.404 3.605-.993 1.985-2.926 2.821-4.572 3.971-2.508 1.724-4.31 4.023-5.12 6.871-.366 1.28.287 2.847.47 4.258.13.993.313 1.985.47 3.057-2.978-1.15-4.859-3.292-6.27-5.826-.548-.993-.6-2.273-.862-3.423-.183-.836-.183-1.724-.496-2.481-.862-2.064-.549-4.154-.235-6.218.183-1.202.209-2.325.078-3.579-.209-1.777.287-3.632.496-5.46.105-.889.314-1.777.392-2.665.079-.836-.235-1.489-.679-2.247-.496-.888-.549-2.116-.496-3.187.052-.68.104-1.019-.523-1.463-2.325-1.62-4.441-3.762-5.356-6.27-1.123-3.135-2.063-6.662.418-9.901 1.15-1.516 2.273-3.135 2.927-4.912.888-2.403.313-3.5-1.411-4.52-1.097-.626-1.803-1.567-1.907-2.847-.105-1.306-.706-2.247-1.672-3.056-.575-.47-.915-.941-.523-1.777.34-.784.026-1.359-.862-1.777-1.75-.81-3.631-1.645-4.05-3.762-.208-1.018-.208-2.325.288-3.187 2.351-4.206 5.33-7.994 8.726-11.39a62.817 62.817 0 013.344-3.109c-12.436 10.58-20.351 26.334-20.351 43.942 0 31.715 25.706 57.422 57.422 57.422 12.54 0 24.139-4.023 33.57-10.842l.079-.679c.339-.81.653-1.619.992-2.429.653-1.437 1.28-2.9 1.986-4.285.653-1.254.47-2.534.078-3.71-.993-3.108.261-5.512 2.195-7.706.078-.079.156-.157.209-.262 1.019-1.149 2.142-2.22 2.142-3.866 0-1.306-.078-2.743-.575-3.919-.13-.313-.235-.627-.339-.94-.497-1.855-.131-3.58.679-5.304.026-.078.078-.156.104-.235a7.131 7.131 0 00.523-3.736 5.01 5.01 0 00-.157-.862c-.078-.287-.549-.757-.575-.757-1.515.966-2.56-.601-3.866-.68-1.724-.104-3.475-.391-5.147-.156-1.776.26-3.396-.079-4.702-1.045-2.404-1.777-4.65-3.815-6.897-5.8-.418-.366-.627-1.123-.706-1.698-.261-2.743.445-5.356 1.15-7.968.444-1.62.705-3.161.47-4.912-.209-1.724 1.15-2.978 2.482-3.97.496-.366 1.176-.523 1.724-.837.575-.34 1.307-.627 1.646-1.15 2.064-3.134 5.277-3.552 8.543-3.5 2.56.052 5.199.183 7.628.862 1.725.47 3.266 1.777 4.729 2.9 1.254.94 2.456 1.646 4.102 1.463.757-.078 1.541.105 2.299.21a43.538 43.538 0 00-1.855-4.86zm-86.735-1.855c.366-1.49 1.698-2.168 2.665-1.437 1.254.94 2.613 1.332 4.154 1.385.366.026.705.261 1.071.418-.052.183-.078.34-.13.522-2.587-.313-5.147-.6-7.76-.888zm10.973 4.676c-1.646-.156-3.188-.313-4.86-.47 1.646-1.28 3.266-2.064 5.173-.287-.104.261-.235.575-.313.757zm36.992-29.077c-.209.523-.47 1.385-.836 1.463-.653.131-1.41-.156-2.063-.444-.653-.26-1.202-.731-1.803-1.123.052-.157.131-.34.183-.496 1.097-.418 2.194-.915 3.318-1.202.418-.105 1.071.209 1.437.549.156.209-.079.862-.236 1.253zm12.044 9.745c-.235-.444-.418-.862-.653-1.306 1.071-.47 1.541-1.202 1.228-2.378-.026-.104.836-.6.914-.522 1.62 1.515 3.71 2.508 4.546 4.885.418 1.176-.131 1.542-.732 2.012-1.306.993-2.769.444-3.109-1.123-.156-.654-.235-1.333-.365-2.195-.679.209-1.228.392-1.829.627z" fill="url(#paint0_linear_4173_35229)"/><path fill-rule="evenodd" clip-rule="evenodd" d="M114.359 24.442h.203c.695 0 1.424 0 2.158.027 3.889.134 7.674.644 11.326 1.503 26.085 6.119 45.489 29.518 45.489 57.457 0 32.58-26.407 58.987-58.987 58.987s-58.987-26.407-58.987-58.987c0-18.318 8.338-34.682 21.446-45.496a1.76 1.76 0 01.179-.135c10.108-8.31 23.069-13.303 37.173-13.356zM77.363 37.696c.004-.001.005-.002.002 0h-.002zm37.185-10.125h-.18c-13.383.052-25.67 4.796-35.245 12.686a1.674 1.674 0 01-.178.135C66.559 50.635 58.69 66.106 58.69 83.428c0 30.851 25.007 55.858 55.858 55.858 30.851 0 55.858-25.007 55.858-55.858 0-26.453-18.373-48.616-43.074-54.41h-.002a55.028 55.028 0 00-10.72-1.422h-.004a55.866 55.866 0 00-2.058-.026z" fill="#5C5C5C"/><path d="M141.795 17.416c5.607 0 10.142 4.63 10.142 10.356 0 4.568-6.435 13.61-9.07 17.114-.551.75-1.654.75-2.206 0-2.635-3.505-9.07-12.546-9.07-17.114.061-5.726 4.596-10.356 10.204-10.356z" fill="#D1D1D1"/><path d="M141.538 31.884a4.07 4.07 0 100-8.138 4.07 4.07 0 000 8.139z" fill="#3A3A3A"/><path d="M90.944 26.91c8.474 0 15.327 6.723 15.327 15.034 0 6.63-9.724 19.756-13.706 24.842-.834 1.09-2.5 1.09-3.334 0-3.982-5.086-13.706-18.212-13.706-24.842.093-8.312 6.945-15.033 15.42-15.033z" fill="#D1D1D1"/><path d="M90.897 48.162c3.247 0 5.878-2.733 5.878-6.104s-2.631-6.104-5.878-6.104c-3.246 0-5.878 2.733-5.878 6.104 0 3.37 2.632 6.104 5.878 6.104z" fill="#3A3A3A"/></g><defs><linearGradient id="paint0_linear_4173_35229" x1="111.94" y1="23.27" x2="111.94" y2="142.09" gradientUnits="userSpaceOnUse"><stop stop-color="#6C6C6C"/><stop offset=".996" stop-color="#464646"/></linearGradient><clipPath id="clip0_4173_35229"><path fill="#fff" d="M0 0h230v160H0z"/></clipPath><filter id="filter0_d_4173_35229" x="39.561" y="19.916" width="149.974" height="149.974" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11.475"/><feGaussianBlur stdDeviation="8"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.16 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4173_35229"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4173_35229" result="shape"/></filter></defs></svg>`));
-export const LoginDark = (/*#__PURE__*/createFluentIllustration('LoginDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.31 144.8c37.64 0 68.105-30.645 68.105-68.455S152.95 8 115.31 8C77.78 8 47.205 38.645 47.205 76.455S77.781 144.8 115.31 144.8z" fill="#242424"/><path d="M37.381 92.365c2.539 0 4.526-1.985 4.526-4.52 0-2.535-1.987-4.52-4.526-4.52-2.539 0-4.526 1.985-4.526 4.52 0 2.535 2.098 4.52 4.526 4.52zM32.856 77.704c1.655 0 3.09-1.433 3.09-3.087 0-1.653-1.434-3.087-3.09-3.087s-3.09 1.434-3.09 3.087c0 1.764 1.434 3.087 3.09 3.087zM194.785 103.718a5.697 5.697 0 005.74-5.732 5.697 5.697 0 00-5.74-5.732 5.697 5.697 0 00-5.74 5.732 5.697 5.697 0 005.74 5.732z" fill="#3A3A3A"/><circle cx="113.767" cy="40.837" r="2.662" fill="#C4C4C4"/><path d="M165.189 133.474H65.247c-3.724 0-6.828-3.077-6.828-6.77V25.759c0-3.693 3.104-6.77 6.828-6.77h99.942c3.724 0 6.828 3.077 6.828 6.77v100.945c0 3.693-3.104 6.77-6.828 6.77z" fill="url(#paint0_linear_3840_873)"/><path d="M114.076 56.812c7.842 0 14.2-6.358 14.2-14.2s-6.358-14.2-14.2-14.2c-7.843 0-14.2 6.358-14.2 14.2s6.357 14.2 14.2 14.2z" fill="#3A3A3A"/><path d="M114.076 43.5a5.325 5.325 0 100-10.65 5.325 5.325 0 000 10.65zM124.726 52.379c-2.575 2.732-6.196 4.433-10.206 4.433-4.011 0-7.632-1.701-10.207-4.433 1.422-4.128 5.454-7.105 10.207-7.105 4.752 0 8.784 2.977 10.206 7.105z" fill="#CACACA"/><path d="M152.749 78.999H75.673c-.922 0-1.843-1.024-1.843-2.048v-9.216c0-1.025.921-2.049 1.843-2.049h77.076c.921 0 1.842 1.024 1.842 2.049v9.216c0 1.024-.921 2.048-1.842 2.048zM152.749 98.524H75.673c-.922 0-1.843-1.024-1.843-2.048V87.26c0-1.025.921-2.049 1.843-2.049h77.076c.921 0 1.842 1.024 1.842 2.049v9.216c0 1.024-.921 2.048-1.842 2.048zM152.675 106.511h-28.116c-.958 0-1.917.731-1.917 2.193v8.039c0 1.096.639 2.193 1.917 2.193h28.116c.958 0 1.917-.731 1.917-2.193v-8.039c0-1.158-.959-2.193-1.917-2.193z" fill="#3A3A3A"/><path d="M146.266 111.836h-14.41c-.613 0-1.226.444-1.226.887 0 .444.613.888 1.226.888h14.41c.613 0 1.226-.444 1.226-.888 0-.665-.613-.887-1.226-.887zM79.155 71.011h34.612v2.662H79.155z" fill="#777"/><rect x="79.155" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><rect x="84.48" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><rect x="89.805" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><rect x="95.13" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><rect x="99.567" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><rect x="104.892" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><defs><linearGradient id="paint0_linear_3840_873" x1="115.181" y1="16.34" x2="115.181" y2="134.709" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient></defs></svg>`));
-export const LoginLight = (/*#__PURE__*/createFluentIllustration('LoginLight', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M116.31 144.8c37.64 0 68.105-30.645 68.105-68.455S153.95 8 116.31 8C78.78 8 48.205 38.645 48.205 76.455S78.781 144.8 116.31 144.8z" fill="#EBEBEB"/><path d="M38.381 92.365c2.539 0 4.526-1.985 4.526-4.52 0-2.535-1.987-4.52-4.526-4.52-2.539 0-4.526 1.985-4.526 4.52 0 2.535 2.098 4.52 4.526 4.52zM33.856 77.704c1.655 0 3.09-1.433 3.09-3.087 0-1.653-1.434-3.087-3.09-3.087s-3.09 1.434-3.09 3.087c0 1.764 1.434 3.087 3.09 3.087zM195.785 103.719a5.697 5.697 0 005.74-5.733 5.697 5.697 0 00-5.74-5.732 5.697 5.697 0 00-5.74 5.732 5.697 5.697 0 005.74 5.733z" fill="#EDE5DA"/><circle cx="114.767" cy="40.837" r="2.662" fill="#C4C4C4"/><g filter="url(#filter0_d_3840_855)"><path d="M166.189 133.474H66.247c-3.724 0-6.828-3.077-6.828-6.771V25.759c0-3.693 3.104-6.77 6.828-6.77h99.942c3.724 0 6.828 3.077 6.828 6.77v100.944c0 3.694-3.104 6.771-6.828 6.771z" fill="url(#paint0_linear_3840_855)"/></g><path d="M115.076 56.812c7.842 0 14.2-6.358 14.2-14.2s-6.358-14.2-14.2-14.2c-7.843 0-14.2 6.358-14.2 14.2s6.357 14.2 14.2 14.2z" fill="#EDE5DA"/><path d="M115.076 43.5a5.325 5.325 0 100-10.65 5.325 5.325 0 000 10.65zM125.726 52.379c-2.575 2.732-6.196 4.433-10.206 4.433-4.011 0-7.632-1.701-10.207-4.433 1.422-4.128 5.454-7.105 10.207-7.105 4.752 0 8.784 2.977 10.206 7.105z" fill="#BCB9B3"/><path d="M153.749 78.999H76.673c-.922 0-1.843-1.024-1.843-2.048v-9.216c0-1.025.921-2.049 1.843-2.049h77.076c.921 0 1.842 1.024 1.842 2.049v9.216c0 1.024-.921 2.048-1.842 2.048zM153.749 98.524H76.673c-.922 0-1.843-1.024-1.843-2.048V87.26c0-1.025.921-2.049 1.843-2.049h77.076c.921 0 1.842 1.024 1.842 2.049v9.216c0 1.024-.921 2.048-1.842 2.048zM153.675 106.511h-28.116c-.958 0-1.917.731-1.917 2.193v8.039c0 1.096.639 2.193 1.917 2.193h28.116c.958 0 1.917-.731 1.917-2.193v-8.039c0-1.158-.959-2.193-1.917-2.193z" fill="#EDE5DA"/><path d="M147.266 111.836h-14.41c-.613 0-1.226.444-1.226.888 0 .443.613.887 1.226.887h14.41c.613 0 1.226-.444 1.226-.887 0-.666-.613-.888-1.226-.888zM80.155 71.011h34.612v2.662H80.155z" fill="#BCB9B3"/><rect x="80.155" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><rect x="85.48" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><rect x="90.805" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><rect x="96.13" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><rect x="100.567" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><rect x="105.892" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><defs><linearGradient id="paint0_linear_3840_855" x1="116.181" y1="16.34" x2="116.181" y2="134.708" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><filter id="filter0_d_3840_855" x="48.619" y="15.388" width="135.198" height="136.086" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="7.2"/><feGaussianBlur stdDeviation="5.4"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.17 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_855"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_855" result="shape"/></filter></defs></svg>`));
-export const MediaDark = (/*#__PURE__*/createFluentIllustration('MediaDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.612 147.8c37.749 0 68.351-30.624 68.351-68.4 0-37.776-30.602-68.4-68.351-68.4-37.749 0-68.35 30.624-68.35 68.4 0 37.776 30.601 68.4 68.35 68.4z" fill="#242424"/><path d="M178.664 38.646v83.629c0 2.948-2.42 5.376-5.359 5.376H58.978c-2.938 0-5.358-2.428-5.358-5.376V38.646c0-2.948 2.42-5.376 5.358-5.376h114.327c2.939 0 5.359 2.428 5.359 5.376z" fill="url(#paint0_linear_3831_4976)"/><path d="M139.639 73.46c-.665-2.458-2.216-4.581-4.433-5.922-2.438-1.452-4.987-1.676-7.758-1.9-4.876-.558-9.863-.67-14.85-.446-4.433.112-9.531-.112-13.632 1.788-5.873 2.57-6.65 8.267-6.87 14.412-.222 4.58-.222 10.055.997 14.524.997 3.575 3.768 6.368 7.425 7.15 4.655 1.118 14.629 1.341 19.505 1.229h1.33c3.879-.111 7.979-.111 11.526-1.452 2.992-1.006 4.765-2.793 5.874-5.028 1.329-2.681 1.662-6.033 1.773-9.72.443-4.804.332-10.278-.887-14.635z" fill="url(#paint1_linear_3831_4976)"/><path d="M112.131 78.3c0-1.474.745-1.965 1.955-1.179 3.257 2.064 6.607 4.226 9.864 6.29 1.209.786 1.209 1.67 0 2.555-3.257 2.064-6.514 4.226-9.771 6.29-1.303.785-1.955.392-2.048-.984V78.3z" fill="#656565"/><path d="M178.664 39.068v3.882H53.62v-3.882c0-2.932 2.42-5.347 5.358-5.347h114.328c2.938 0 5.358 2.415 5.358 5.347z" fill="#3A3A3A"/><path d="M58.866 40.017a1.638 1.638 0 100-3.277 1.638 1.638 0 000 3.277z" fill="#CCC"/><path opacity=".6" d="M63.952 40.017a1.638 1.638 0 10-.001-3.277 1.638 1.638 0 000 3.277z" fill="#5F5F5F"/><path opacity=".4" d="M69.123 40.017a1.638 1.638 0 100-3.277 1.638 1.638 0 000 3.277z" fill="#5C5C5C"/><defs><linearGradient id="paint0_linear_3831_4976" x1="116.101" y1="31.087" x2="116.101" y2="128.669" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint1_linear_3831_4976" x1="116.373" y1="65.084" x2="116.373" y2="104.321" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient></defs></svg>`));
-export const MediaErrorDark = (/*#__PURE__*/createFluentIllustration('MediaErrorDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.613 147.8c37.749 0 68.35-30.624 68.35-68.4 0-37.776-30.601-68.4-68.35-68.4-37.75 0-68.351 30.624-68.351 68.4 0 37.776 30.601 68.4 68.351 68.4z" fill="#242424"/><path d="M178.665 38.646v83.629c0 2.948-2.42 5.376-5.359 5.376H58.979c-2.939 0-5.359-2.428-5.359-5.376V38.646c0-2.948 2.42-5.376 5.359-5.376h114.327c2.939 0 5.359 2.428 5.359 5.376z" fill="url(#paint0_linear_4732_428)"/><path d="M139.64 73.46c-.665-2.458-2.217-4.581-4.433-5.922-2.438-1.452-4.987-1.676-7.758-1.9-4.876-.558-9.863-.67-14.85-.446-4.433.112-9.531-.112-13.632 1.788-5.873 2.57-6.65 8.267-6.87 14.412-.222 4.58-.222 10.055.997 14.524.997 3.575 3.768 6.368 7.425 7.15 4.654 1.118 14.629 1.341 19.505 1.229h1.33c3.879-.111 7.979-.111 11.526-1.452 2.992-1.006 4.765-2.793 5.873-5.028 1.33-2.681 1.663-6.033 1.774-9.72.443-4.804.332-10.278-.887-14.635z" fill="url(#paint1_linear_4732_428)"/><path d="M112.132 78.3c0-1.474.745-1.965 1.955-1.179 3.257 2.064 6.607 4.226 9.864 6.29 1.209.786 1.209 1.67 0 2.555-3.257 2.064-6.514 4.226-9.771 6.29-1.303.785-1.955.392-2.048-.984V78.3z" fill="#656565"/><path d="M178.665 39.068v3.882H53.62v-3.882c0-2.932 2.42-5.347 5.359-5.347h114.327c2.939 0 5.359 2.415 5.359 5.347z" fill="#3A3A3A"/><path d="M58.867 40.017a1.638 1.638 0 10-.002-3.277 1.638 1.638 0 00.002 3.277z" fill="#CCC"/><path opacity=".6" d="M63.952 40.017a1.638 1.638 0 10-.001-3.277 1.638 1.638 0 00.001 3.277z" fill="#5F5F5F"/><path opacity=".4" d="M69.124 40.017a1.638 1.638 0 10-.001-3.277 1.638 1.638 0 00.001 3.277z" fill="#5C5C5C"/><path d="M166.047 100.559l-19.458 28.384c-2.493 3.637.363 8.364 5.055 8.364h38.916c4.692 0 7.548-4.727 5.055-8.364l-19.458-28.384c-2.339-3.412-7.771-3.412-10.11 0z" fill="url(#paint2_linear_4732_428)"/><path d="M171.236 110.098a2.221 2.221 0 00-2.218 2.345l.552 9.86a1.67 1.67 0 003.332 0l.552-9.86a2.221 2.221 0 00-2.218-2.345zm-.007 20.356c1.192 0 2.123-.964 2.123-2.172 0-1.209-.931-2.157-2.123-2.157-1.178 0-2.109.948-2.109 2.157 0 1.208.931 2.172 2.109 2.172z" fill="#242424"/><defs><linearGradient id="paint0_linear_4732_428" x1="116.102" y1="31.087" x2="116.102" y2="128.669" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint1_linear_4732_428" x1="116.374" y1="65.084" x2="116.374" y2="104.321" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint2_linear_4732_428" x1="151.627" y1="146.992" x2="187.23" y2="108.187" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient></defs></svg>`));
