+++ illustrations/src/illustrations/chunk-1.tsx
@@ -1,4 +1,8 @@
 import { createFluentIllustration } from "../utils/createFluentIllustration";
+export const LoginDark = (/*#__PURE__*/createFluentIllustration('LoginDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.31 144.8c37.64 0 68.105-30.645 68.105-68.455S152.95 8 115.31 8C77.78 8 47.205 38.645 47.205 76.455S77.781 144.8 115.31 144.8z" fill="#242424"/><path d="M37.381 92.365c2.539 0 4.526-1.985 4.526-4.52 0-2.535-1.987-4.52-4.526-4.52-2.539 0-4.526 1.985-4.526 4.52 0 2.535 2.098 4.52 4.526 4.52zM32.856 77.704c1.655 0 3.09-1.433 3.09-3.087 0-1.653-1.434-3.087-3.09-3.087s-3.09 1.434-3.09 3.087c0 1.764 1.434 3.087 3.09 3.087zM194.785 103.718a5.697 5.697 0 005.74-5.732 5.697 5.697 0 00-5.74-5.732 5.697 5.697 0 00-5.74 5.732 5.697 5.697 0 005.74 5.732z" fill="#3A3A3A"/><circle cx="113.767" cy="40.837" r="2.662" fill="#C4C4C4"/><path d="M165.189 133.474H65.247c-3.724 0-6.828-3.077-6.828-6.77V25.759c0-3.693 3.104-6.77 6.828-6.77h99.942c3.724 0 6.828 3.077 6.828 6.77v100.945c0 3.693-3.104 6.77-6.828 6.77z" fill="url(#paint0_linear_3840_873)"/><path d="M114.076 56.812c7.842 0 14.2-6.358 14.2-14.2s-6.358-14.2-14.2-14.2c-7.843 0-14.2 6.358-14.2 14.2s6.357 14.2 14.2 14.2z" fill="#3A3A3A"/><path d="M114.076 43.5a5.325 5.325 0 100-10.65 5.325 5.325 0 000 10.65zM124.726 52.379c-2.575 2.732-6.196 4.433-10.206 4.433-4.011 0-7.632-1.701-10.207-4.433 1.422-4.128 5.454-7.105 10.207-7.105 4.752 0 8.784 2.977 10.206 7.105z" fill="#CACACA"/><path d="M152.749 78.999H75.673c-.922 0-1.843-1.024-1.843-2.048v-9.216c0-1.025.921-2.049 1.843-2.049h77.076c.921 0 1.842 1.024 1.842 2.049v9.216c0 1.024-.921 2.048-1.842 2.048zM152.749 98.524H75.673c-.922 0-1.843-1.024-1.843-2.048V87.26c0-1.025.921-2.049 1.843-2.049h77.076c.921 0 1.842 1.024 1.842 2.049v9.216c0 1.024-.921 2.048-1.842 2.048zM152.675 106.511h-28.116c-.958 0-1.917.731-1.917 2.193v8.039c0 1.096.639 2.193 1.917 2.193h28.116c.958 0 1.917-.731 1.917-2.193v-8.039c0-1.158-.959-2.193-1.917-2.193z" fill="#3A3A3A"/><path d="M146.266 111.836h-14.41c-.613 0-1.226.444-1.226.887 0 .444.613.888 1.226.888h14.41c.613 0 1.226-.444 1.226-.888 0-.665-.613-.887-1.226-.887zM79.155 71.011h34.612v2.662H79.155z" fill="#777"/><rect x="79.155" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><rect x="84.48" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><rect x="89.805" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><rect x="95.13" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><rect x="99.567" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><rect x="104.892" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#777"/><defs><linearGradient id="paint0_linear_3840_873" x1="115.181" y1="16.34" x2="115.181" y2="134.709" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient></defs></svg>`));
+export const LoginLight = (/*#__PURE__*/createFluentIllustration('LoginLight', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M116.31 144.8c37.64 0 68.105-30.645 68.105-68.455S153.95 8 116.31 8C78.78 8 48.205 38.645 48.205 76.455S78.781 144.8 116.31 144.8z" fill="#EBEBEB"/><path d="M38.381 92.365c2.539 0 4.526-1.985 4.526-4.52 0-2.535-1.987-4.52-4.526-4.52-2.539 0-4.526 1.985-4.526 4.52 0 2.535 2.098 4.52 4.526 4.52zM33.856 77.704c1.655 0 3.09-1.433 3.09-3.087 0-1.653-1.434-3.087-3.09-3.087s-3.09 1.434-3.09 3.087c0 1.764 1.434 3.087 3.09 3.087zM195.785 103.719a5.697 5.697 0 005.74-5.733 5.697 5.697 0 00-5.74-5.732 5.697 5.697 0 00-5.74 5.732 5.697 5.697 0 005.74 5.733z" fill="#EDE5DA"/><circle cx="114.767" cy="40.837" r="2.662" fill="#C4C4C4"/><g filter="url(#filter0_d_3840_855)"><path d="M166.189 133.474H66.247c-3.724 0-6.828-3.077-6.828-6.771V25.759c0-3.693 3.104-6.77 6.828-6.77h99.942c3.724 0 6.828 3.077 6.828 6.77v100.944c0 3.694-3.104 6.771-6.828 6.771z" fill="url(#paint0_linear_3840_855)"/></g><path d="M115.076 56.812c7.842 0 14.2-6.358 14.2-14.2s-6.358-14.2-14.2-14.2c-7.843 0-14.2 6.358-14.2 14.2s6.357 14.2 14.2 14.2z" fill="#EDE5DA"/><path d="M115.076 43.5a5.325 5.325 0 100-10.65 5.325 5.325 0 000 10.65zM125.726 52.379c-2.575 2.732-6.196 4.433-10.206 4.433-4.011 0-7.632-1.701-10.207-4.433 1.422-4.128 5.454-7.105 10.207-7.105 4.752 0 8.784 2.977 10.206 7.105z" fill="#BCB9B3"/><path d="M153.749 78.999H76.673c-.922 0-1.843-1.024-1.843-2.048v-9.216c0-1.025.921-2.049 1.843-2.049h77.076c.921 0 1.842 1.024 1.842 2.049v9.216c0 1.024-.921 2.048-1.842 2.048zM153.749 98.524H76.673c-.922 0-1.843-1.024-1.843-2.048V87.26c0-1.025.921-2.049 1.843-2.049h77.076c.921 0 1.842 1.024 1.842 2.049v9.216c0 1.024-.921 2.048-1.842 2.048zM153.675 106.511h-28.116c-.958 0-1.917.731-1.917 2.193v8.039c0 1.096.639 2.193 1.917 2.193h28.116c.958 0 1.917-.731 1.917-2.193v-8.039c0-1.158-.959-2.193-1.917-2.193z" fill="#EDE5DA"/><path d="M147.266 111.836h-14.41c-.613 0-1.226.444-1.226.888 0 .443.613.887 1.226.887h14.41c.613 0 1.226-.444 1.226-.887 0-.666-.613-.888-1.226-.888zM80.155 71.011h34.612v2.662H80.155z" fill="#BCB9B3"/><rect x="80.155" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><rect x="85.48" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><rect x="90.805" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><rect x="96.13" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><rect x="100.567" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><rect x="105.892" y="90.536" width="2.662" height="2.662" rx="1.331" fill="#BCB9B3"/><defs><linearGradient id="paint0_linear_3840_855" x1="116.181" y1="16.34" x2="116.181" y2="134.708" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><filter id="filter0_d_3840_855" x="48.619" y="15.388" width="135.198" height="136.086" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="7.2"/><feGaussianBlur stdDeviation="5.4"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.17 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_855"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_855" result="shape"/></filter></defs></svg>`));
+export const MediaDark = (/*#__PURE__*/createFluentIllustration('MediaDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.612 147.8c37.749 0 68.351-30.624 68.351-68.4 0-37.776-30.602-68.4-68.351-68.4-37.749 0-68.35 30.624-68.35 68.4 0 37.776 30.601 68.4 68.35 68.4z" fill="#242424"/><path d="M178.664 38.646v83.629c0 2.948-2.42 5.376-5.359 5.376H58.978c-2.938 0-5.358-2.428-5.358-5.376V38.646c0-2.948 2.42-5.376 5.358-5.376h114.327c2.939 0 5.359 2.428 5.359 5.376z" fill="url(#paint0_linear_3831_4976)"/><path d="M139.639 73.46c-.665-2.458-2.216-4.581-4.433-5.922-2.438-1.452-4.987-1.676-7.758-1.9-4.876-.558-9.863-.67-14.85-.446-4.433.112-9.531-.112-13.632 1.788-5.873 2.57-6.65 8.267-6.87 14.412-.222 4.58-.222 10.055.997 14.524.997 3.575 3.768 6.368 7.425 7.15 4.655 1.118 14.629 1.341 19.505 1.229h1.33c3.879-.111 7.979-.111 11.526-1.452 2.992-1.006 4.765-2.793 5.874-5.028 1.329-2.681 1.662-6.033 1.773-9.72.443-4.804.332-10.278-.887-14.635z" fill="url(#paint1_linear_3831_4976)"/><path d="M112.131 78.3c0-1.474.745-1.965 1.955-1.179 3.257 2.064 6.607 4.226 9.864 6.29 1.209.786 1.209 1.67 0 2.555-3.257 2.064-6.514 4.226-9.771 6.29-1.303.785-1.955.392-2.048-.984V78.3z" fill="#656565"/><path d="M178.664 39.068v3.882H53.62v-3.882c0-2.932 2.42-5.347 5.358-5.347h114.328c2.938 0 5.358 2.415 5.358 5.347z" fill="#3A3A3A"/><path d="M58.866 40.017a1.638 1.638 0 100-3.277 1.638 1.638 0 000 3.277z" fill="#CCC"/><path opacity=".6" d="M63.952 40.017a1.638 1.638 0 10-.001-3.277 1.638 1.638 0 000 3.277z" fill="#5F5F5F"/><path opacity=".4" d="M69.123 40.017a1.638 1.638 0 100-3.277 1.638 1.638 0 000 3.277z" fill="#5C5C5C"/><defs><linearGradient id="paint0_linear_3831_4976" x1="116.101" y1="31.087" x2="116.101" y2="128.669" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint1_linear_3831_4976" x1="116.373" y1="65.084" x2="116.373" y2="104.321" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient></defs></svg>`));
+export const MediaErrorDark = (/*#__PURE__*/createFluentIllustration('MediaErrorDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.613 147.8c37.749 0 68.35-30.624 68.35-68.4 0-37.776-30.601-68.4-68.35-68.4-37.75 0-68.351 30.624-68.351 68.4 0 37.776 30.601 68.4 68.351 68.4z" fill="#242424"/><path d="M178.665 38.646v83.629c0 2.948-2.42 5.376-5.359 5.376H58.979c-2.939 0-5.359-2.428-5.359-5.376V38.646c0-2.948 2.42-5.376 5.359-5.376h114.327c2.939 0 5.359 2.428 5.359 5.376z" fill="url(#paint0_linear_4732_428)"/><path d="M139.64 73.46c-.665-2.458-2.217-4.581-4.433-5.922-2.438-1.452-4.987-1.676-7.758-1.9-4.876-.558-9.863-.67-14.85-.446-4.433.112-9.531-.112-13.632 1.788-5.873 2.57-6.65 8.267-6.87 14.412-.222 4.58-.222 10.055.997 14.524.997 3.575 3.768 6.368 7.425 7.15 4.654 1.118 14.629 1.341 19.505 1.229h1.33c3.879-.111 7.979-.111 11.526-1.452 2.992-1.006 4.765-2.793 5.873-5.028 1.33-2.681 1.663-6.033 1.774-9.72.443-4.804.332-10.278-.887-14.635z" fill="url(#paint1_linear_4732_428)"/><path d="M112.132 78.3c0-1.474.745-1.965 1.955-1.179 3.257 2.064 6.607 4.226 9.864 6.29 1.209.786 1.209 1.67 0 2.555-3.257 2.064-6.514 4.226-9.771 6.29-1.303.785-1.955.392-2.048-.984V78.3z" fill="#656565"/><path d="M178.665 39.068v3.882H53.62v-3.882c0-2.932 2.42-5.347 5.359-5.347h114.327c2.939 0 5.359 2.415 5.359 5.347z" fill="#3A3A3A"/><path d="M58.867 40.017a1.638 1.638 0 10-.002-3.277 1.638 1.638 0 00.002 3.277z" fill="#CCC"/><path opacity=".6" d="M63.952 40.017a1.638 1.638 0 10-.001-3.277 1.638 1.638 0 00.001 3.277z" fill="#5F5F5F"/><path opacity=".4" d="M69.124 40.017a1.638 1.638 0 10-.001-3.277 1.638 1.638 0 00.001 3.277z" fill="#5C5C5C"/><path d="M166.047 100.559l-19.458 28.384c-2.493 3.637.363 8.364 5.055 8.364h38.916c4.692 0 7.548-4.727 5.055-8.364l-19.458-28.384c-2.339-3.412-7.771-3.412-10.11 0z" fill="url(#paint2_linear_4732_428)"/><path d="M171.236 110.098a2.221 2.221 0 00-2.218 2.345l.552 9.86a1.67 1.67 0 003.332 0l.552-9.86a2.221 2.221 0 00-2.218-2.345zm-.007 20.356c1.192 0 2.123-.964 2.123-2.172 0-1.209-.931-2.157-2.123-2.157-1.178 0-2.109.948-2.109 2.157 0 1.208.931 2.172 2.109 2.172z" fill="#242424"/><defs><linearGradient id="paint0_linear_4732_428" x1="116.102" y1="31.087" x2="116.102" y2="128.669" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint1_linear_4732_428" x1="116.374" y1="65.084" x2="116.374" y2="104.321" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint2_linear_4732_428" x1="151.627" y1="146.992" x2="187.23" y2="108.187" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient></defs></svg>`));
 export const MediaErrorLight = (/*#__PURE__*/createFluentIllustration('MediaErrorLight', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.613 147.8c37.749 0 68.35-30.624 68.35-68.4 0-37.776-30.601-68.4-68.35-68.4-37.75 0-68.351 30.624-68.351 68.4 0 37.776 30.601 68.4 68.351 68.4z" fill="#EBEBEB"/><g filter="url(#filter0_d_4732_412)"><path d="M178.665 38.646v83.629c0 2.948-2.42 5.376-5.359 5.376H58.979c-2.939 0-5.359-2.428-5.359-5.376V38.646c0-2.948 2.42-5.376 5.359-5.376h114.327c2.939 0 5.359 2.428 5.359 5.376z" fill="url(#paint0_linear_4732_412)"/></g><path d="M139.644 73.46c-.665-2.459-2.217-4.581-4.433-5.922-2.438-1.452-4.987-1.676-7.758-1.9-4.876-.558-9.863-.67-14.85-.446-4.433.112-9.531-.112-13.632 1.787-5.874 2.57-6.65 8.268-6.871 14.413-.222 4.58-.222 10.055.997 14.524.998 3.575 3.769 6.368 7.426 7.15 4.654 1.118 14.629 1.341 19.505 1.229h1.33c3.879-.111 7.979-.111 11.526-1.452 2.992-1.006 4.765-2.793 5.873-5.028 1.33-2.681 1.663-6.033 1.773-9.72.444-4.804.333-10.278-.886-14.636z" fill="url(#paint1_linear_4732_412)"/><path d="M112.136 78.3c0-1.474.745-1.965 1.954-1.18 3.257 2.065 6.607 4.227 9.865 6.29 1.209.787 1.209 1.671 0 2.556-3.258 2.063-6.515 4.225-9.772 6.29-1.302.785-1.954.392-2.047-.984V78.3z" fill="#fff"/><path d="M178.665 39.068v3.882H53.62v-3.882c0-2.932 2.42-5.348 5.359-5.348h114.327c2.939 0 5.359 2.416 5.359 5.348z" fill="#D0C6B8"/><path d="M58.867 40.017a1.638 1.638 0 10-.002-3.277 1.638 1.638 0 00.002 3.277z" fill="#F4F1EA"/><path opacity=".6" d="M63.952 40.017a1.638 1.638 0 10-.001-3.277 1.638 1.638 0 00.001 3.277z" fill="#A09F9D"/><path opacity=".4" d="M69.124 40.017a1.638 1.638 0 10-.001-3.277 1.638 1.638 0 00.001 3.277z" fill="#949390"/><path d="M166.047 100.559l-19.458 28.384c-2.493 3.637.363 8.364 5.055 8.364h38.916c4.692 0 7.548-4.727 5.055-8.364l-19.458-28.384c-2.339-3.412-7.771-3.412-10.11 0z" fill="url(#paint2_linear_4732_412)"/><path d="M171.226 109.626c-1.394 0-2.501 1.072-2.424 2.345l.603 9.86c.054.884.853 1.575 1.821 1.575.967 0 1.766-.691 1.82-1.575l.603-9.86c.077-1.273-1.03-2.345-2.423-2.345zm-.008 20.356c1.302 0 2.319-.963 2.319-2.172 0-1.208-1.017-2.156-2.319-2.156-1.287 0-2.304.948-2.304 2.156 0 1.209 1.017 2.172 2.304 2.172z" fill="#fff"/><defs><linearGradient id="paint0_linear_4732_412" x1="116.102" y1="31.087" x2="116.102" y2="128.669" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_4732_412" x1="115.926" y1="59.128" x2="116.442" y2="88.469" gradientUnits="userSpaceOnUse"><stop stop-color="#E1DAD0"/><stop offset="1" stop-color="#CCC1B1"/></linearGradient><linearGradient id="paint2_linear_4732_412" x1="194.519" y1="98" x2="147.219" y2="98.213" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><filter id="filter0_d_4732_412" x="41.62" y="32.07" width="149.045" height="118.381" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="10.8"/><feGaussianBlur stdDeviation="6"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.24 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4732_412"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4732_412" result="shape"/></filter></defs></svg>`));
 export const MediaLight = (/*#__PURE__*/createFluentIllustration('MediaLight', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.612 147.8c37.749 0 68.351-30.624 68.351-68.4 0-37.776-30.602-68.4-68.351-68.4-37.749 0-68.35 30.624-68.35 68.4 0 37.776 30.601 68.4 68.35 68.4z" fill="#EBEBEB"/><g filter="url(#filter0_d_3831_4960)"><path d="M178.664 38.646v83.629c0 2.948-2.42 5.376-5.359 5.376H58.978c-2.938 0-5.358-2.428-5.358-5.376V38.646c0-2.948 2.42-5.376 5.358-5.376h114.327c2.939 0 5.359 2.428 5.359 5.376z" fill="url(#paint0_linear_3831_4960)"/></g><path d="M139.643 73.46c-.665-2.459-2.216-4.581-4.433-5.922-2.438-1.452-4.987-1.676-7.758-1.9-4.876-.558-9.863-.67-14.85-.446-4.433.112-9.531-.112-13.632 1.787-5.873 2.57-6.65 8.268-6.87 14.413-.222 4.58-.222 10.055.997 14.524.997 3.575 3.768 6.368 7.425 7.15 4.655 1.118 14.629 1.341 19.505 1.229h1.33c3.879-.111 7.979-.111 11.526-1.452 2.992-1.006 4.765-2.793 5.873-5.028 1.33-2.681 1.663-6.033 1.774-9.72.443-4.804.332-10.278-.887-14.636z" fill="url(#paint1_linear_3831_4960)"/><path d="M112.135 78.3c0-1.474.745-1.965 1.954-1.18 3.257 2.065 6.607 4.227 9.864 6.29 1.21.787 1.21 1.671 0 2.556-3.257 2.063-6.514 4.225-9.771 6.29-1.302.785-1.954.392-2.047-.984V78.3z" fill="#fff"/><path d="M178.664 39.068v3.882H53.62v-3.882c0-2.932 2.42-5.348 5.358-5.348h114.328c2.938 0 5.358 2.416 5.358 5.348z" fill="#D0C6B8"/><path d="M58.866 40.017a1.638 1.638 0 100-3.277 1.638 1.638 0 000 3.277z" fill="#F4F1EA"/><path opacity=".6" d="M63.952 40.017a1.638 1.638 0 10-.001-3.277 1.638 1.638 0 000 3.277z" fill="#A09F9D"/><path opacity=".4" d="M69.123 40.017a1.638 1.638 0 100-3.277 1.638 1.638 0 000 3.277z" fill="#949390"/><defs><linearGradient id="paint0_linear_3831_4960" x1="116.101" y1="31.087" x2="116.101" y2="128.669" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_3831_4960" x1="115.925" y1="59.128" x2="116.442" y2="88.469" gradientUnits="userSpaceOnUse"><stop stop-color="#E1DAD0"/><stop offset="1" stop-color="#CCC1B1"/></linearGradient><filter id="filter0_d_3831_4960" x="41.62" y="32.07" width="149.045" height="118.381" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="10.8"/><feGaussianBlur stdDeviation="6"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.24 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3831_4960"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3831_4960" result="shape"/></filter></defs></svg>`));
 export const MilPermissionsDark = (/*#__PURE__*/createFluentIllustration('MilPermissionsDark', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M114.636 144.8c34.267 0 62.048-27.801 62.048-62.215S148.781 20.37 114.636 20.37c-34.267 0-62.048 27.8-62.048 62.215 0 34.414 27.781 62.215 62.048 62.215z" fill="#242424"/><path d="M172.645 41.802a5.02 5.02 0 005.018-5.021 5.02 5.02 0 00-5.018-5.021 5.019 5.019 0 00-5.017 5.02 5.019 5.019 0 005.017 5.022zM179.989 22.207a3.428 3.428 0 003.426-3.43 3.428 3.428 0 10-6.853 0 3.428 3.428 0 003.427 3.43zM55.892 41.68a3.428 3.428 0 003.427-3.43 3.428 3.428 0 10-3.427 3.43zM34.965 107.814a6.367 6.367 0 006.364-6.369 6.366 6.366 0 00-6.364-6.368 6.366 6.366 0 00-6.364 6.368 6.367 6.367 0 006.364 6.369z" fill="#5C5C5C"/><mask id="a" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="52" y="8" width="125" height="137"><path d="M168.393 8H63.068v40.251c-6.5 9.834-10.287 21.634-10.287 34.334 0 34.414 27.8 62.215 62.093 62.215 34.292 0 62.092-27.801 62.092-62.215 0-11.52-3.128-22.298-8.573-31.541V8z" fill="#5C5C5C"/></mask><g mask="url(#a)"><path d="M67.377 147.616l-1.102-129.34c0-3.034 2.48-5.377 5.51-5.377h86.528c3.031 0 5.511 2.481 5.511 5.377v123.963c0 3.033-2.48 5.377-5.511 5.377H67.377z" fill="url(#paint0_linear_3831_4848)"/></g><ellipse cx="114.774" cy="82.325" rx="38.778" ry="46.318" fill="#3A3A3A"/><path d="M153.552 82.325c0 25.58-17.361 46.318-38.778 46.318-21.416 0-38.778-20.738-38.778-46.319 0-25.58 17.362-46.318 38.778-46.318 21.417 0 38.778 20.738 38.778 46.319z" fill="#3A3A3A"/><path d="M109.486 113.268c0-.973.789-1.762 1.763-1.762h6.445c.973 0 1.763.789 1.763 1.762v.962c0 .973-.79 1.762-1.763 1.762h-6.445a1.762 1.762 0 01-1.763-1.762v-.962z" fill="#F5F2EC"/><path d="M114.952 113.749c.974 0 1.763.789 1.763 1.762v11.929c0 .974-.789 1.763-1.763 1.763h-.961a1.763 1.763 0 01-1.763-1.763v-11.929c0-.973.789-1.762 1.763-1.762h.961zM140.332 94.761a.882.882 0 01-.881-.881v-6.17c0-.486.395-.88.881-.88a3.526 3.526 0 013.526 3.524v.882a3.526 3.526 0 01-3.526 3.525zM89.216 86.829c.487 0 .881.394.881.88v6.17a.881.881 0 01-.881.881 3.525 3.525 0 01-3.525-3.525v-.881a3.525 3.525 0 013.525-3.525z" fill="#F5F2EC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M94.504 86.388c.973 0 1.762.79 1.762 1.763v3.084c0 10.222 8.287 18.508 18.508 18.508 10.222 0 18.508-8.286 18.508-18.508v-3.084a1.762 1.762 0 113.525 0v3.084c0 12.169-9.864 22.033-22.033 22.033-12.168 0-22.033-9.864-22.033-22.033v-3.084c0-.974.79-1.763 1.763-1.763z" fill="#F5F2EC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M98.91 74.49c0-8.76 7.103-15.863 15.864-15.863 8.762 0 15.864 7.102 15.864 15.863v16.745c0 8.762-7.102 15.864-15.864 15.864-8.761 0-15.864-7.102-15.864-15.864V74.49zm15.864-12.338c-6.814 0-12.338 5.524-12.338 12.338v16.745c0 6.815 5.524 12.339 12.338 12.339 6.815 0 12.339-5.524 12.339-12.339V74.49c0-6.814-5.524-12.338-12.339-12.338z" fill="#F5F2EC"/><path d="M98.91 86.829h31.728v4.406c0 8.761-7.102 15.864-15.864 15.864-8.761 0-15.864-7.103-15.864-15.864V86.83z" fill="#F5F2EC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M111.249 81.541c0 .974-.789 1.763-1.763 1.763h-8.813v-3.526h8.813c.974 0 1.763.79 1.763 1.763zM111.249 74.49c0 .974-.789 1.763-1.763 1.763h-8.813v-3.525h8.813c.974 0 1.763.789 1.763 1.762zM118.299 74.49c0-.973.79-1.762 1.763-1.762h8.813v3.525h-8.813c-.973 0-1.763-.79-1.763-1.763zM118.299 81.541c0-.973.79-1.762 1.763-1.762h8.813v3.525h-8.813c-.973 0-1.763-.79-1.763-1.763z" fill="#F5F2EC"/><defs><linearGradient id="paint0_linear_3831_4848" x1="115.081" y1="150.733" x2="115.081" y2="11.446" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient></defs></svg>`));
@@ -29,6 +33,8 @@ export const NotificationsDark = (/*#__PURE__*/createFluentIllustration('Notific
 export const NotificationsLight = (/*#__PURE__*/createFluentIllustration('NotificationsLight', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M115.01 141.738c29.877 0 54.098-24.221 54.098-54.204S144.78 33.33 115.01 33.33c-29.876 0-54.097 24.221-54.097 54.204s24.221 54.204 54.097 54.204z" fill="#EBEBEB"/><path d="M172.468 122.507a4.374 4.374 0 100-8.749 4.374 4.374 0 000 8.749zM178.871 105.435a2.987 2.987 0 10.001-5.975 2.987 2.987 0 00-.001 5.975zM63.794 51.896a2.988 2.988 0 100-5.975 2.988 2.988 0 000 5.975zM45.548 109.515a5.55 5.55 0 100-11.098 5.55 5.55 0 000 11.098z" fill="#E8DFD3"/><g filter="url(#filter0_dd_3692_28347)"><path d="M117.615 46.422h-4.565c-1.058 0-1.948.98-1.948 2.145v2.513c0 2.574 1.892 4.657 4.231 4.657 2.338 0 4.231-2.083 4.231-4.657v-2.513c0-1.226-.891-2.145-1.949-2.145z" fill="#FCFCFC"/><path d="M113.011 23.18v-.418a2.21 2.21 0 012.235-2.236c1.253 0 2.235.982 2.235 2.236v.418a9.87 9.87 0 017.631 9.615v1.255c0 10.579 7.958 17.504 7.958 17.504H97.367s8.013-6.925 8.013-17.504v-1.255a9.87 9.87 0 017.631-9.616z" fill="url(#paint0_linear_3692_28347)"/></g><path d="M126.364 23.76a4.81 4.81 0 100 9.622 4.81 4.81 0 000-9.622z" fill="#D9D2C7"/><g filter="url(#filter1_d_3692_28347)"><rect x="79.1" y="111.395" width="76.953" height="30.886" rx="6" fill="url(#paint1_linear_3692_28347)"/></g><path fill-rule="evenodd" clip-rule="evenodd" d="M136.126 124.161H90.092v-2h46.034v2zM116.888 132.25H90.092v-2h26.796v2z" fill="#EDE5DA"/><g filter="url(#filter2_d_3692_28347)"><rect x="70.943" y="63.206" width="91.144" height="40.183" rx="6" fill="url(#paint2_linear_3692_28347)"/></g><path d="M88.31 89.731a8.854 8.854 0 01-8.888-8.894 8.854 8.854 0 018.888-8.894 8.854 8.854 0 018.888 8.894c0 4.843-3.95 8.894-8.888 8.894z" fill="#D9D2C7"/><path d="M93.642 85.383C92.26 84 90.976 83.21 88.31 83.21c-2.765 0-3.95.89-5.332 2.174a7.229 7.229 0 005.431 2.47c1.975 0 3.753-.79 5.037-2.074 0-.1.197-.198.197-.396.099 0 0 0 0 0zM88.308 81.529a2.964 2.964 0 002.963-2.965 2.964 2.964 0 10-2.963 2.965z" fill="#fff"/><path fill-rule="evenodd" clip-rule="evenodd" d="M149.067 84.298h-45.572v-2h45.572v2zM149.067 76.16h-45.572v-2h45.572v2zM131.164 92.436h-27.669v-2h27.669v2z" fill="#EDE5DA"/><path d="M131.269 21.406l1.719-3.906M134.678 25.352l6.76-3.436" stroke="#EDE5DA" stroke-width="2" stroke-miterlimit="10" stroke-linecap="round"/><defs><linearGradient id="paint0_linear_3692_28347" x1="115.207" y1="19.808" x2="115.207" y2="51.888" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_3692_28347" x1="117.551" y1="110.68" x2="117.551" y2="142.613" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint2_linear_3692_28347" x1="116.486" y1="62.277" x2="116.486" y2="103.823" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><filter id="filter0_dd_3692_28347" x="81.367" y="2.526" width="67.702" height="67.211" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="-2"/><feGaussianBlur stdDeviation="8"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.16 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3692_28347"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset/><feGaussianBlur stdDeviation="1.5"/><feComposite in2="hardAlpha" operator="out"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.16 0"/><feBlend in2="effect1_dropShadow_3692_28347" result="effect2_dropShadow_3692_28347"/><feBlend in="SourceGraphic" in2="effect2_dropShadow_3692_28347" result="shape"/></filter><filter id="filter1_d_3692_28347" x="63.1" y="95.395" width="108.952" height="62.886" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset/><feGaussianBlur stdDeviation="8"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.24 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3692_28347"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3692_28347" result="shape"/></filter><filter id="filter2_d_3692_28347" x="54.943" y="51.206" width="123.145" height="72.183" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="4"/><feGaussianBlur stdDeviation="8"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3692_28347"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3692_28347" result="shape"/></filter></defs></svg>`));
 export const OrganizationDark = (/*#__PURE__*/createFluentIllustration('OrganizationDark', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><g clip-path="url(#clip0_4731_668)"><path d="M112.405 134.389c27.778 0 50.262-22.602 50.262-50.489 0-27.887-22.484-50.408-50.262-50.408-27.779 0-50.262 22.602-50.262 50.49 0 27.968 22.564 50.407 50.262 50.407z" fill="#242424"/><path fill-rule="evenodd" clip-rule="evenodd" d="M151.495 36.896c.474 0 .859.385.859.859v15.771a.859.859 0 01-1.717 0V37.755c0-.474.384-.859.858-.859z" fill="#353535"/><path d="M145.101 52.248h12.787v12.787h-12.787V52.248z" fill="#353535"/><g filter="url(#filter0_d_4731_668)"><path d="M123.788 48.837l39.215 18.29v73.781h-39.215v-92.07z" fill="#4D4D4D"/></g><path d="M132.313 74.546a1.84 1.84 0 011.839-1.838h5.7c1.016 0 1.839.823 1.839 1.838v3.143a1.839 1.839 0 01-1.839 1.839h-5.7a1.84 1.84 0 01-1.839-1.84v-3.142zM132.313 87.334a1.84 1.84 0 011.839-1.839h5.7c1.016 0 1.839.823 1.839 1.839v3.995a1.839 1.839 0 01-1.839 1.839h-5.7a1.84 1.84 0 01-1.839-1.839v-3.995zM132.313 100.121a1.84 1.84 0 011.839-1.838h5.7c1.016 0 1.839.823 1.839 1.838v3.995a1.839 1.839 0 01-1.839 1.839h-5.7a1.84 1.84 0 01-1.839-1.839v-3.995zM132.313 113.761c0-1.015.824-1.838 1.839-1.838h5.7c1.016 0 1.839.823 1.839 1.838v3.995a1.839 1.839 0 01-1.839 1.839h-5.7a1.84 1.84 0 01-1.839-1.839v-3.995zM132.313 126.549a1.84 1.84 0 011.839-1.839h5.7c1.016 0 1.839.823 1.839 1.839v3.995a1.84 1.84 0 01-1.839 1.839h-5.7a1.84 1.84 0 01-1.839-1.839v-3.995zM145.954 74.546c0-1.015.823-1.838 1.839-1.838h4.847c1.016 0 1.839.823 1.839 1.838v3.143a1.839 1.839 0 01-1.839 1.839h-4.847a1.839 1.839 0 01-1.839-1.84v-3.142zM145.954 87.334c0-1.016.823-1.839 1.839-1.839h4.847c1.016 0 1.839.823 1.839 1.839v3.995a1.839 1.839 0 01-1.839 1.839h-4.847a1.839 1.839 0 01-1.839-1.839v-3.995zM145.954 100.121c0-1.015.823-1.838 1.839-1.838h4.847c1.016 0 1.839.823 1.839 1.838v3.995a1.839 1.839 0 01-1.839 1.839h-4.847a1.839 1.839 0 01-1.839-1.839v-3.995zM145.954 113.761c0-1.015.823-1.838 1.839-1.838h4.847c1.016 0 1.839.823 1.839 1.838v3.995a1.839 1.839 0 01-1.839 1.839h-4.847a1.839 1.839 0 01-1.839-1.839v-3.995zM145.954 126.549c0-1.016.823-1.839 1.839-1.839h4.847c1.016 0 1.839.823 1.839 1.839v3.995a1.84 1.84 0 01-1.839 1.839h-4.847a1.84 1.84 0 01-1.839-1.839v-3.995z" fill="#3A3A3A"/><g filter="url(#filter1_d_4731_668)"><path d="M89.689 19h34.952v121.908H89.688V19z" fill="#5C5C5C"/></g><path d="M95.656 29.364c0-1.016.823-1.839 1.839-1.839h19.34c1.016 0 1.839.823 1.839 1.839v2.29a1.839 1.839 0 01-1.839 1.838h-19.34a1.839 1.839 0 01-1.839-1.838v-2.29zM95.656 42.151c0-1.015.823-1.839 1.839-1.839h19.34c1.016 0 1.839.824 1.839 1.84v1.437a1.839 1.839 0 01-1.839 1.839h-19.34a1.839 1.839 0 01-1.839-1.84v-1.437zM95.656 54.939c0-1.016.823-1.839 1.839-1.839h19.34c1.016 0 1.839.823 1.839 1.839v1.437a1.839 1.839 0 01-1.839 1.839h-19.34a1.839 1.839 0 01-1.839-1.839V54.94zM95.656 67.726c0-1.015.823-1.838 1.839-1.838h19.34c1.016 0 1.839.823 1.839 1.838v1.438a1.839 1.839 0 01-1.839 1.839h-19.34a1.839 1.839 0 01-1.839-1.839v-1.438zM95.656 80.514c0-1.016.823-1.84 1.839-1.84h19.34c1.016 0 1.839.824 1.839 1.84v1.437a1.839 1.839 0 01-1.839 1.839h-19.34a1.839 1.839 0 01-1.839-1.839v-1.437zM95.656 93.302c0-1.016.823-1.84 1.839-1.84h19.34c1.016 0 1.839.824 1.839 1.84v1.437a1.839 1.839 0 01-1.839 1.839h-19.34a1.839 1.839 0 01-1.839-1.839v-1.438zM95.656 106.089a1.84 1.84 0 011.839-1.839h19.34a1.84 1.84 0 011.839 1.839v1.438a1.838 1.838 0 01-1.839 1.838h-19.34a1.838 1.838 0 01-1.839-1.838v-1.438zM95.656 118.876a1.84 1.84 0 011.839-1.839h19.34a1.84 1.84 0 011.839 1.839v1.438a1.838 1.838 0 01-1.839 1.838h-19.34a1.838 1.838 0 01-1.839-1.838v-1.438z" fill="#3A3A3A"/><g filter="url(#filter2_d_4731_668)"><path d="M70.081 64.183h38.363v76.725H70.081V64.183z" fill="#707070"/></g><path d="M77.872 73.055c0-1.015.823-1.839 1.839-1.839h5.295c1.015 0 1.838.824 1.838 1.84v3.833a1.839 1.839 0 01-1.838 1.839H79.71a1.839 1.839 0 01-1.839-1.839v-3.834zM77.872 86.201c0-1.015.823-1.839 1.839-1.839h5.295c1.015 0 1.838.824 1.838 1.84v3.833a1.839 1.839 0 01-1.838 1.839H79.71a1.839 1.839 0 01-1.839-1.839v-3.834zM77.872 99.346c0-1.016.823-1.839 1.839-1.839h5.295c1.015 0 1.838.823 1.838 1.839v3.834a1.839 1.839 0 01-1.838 1.839H79.71a1.839 1.839 0 01-1.839-1.839v-3.834zM77.872 112.492c0-1.016.823-1.839 1.839-1.839h5.295c1.015 0 1.838.823 1.838 1.839v3.834a1.839 1.839 0 01-1.838 1.839H79.71a1.839 1.839 0 01-1.839-1.839v-3.834zM80.311 128.254a1.84 1.84 0 011.84-1.839h7.538v14.493h-9.377v-12.654zM91.435 73.055c0-1.015.823-1.839 1.838-1.839h5.295c1.016 0 1.839.824 1.839 1.84v3.833a1.839 1.839 0 01-1.839 1.84h-5.295a1.839 1.839 0 01-1.838-1.84v-3.834zM91.435 86.201c0-1.015.823-1.839 1.838-1.839h5.295c1.016 0 1.839.824 1.839 1.84v3.833a1.839 1.839 0 01-1.839 1.839h-5.295a1.839 1.839 0 01-1.838-1.839v-3.834zM91.435 99.346c0-1.016.823-1.839 1.838-1.839h5.295c1.016 0 1.839.823 1.839 1.839v3.834a1.839 1.839 0 01-1.839 1.839h-5.295a1.839 1.839 0 01-1.838-1.839v-3.834zM91.435 112.492c0-1.016.823-1.839 1.838-1.839h5.295c1.016 0 1.839.823 1.839 1.839v3.834a1.839 1.839 0 01-1.839 1.839h-5.295a1.839 1.839 0 01-1.838-1.839v-3.834zM89.69 126.415h7.538a1.84 1.84 0 011.839 1.839v12.654H89.69v-14.493z" fill="#3A3A3A"/><path fill-rule="evenodd" clip-rule="evenodd" d="M51 140.908a.92.92 0 01.92-.92h128.709a.92.92 0 010 1.839H51.919a.919.919 0 01-.919-.919z" fill="#3A3A3A"/></g><defs><filter id="filter0_d_4731_668" x="111.788" y="44.837" width="63.215" height="116.07" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="8"/><feGaussianBlur stdDeviation="6"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.16 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4731_668"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4731_668" result="shape"/></filter><filter id="filter1_d_4731_668" x="77.689" y="15" width="58.952" height="145.908" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="8"/><feGaussianBlur stdDeviation="6"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.16 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4731_668"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4731_668" result="shape"/></filter><filter id="filter2_d_4731_668" x="58.081" y="60.183" width="62.362" height="100.725" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="8"/><feGaussianBlur stdDeviation="6"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.16 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4731_668"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4731_668" result="shape"/></filter><clipPath id="clip0_4731_668"><path fill="#fff" transform="translate(.792)" d="M0 0h230v160H0z"/></clipPath></defs></svg>`));
 export const OrganizationLight = (/*#__PURE__*/createFluentIllustration('OrganizationLight', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><g clip-path="url(#clip0_3267_4900)"><path d="M111.405 134.389c27.778 0 50.262-22.602 50.262-50.489 0-27.887-22.484-50.408-50.262-50.408-27.779 0-50.262 22.603-50.262 50.49 0 27.968 22.564 50.407 50.262 50.407z" fill="#EBEBEB"/><path fill-rule="evenodd" clip-rule="evenodd" d="M150.498 36.896c.474 0 .859.385.859.859v15.771a.859.859 0 01-1.717 0V37.755c0-.474.384-.859.858-.859z" fill="#D9D2C7"/><path d="M144.104 52.248h12.788v12.787h-12.788V52.248z" fill="#D9D2C7"/><g filter="url(#filter0_d_3267_4900)"><path d="M122.792 48.837l39.215 18.29v73.781h-39.215v-92.07z" fill="url(#paint0_linear_3267_4900)"/></g><path d="M131.317 74.546a1.84 1.84 0 011.839-1.839h5.7c1.016 0 1.839.824 1.839 1.84v3.142a1.839 1.839 0 01-1.839 1.839h-5.7a1.84 1.84 0 01-1.839-1.84v-3.142zM131.317 87.334a1.84 1.84 0 011.839-1.839h5.7c1.016 0 1.839.823 1.839 1.839v3.995a1.839 1.839 0 01-1.839 1.839h-5.7a1.84 1.84 0 01-1.839-1.839v-3.995zM131.317 100.121a1.84 1.84 0 011.839-1.838h5.7c1.016 0 1.839.823 1.839 1.838v3.996a1.838 1.838 0 01-1.839 1.838h-5.7a1.839 1.839 0 01-1.839-1.838v-3.996zM131.317 113.761c0-1.015.824-1.838 1.839-1.838h5.7c1.016 0 1.839.823 1.839 1.838v3.995a1.839 1.839 0 01-1.839 1.839h-5.7a1.84 1.84 0 01-1.839-1.839v-3.995zM131.317 126.549a1.84 1.84 0 011.839-1.839h5.7c1.016 0 1.839.823 1.839 1.839v3.995a1.84 1.84 0 01-1.839 1.839h-5.7a1.84 1.84 0 01-1.839-1.839v-3.995zM144.957 74.546c0-1.015.823-1.839 1.839-1.839h4.847c1.016 0 1.839.824 1.839 1.84v3.142a1.839 1.839 0 01-1.839 1.839h-4.847a1.839 1.839 0 01-1.839-1.84v-3.142zM144.957 87.334c0-1.016.823-1.839 1.839-1.839h4.847c1.016 0 1.839.823 1.839 1.839v3.995a1.839 1.839 0 01-1.839 1.839h-4.847a1.839 1.839 0 01-1.839-1.839v-3.995zM144.957 100.121c0-1.015.823-1.838 1.839-1.838h4.847c1.016 0 1.839.823 1.839 1.838v3.996a1.838 1.838 0 01-1.839 1.838h-4.847a1.838 1.838 0 01-1.839-1.838v-3.996zM144.957 113.761c0-1.015.823-1.838 1.839-1.838h4.847c1.016 0 1.839.823 1.839 1.838v3.995a1.839 1.839 0 01-1.839 1.839h-4.847a1.839 1.839 0 01-1.839-1.839v-3.995zM144.957 126.549c0-1.016.823-1.839 1.839-1.839h4.847c1.016 0 1.839.823 1.839 1.839v3.995a1.84 1.84 0 01-1.839 1.839h-4.847a1.84 1.84 0 01-1.839-1.839v-3.995z" fill="#DED4C6"/><g filter="url(#filter1_d_3267_4900)"><path d="M88.692 19h34.953v121.908H88.692V19z" fill="url(#paint1_linear_3267_4900)"/></g><path d="M94.66 29.364c0-1.016.822-1.839 1.838-1.839h19.34c1.016 0 1.839.823 1.839 1.839v2.29a1.839 1.839 0 01-1.839 1.839h-19.34a1.839 1.839 0 01-1.839-1.84v-2.29zM94.66 42.151c0-1.015.822-1.838 1.838-1.838h19.34c1.016 0 1.839.823 1.839 1.838v1.438a1.839 1.839 0 01-1.839 1.839h-19.34a1.839 1.839 0 01-1.839-1.84v-1.437zM94.66 54.939c0-1.016.822-1.839 1.838-1.839h19.34c1.016 0 1.839.823 1.839 1.839v1.437a1.839 1.839 0 01-1.839 1.84h-19.34a1.839 1.839 0 01-1.839-1.84V54.94z" fill="#fff"/><path d="M94.66 67.726c0-1.015.822-1.838 1.838-1.838h19.34c1.016 0 1.839.823 1.839 1.838v1.438a1.839 1.839 0 01-1.839 1.839h-19.34a1.839 1.839 0 01-1.839-1.84v-1.437zM94.66 80.514c0-1.016.822-1.839 1.838-1.839h19.34c1.016 0 1.839.823 1.839 1.839v1.438a1.839 1.839 0 01-1.839 1.838h-19.34a1.839 1.839 0 01-1.839-1.839v-1.437zM94.66 93.301c0-1.015.822-1.838 1.838-1.838h19.34c1.016 0 1.839.823 1.839 1.838v1.438a1.839 1.839 0 01-1.839 1.839h-19.34a1.839 1.839 0 01-1.839-1.84v-1.437zM94.66 106.089a1.84 1.84 0 011.838-1.839h19.34a1.84 1.84 0 011.839 1.839v1.438a1.838 1.838 0 01-1.839 1.838h-19.34a1.838 1.838 0 01-1.839-1.838v-1.438zM94.66 118.876c0-1.015.822-1.838 1.838-1.838h19.34c1.016 0 1.839.823 1.839 1.838v1.438a1.839 1.839 0 01-1.839 1.839h-19.34a1.839 1.839 0 01-1.839-1.839v-1.438z" fill="#F4EFE9"/><g filter="url(#filter2_d_3267_4900)"><path d="M69.084 64.183h38.363v76.725H69.084V64.183z" fill="#FCFBF8"/></g><path d="M76.875 73.055c0-1.015.823-1.838 1.839-1.838h5.294c1.016 0 1.84.823 1.84 1.838v3.835a1.839 1.839 0 01-1.84 1.838h-5.294a1.839 1.839 0 01-1.839-1.838v-3.835zM76.875 86.2c0-1.015.823-1.838 1.839-1.838h5.294c1.016 0 1.84.823 1.84 1.839v3.834a1.839 1.839 0 01-1.84 1.839h-5.294a1.839 1.839 0 01-1.839-1.839v-3.834zM76.875 99.346c0-1.016.823-1.839 1.839-1.839h5.294c1.016 0 1.84.823 1.84 1.839v3.834a1.839 1.839 0 01-1.84 1.839h-5.294a1.839 1.839 0 01-1.839-1.839v-3.834zM76.875 112.491c0-1.015.823-1.838 1.839-1.838h5.294c1.016 0 1.84.823 1.84 1.838v3.835a1.838 1.838 0 01-1.84 1.838h-5.294a1.838 1.838 0 01-1.839-1.838v-3.835zM79.314 128.254c0-1.016.824-1.839 1.84-1.839h7.538v14.493h-9.377v-12.654zM90.438 73.055c0-1.015.823-1.838 1.838-1.838h5.295c1.016 0 1.839.823 1.839 1.838v3.835a1.839 1.839 0 01-1.839 1.838h-5.295a1.839 1.839 0 01-1.838-1.838v-3.835zM90.438 86.2c0-1.015.823-1.838 1.838-1.838h5.295c1.016 0 1.839.823 1.839 1.839v3.834a1.839 1.839 0 01-1.839 1.839h-5.295a1.839 1.839 0 01-1.838-1.839v-3.834zM90.438 99.346c0-1.016.823-1.839 1.838-1.839h5.295c1.016 0 1.839.823 1.839 1.839v3.834a1.839 1.839 0 01-1.839 1.839h-5.295a1.839 1.839 0 01-1.838-1.839v-3.834zM90.438 112.491c0-1.015.823-1.838 1.838-1.838h5.295c1.016 0 1.839.823 1.839 1.838v3.835a1.838 1.838 0 01-1.839 1.838h-5.295a1.838 1.838 0 01-1.838-1.838v-3.835zM88.692 126.415h7.54c1.015 0 1.838.823 1.838 1.839v12.654h-9.378v-14.493z" fill="#DED4C6"/><path fill-rule="evenodd" clip-rule="evenodd" d="M50 140.908a.92.92 0 01.92-.92h128.709a.92.92 0 010 1.839H50.919a.92.92 0 01-.919-.919z" fill="#EDE5DA"/></g><defs><filter id="filter0_d_3267_4900" x="106.792" y="40.837" width="71.215" height="124.07" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="8"/><feGaussianBlur stdDeviation="8"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.16 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3267_4900"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3267_4900" result="shape"/></filter><filter id="filter1_d_3267_4900" x="76.692" y="15" width="58.952" height="145.908" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="8"/><feGaussianBlur stdDeviation="6"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.16 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3267_4900"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3267_4900" result="shape"/></filter><filter id="filter2_d_3267_4900" x="53.084" y="56.183" width="70.362" height="108.725" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="8"/><feGaussianBlur stdDeviation="8"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.16 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3267_4900"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3267_4900" result="shape"/></filter><linearGradient id="paint0_linear_3267_4900" x1="141.921" y1="54.628" x2="142.387" y2="141.9" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFDFD"/><stop offset=".996" stop-color="#F4F1ED"/></linearGradient><linearGradient id="paint1_linear_3267_4900" x1="106.157" y1="16.18" x2="106.157" y2="142.222" gradientUnits="userSpaceOnUse"><stop stop-color="#EAE7E1"/><stop offset=".996" stop-color="#B4B09F"/></linearGradient><clipPath id="clip0_3267_4900"><path fill="#fff" d="M0 0h230v160H0z"/></clipPath></defs></svg>`));
+export const ProhibitedDark = (/*#__PURE__*/createFluentIllustration('ProhibitedDark', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M114.56 137.302c31.236 0 56.56-25.323 56.56-56.671s-25.436-56.672-56.56-56.672C83.324 23.96 58 49.284 58 80.631c0 31.348 25.324 56.671 56.56 56.671z" fill="#242424"/><path d="M156.281 127.224a2.573 2.573 0 100 5.146 2.573 2.573 0 000-5.146zM169.315 120.089a1.757 1.757 0 100 3.515 1.757 1.757 0 000-3.515zM78.943 26.514a1.757 1.757 0 100-3.514 1.757 1.757 0 000 3.514zM70.856 37.2a3.263 3.263 0 100-6.526 3.263 3.263 0 000 6.527z" fill="#323232"/><path d="M58 55.002a8 8 0 018-8h97.243a8 8 0 018 8v49.296a8 8 0 01-8 8H114.98c-2.764 0-5.327 1.426-6.938 3.672-4.297 5.991-13.493 17.359-22.37 18.915-1.088.191-1.988-.717-1.988-1.822v-12.765a8 8 0 00-8-8H66a8 8 0 01-8-8V55.001z" fill="url(#paint0_linear_5118_410)"/><path fill-rule="evenodd" clip-rule="evenodd" d="M185.987 50.857c.039.8-.577 1.48-1.377 1.52l-7.918.392a1.451 1.451 0 01-.144-2.897l7.918-.392c.8-.04 1.481.577 1.521 1.377zM181.719 41.992a1.45 1.45 0 01-.592 1.965l-4 2.146a1.45 1.45 0 01-1.372-2.556l4-2.147a1.45 1.45 0 011.964.593zM170.417 43.892a1.451 1.451 0 01-.982-1.8l2.939-9.986a1.451 1.451 0 012.783.82l-2.939 9.984a1.45 1.45 0 01-1.801.982z" fill="#777"/><path d="M152.563 69.154a1 1 0 011 1.032l-.4 12.35a1 1 0 01-1 .967h-2.214a1 1 0 01-.999-.968l-.4-12.349a1 1 0 011-1.032h3.013zm-1.512 20.974c-.72 0-1.339-.256-1.858-.768a2.533 2.533 0 01-.767-1.858c0-.714.256-1.323.767-1.828a2.554 2.554 0 011.858-.767c.694 0 1.303.256 1.828.767.532.505.798 1.114.798 1.828 0 .485-.125.926-.374 1.323a2.753 2.753 0 01-.959.95 2.455 2.455 0 01-1.293.352zM134.451 85.957v-1.09c0-.849.181-1.633.545-2.353.37-.72.905-1.3 1.606-1.737.7-.438 1.555-.657 2.565-.657 1.03 0 1.895.22 2.595.657a4.147 4.147 0 011.595 1.727c.364.713.546 1.5.546 2.362v1.091c0 .848-.182 1.633-.546 2.353-.363.72-.899 1.3-1.605 1.737-.707.437-1.569.656-2.585.656-1.024 0-1.885-.219-2.586-.656a4.3 4.3 0 01-1.595-1.737 5.224 5.224 0 01-.535-2.353zm3.221-1.09v1.09c0 .43.104.845.313 1.242.215.397.609.596 1.182.596.585 0 .979-.192 1.181-.576.202-.39.303-.81.303-1.262v-1.09a2.85 2.85 0 00-.293-1.273c-.188-.404-.586-.606-1.191-.606-.566 0-.956.202-1.172.606a2.633 2.633 0 00-.323 1.272zm-13.945-10.745v-1.09c0-.862.185-1.65.555-2.364a4.38 4.38 0 011.616-1.726c.7-.438 1.548-.657 2.544-.657 1.03 0 1.895.22 2.596.657a4.216 4.216 0 011.605 1.727c.364.713.545 1.5.545 2.362v1.091c0 .862-.185 1.65-.555 2.363a4.162 4.162 0 01-1.606 1.707c-.706.424-1.568.636-2.585.636-1.016 0-1.878-.216-2.585-.647a4.248 4.248 0 01-1.595-1.716 5.162 5.162 0 01-.535-2.343zm3.251-1.09v1.09c0 .451.104.872.313 1.262.216.384.599.576 1.151.576.593 0 .987-.192 1.182-.576.202-.39.303-.81.303-1.262v-1.09c0-.452-.094-.876-.283-1.273-.188-.404-.589-.606-1.202-.606-.565 0-.952.206-1.161.616a2.83 2.83 0 00-.303 1.262zm.174 16.803a1 1 0 01-.824-1.567l12.843-18.68c.187-.272.495-.434.824-.434h.33a1 1 0 01.824 1.566l-12.843 18.681a1 1 0 01-.824.434h-.33zM109.917 90.148c-1.454 0-2.709-.26-3.766-.778-1.051-.518-1.858-1.215-2.424-2.09a5.335 5.335 0 01-.848-2.939c0-.794.155-1.494.464-2.1a6.155 6.155 0 011.303-1.686 15.94 15.94 0 011.949-1.525l4.342-2.777c.492-.31.838-.636 1.04-.98.209-.343.313-.686.313-1.03 0-.37-.144-.706-.434-1.01-.283-.302-.703-.45-1.262-.444-.357 0-.667.074-.929.222a1.633 1.633 0 00-.616.596c-.142.25-.212.532-.212.849 0 .41.118.83.353 1.262.236.43.552.888.949 1.373.404.485.852 1.02 1.343 1.606l8.426 9.473c.574.645.116 1.665-.747 1.665h-2.363a1 1 0 01-.735-.322l-7.994-8.675a38.316 38.316 0 01-1.484-1.768 10.517 10.517 0 01-1.313-2.14 6.046 6.046 0 01-.535-2.545c0-1.03.242-1.963.727-2.797a5.32 5.32 0 012.04-1.99c.882-.498 1.908-.747 3.08-.747 1.131 0 2.107.232 2.928.697.821.464 1.454 1.08 1.899 1.848.444.767.666 1.609.666 2.524 0 .97-.239 1.862-.717 2.676-.478.808-1.154 1.525-2.03 2.151l-4.251 3.04c-.451.323-.804.67-1.06 1.04-.256.363-.384.76-.384 1.191 0 .445.104.832.313 1.162.216.33.519.589.909.777.39.189.848.283 1.373.283.7 0 1.401-.165 2.101-.495a6.61 6.61 0 001.918-1.414 7.151 7.151 0 001.394-2.17 6.795 6.795 0 00.464-1.77c.069-.548.509-.998 1.061-.998h1.979c.552 0 1.003.449.97 1a13.697 13.697 0 01-.364 2.454 9.819 9.819 0 01-1.182 2.848 7.224 7.224 0 01-1.928 2.08c-.243.134-.478.262-.707.384-.229.12-.465.249-.707.383a7.888 7.888 0 01-2.585 1.232 10.43 10.43 0 01-2.727.374zM98.524 69.154a1 1 0 011 1.032l-.4 12.35a1 1 0 01-1 .967H95.91a1 1 0 01-1-.968l-.399-12.35a1 1 0 011-1.031h3.013zm-1.512 20.973c-.72 0-1.34-.255-1.858-.767a2.53 2.53 0 01-.767-1.858c0-.714.255-1.323.767-1.828a2.552 2.552 0 011.858-.767c.694 0 1.303.255 1.828.767.532.505.798 1.114.798 1.828 0 .485-.125.925-.374 1.323a2.755 2.755 0 01-.96.949c-.39.236-.82.353-1.292.353zM84.17 89.834a1 1 0 01-.986-1.162l3.064-18.68a1 1 0 01.987-.838h1.609a1 1 0 01.987 1.162l-3.065 18.68a1 1 0 01-.987.838H84.17zm-9.213-4.927a1 1 0 01-.986-1.167l.277-1.636a1 1 0 01.986-.833h14.371a1 1 0 01.986 1.167l-.277 1.636a1 1 0 01-.986.833h-14.37zm1.943 4.927a1 1 0 01-.987-1.162l3.065-18.68a1 1 0 01.987-.838h1.608a1 1 0 01.987 1.162l-3.065 18.68a1 1 0 01-.987.838H76.9zm-.761-12.117a1 1 0 01-.986-1.167l.277-1.636a1 1 0 01.986-.833h14.37a1 1 0 01.986 1.168l-.277 1.635a1 1 0 01-.986.833H76.14z" fill="#D1D1D1"/><defs><linearGradient id="paint0_linear_5118_410" x1="114.585" y1="41.895" x2="114.585" y2="136.045" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#3A3A3A"/></linearGradient></defs></svg>`));
+export const ProhibitedLight = (/*#__PURE__*/createFluentIllustration('ProhibitedLight', `<svg viewBox="0 0 230 164" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M114.56 137.302c31.236 0 56.56-25.323 56.56-56.671s-25.436-56.672-56.56-56.672C83.324 23.96 58 49.284 58 80.631c0 31.348 25.324 56.671 56.56 56.671z" fill="#F0F0F0"/><path d="M156.281 127.224a2.573 2.573 0 100 5.146 2.573 2.573 0 000-5.146zM169.315 120.089a1.757 1.757 0 100 3.514 1.757 1.757 0 000-3.514zM78.943 26.514a1.757 1.757 0 100-3.514 1.757 1.757 0 000 3.514zM70.856 37.2a3.263 3.263 0 100-6.526 3.263 3.263 0 000 6.526z" fill="#EDE5DA"/><g filter="url(#filter0_d_5118_408)"><path d="M58 52.001a5 5 0 015-5h103.243a5 5 0 015 5v55.296a5 5 0 01-5 5h-52.93c-1.727 0-3.327.884-4.309 2.305-3.363 4.871-11.791 16.045-20.451 19.455-2.57 1.011-4.869-1.233-4.869-3.994v-12.766a5 5 0 00-5-5H63a5 5 0 01-5-5V52.001z" fill="url(#paint0_linear_5118_408)"/></g><path fill-rule="evenodd" clip-rule="evenodd" d="M185.987 50.857c.039.8-.577 1.48-1.377 1.52l-7.918.392a1.451 1.451 0 01-.144-2.897l7.918-.392c.8-.04 1.481.577 1.521 1.377zM181.719 41.992a1.45 1.45 0 01-.592 1.965l-4 2.146a1.45 1.45 0 01-1.372-2.556l4-2.147a1.45 1.45 0 011.964.593zM170.417 43.892a1.451 1.451 0 01-.982-1.8l2.939-9.986a1.451 1.451 0 012.783.82l-2.939 9.984a1.45 1.45 0 01-1.801.982z" fill="#EDE5DA"/><path d="M152.563 69.154a1 1 0 011 1.032l-.4 12.35a1 1 0 01-1 .967h-2.214a1 1 0 01-.999-.968l-.4-12.349a1 1 0 011-1.032h3.013zm-1.512 20.974c-.72 0-1.339-.256-1.858-.768a2.533 2.533 0 01-.767-1.858c0-.714.256-1.323.767-1.828a2.554 2.554 0 011.858-.767c.694 0 1.303.256 1.828.767.532.505.798 1.114.798 1.828 0 .485-.125.926-.374 1.323a2.753 2.753 0 01-.959.95 2.455 2.455 0 01-1.293.352zM134.451 85.957v-1.09c0-.849.181-1.633.545-2.353.37-.72.905-1.3 1.606-1.737.7-.438 1.555-.656 2.565-.656 1.03 0 1.895.218 2.595.656a4.146 4.146 0 011.595 1.727c.364.713.546 1.501.546 2.363v1.09c0 .849-.182 1.633-.546 2.353-.363.72-.899 1.3-1.605 1.737-.707.438-1.569.656-2.585.656-1.024 0-1.885-.218-2.586-.656a4.3 4.3 0 01-1.595-1.737 5.224 5.224 0 01-.535-2.353zm3.221-1.09v1.09c0 .431.104.845.313 1.242.215.398.609.596 1.182.596.585 0 .979-.192 1.181-.576.202-.39.303-.81.303-1.262v-1.09a2.85 2.85 0 00-.293-1.273c-.188-.404-.586-.606-1.191-.606-.566 0-.956.202-1.172.606a2.633 2.633 0 00-.323 1.273zm-13.945-10.745v-1.09c0-.862.185-1.65.555-2.363a4.381 4.381 0 011.616-1.727c.7-.438 1.548-.656 2.544-.656 1.03 0 1.895.218 2.596.656a4.217 4.217 0 011.605 1.727c.364.713.545 1.501.545 2.363v1.09c0 .862-.185 1.65-.555 2.363a4.162 4.162 0 01-1.606 1.707c-.706.424-1.568.636-2.585.636-1.016 0-1.878-.215-2.585-.646a4.249 4.249 0 01-1.595-1.717 5.162 5.162 0 01-.535-2.343zm3.251-1.09v1.09c0 .451.104.872.313 1.263.216.383.599.575 1.151.575.593 0 .987-.192 1.182-.575.202-.39.303-.812.303-1.263v-1.09c0-.451-.094-.875-.283-1.273-.188-.403-.589-.606-1.202-.606-.565 0-.952.206-1.161.616a2.83 2.83 0 00-.303 1.263zm.174 16.803a1 1 0 01-.824-1.567l12.843-18.68a1 1 0 01.824-.434h.33a1 1 0 01.824 1.567L128.306 89.4a1 1 0 01-.824.434h-.33zM109.917 90.148c-1.454 0-2.709-.26-3.766-.778-1.051-.518-1.858-1.215-2.424-2.09a5.335 5.335 0 01-.848-2.939c0-.794.155-1.494.464-2.1a6.155 6.155 0 011.303-1.686 15.94 15.94 0 011.949-1.525l4.342-2.777c.492-.31.838-.636 1.04-.98.209-.343.313-.686.313-1.03 0-.37-.144-.706-.434-1.01-.283-.302-.703-.45-1.262-.444-.357 0-.667.074-.929.222a1.633 1.633 0 00-.616.596c-.142.25-.212.532-.212.849 0 .41.118.83.353 1.262.236.43.552.888.949 1.373.404.485.852 1.02 1.343 1.606l8.426 9.473c.574.645.116 1.665-.747 1.665h-2.363a1 1 0 01-.735-.322l-7.994-8.675a38.316 38.316 0 01-1.484-1.768 10.517 10.517 0 01-1.313-2.14 6.046 6.046 0 01-.535-2.545c0-1.03.242-1.963.727-2.797a5.32 5.32 0 012.04-1.99c.882-.498 1.908-.747 3.08-.747 1.131 0 2.107.232 2.928.697.821.464 1.454 1.08 1.899 1.848.444.767.666 1.609.666 2.524 0 .97-.239 1.862-.717 2.676-.478.808-1.154 1.525-2.03 2.151l-4.251 3.04c-.451.323-.804.67-1.06 1.04-.256.363-.384.76-.384 1.191 0 .445.104.832.313 1.162.216.33.519.589.909.777.39.189.848.283 1.373.283.7 0 1.401-.165 2.101-.495a6.61 6.61 0 001.918-1.414 7.151 7.151 0 001.394-2.17 6.795 6.795 0 00.464-1.77c.069-.548.509-.998 1.061-.998h1.979c.552 0 1.003.449.97 1a13.697 13.697 0 01-.364 2.454 9.819 9.819 0 01-1.182 2.848 7.224 7.224 0 01-1.928 2.08c-.243.134-.478.262-.707.384-.229.12-.465.249-.707.383a7.888 7.888 0 01-2.585 1.232 10.43 10.43 0 01-2.727.374zM98.524 69.154a1 1 0 011 1.032l-.4 12.35a1 1 0 01-1 .967H95.91a1 1 0 01-1-.968l-.399-12.349a1 1 0 011-1.032h3.013zm-1.512 20.974c-.72 0-1.34-.256-1.858-.768a2.53 2.53 0 01-.767-1.858c0-.714.255-1.323.767-1.828a2.552 2.552 0 011.858-.767c.694 0 1.303.256 1.828.767.532.505.798 1.114.798 1.828 0 .485-.125.926-.374 1.323a2.755 2.755 0 01-.96.95c-.39.235-.82.352-1.292.352zM84.17 89.835a1 1 0 01-.986-1.162l3.064-18.681a1 1 0 01.987-.838h1.609a1 1 0 01.987 1.162l-3.065 18.68a1 1 0 01-.987.839H84.17zm-9.213-4.928a1 1 0 01-.986-1.167l.277-1.636a1 1 0 01.986-.832h14.371a1 1 0 01.986 1.167l-.277 1.635a1 1 0 01-.986.833h-14.37zm1.943 4.928a1 1 0 01-.987-1.162l3.065-18.681a1 1 0 01.987-.838h1.608a1 1 0 01.987 1.162l-3.065 18.68a1 1 0 01-.987.839H76.9zm-.761-12.118a1 1 0 01-.986-1.167l.277-1.635a1 1 0 01.986-.833h14.37a1 1 0 01.986 1.167l-.277 1.635a1 1 0 01-.986.833H76.14z" fill="#D9D2C7"/><defs><linearGradient id="paint0_linear_5118_408" x1="114.585" y1="44.964" x2="114.585" y2="136.012" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><filter id="filter0_d_5118_408" x="36" y="32.001" width="157.243" height="131.302" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="7"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_5118_408"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_5118_408" result="shape"/></filter></defs></svg>`));
 export const SettingsDark = (/*#__PURE__*/createFluentIllustration('SettingsDark', `<svg viewBox="0 0 231 169" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M114.668 135.883c30.727 0 55.636-25.065 55.636-55.985 0-30.92-24.909-55.985-55.636-55.985-30.726 0-55.635 25.065-55.635 55.985 0 30.92 24.909 55.985 55.635 55.985z" fill="#242424"/><path d="M134.525 19.761a1.245 1.245 0 00.078-1.751 1.226 1.226 0 00-1.74-.079 1.245 1.245 0 00-.078 1.752c.459.505 1.238.54 1.74.078z" fill="#2A2E37"/><g filter="url(#filter0_d_3840_762)"><path fill-rule="evenodd" clip-rule="evenodd" d="M108.93 135.317c1.716.177 3.137-1.299 4.36-2.515.827-.822 1.654-1.645 2.481-2.741 1.961-2.195 4.141-1.569 6.733-.825.321.092.649.186.984.277 3.872.906 5.109-2.308 6.978-7.165.392-1.019.812-2.11 1.29-3.251 1.674-3.33-.802-4.803-3.467-6.387-.406-.242-.817-.486-1.218-.74-2.351-1.403-2.296-3.604-2.229-6.262.012-.458.024-.93.024-1.413-.258-2.821 1.836-4.082 3.856-5.3 2.298-1.384 4.501-2.711 3.034-6.213-.359-.857-.69-1.682-1.005-2.466-2.099-5.23-3.464-8.63-7.539-7.676-.36.098-.713.2-1.058.3-2.559.737-4.716 1.359-6.659-.574-.539-.603-1.094-1.124-1.618-1.615-1.618-1.519-2.933-2.752-2.516-5.237 1.713-7.302 1.253-7.472-6.397-10.297-.965-.356-2.045-.755-3.25-1.216-3.518-1.312-4.775.866-6.094 3.151-1.16 2.01-2.368 4.104-5.206 3.976-.645 0-1.269.031-1.868.061-2.472.124-4.518.227-5.849-1.98-3.935-6.36-4.358-6.171-12.043-2.745-.926.413-1.959.873-3.115 1.375-3.787 1.613-2.796 4.81-1.716 8.297l.062.2c.654 2.603-.762 4.004-2.474 5.698-.452.447-.924.914-1.384 1.429-1.16 1.318-2.63 2.828-2.371 4.564 3.748 25.063 24.007 44.685 49.274 47.29zM83.249 121.29c7.992 3.015 17.087-.823 20.119-8.772s-.827-16.995-8.82-20.01c-7.992-3.015-17.087.823-20.119 8.772s.827 16.994 8.82 20.01z" fill="#3A3A3A"/></g><g filter="url(#filter1_d_3840_762)"><path d="M153.107 61.392c-.216-3.275.216-5.458-1.948-5.676-1.514-.219-2.596 0-3.246-1.31-.649-1.31-1.515-2.402-.649-3.494 2.164-3.275 1.515-3.057-2.164-6.55-2.38-2.184-3.462 2.401-5.843 1.31-1.298-.655-2.597-.655-3.029-2.184-.866-3.712-.866-3.275-6.06-3.057-1.947 0-1.947 1.747-1.947 3.494-.217 1.528-1.299 1.747-2.813 2.402-2.381 1.31-3.896-3.057-6.06-.655a21.585 21.585 0 00-2.493 3.019c-.216.313-.414.651-.462 1.028-.163 1.275 1.044 1.78 1.815 3.058.039.065.058.057.058.133 0 .067.018.216.041.28.146.405-.041.81-.041 1.216 0 .596-.318 1.162-.514 1.725-.236.675-.497 1.174-1.433 1.331-3.679.655-3.246.874-3.03 6.114.216 3.275 4.112.873 5.194 3.275.352.711.991 1.711.983 2.53-.003.362-.148.713-.317 1.033-.237.446-.594.853-.846 1.294a4.912 4.912 0 00-.501 1.207c-.415 1.602 1.323 2.64 3.061 4.198 2.381 2.184 3.463-2.401 5.843-1.31 2.813 1.092 2.597.437 3.246 3.712.433 2.184 3.03 1.529 5.843 1.31 3.246-.218.865-4.148 3.246-5.24.433-.218 1.082-.437 1.515-.655 1.298-.655 2.164.218 3.462.873 1.948 1.31 3.03-.655 5.194-3.275 1.298-1.528 0-2.62-1.082-3.93-.866-1.092-.217-2.183.216-3.712.866-2.183 4.977-.218 4.761-3.493zm-14.932 4.15c-2.597 2.838-6.925 3.274-9.954.654-2.813-2.62-3.246-6.987-.649-10.044 2.596-2.838 6.924-3.275 9.954-.655 2.813 2.62 3.246 6.987.649 10.044z" fill="#3A3A3A"/></g><path d="M47.155 78.813a4.3 4.3 0 00-1.962 5.726 4.235 4.235 0 005.69 1.954 4.3 4.3 0 001.962-5.727 4.235 4.235 0 00-5.69-1.953zM129.875 141.019a2.82 2.82 0 00-2.723 2.899 2.795 2.795 0 002.875 2.728 2.82 2.82 0 002.723-2.899 2.795 2.795 0 00-2.875-2.728zM121.426 148.115a1.919 1.919 0 00-1.852 1.972 1.902 1.902 0 001.955 1.855 1.919 1.919 0 001.852-1.972 1.902 1.902 0 00-1.955-1.855zM41.494 96.985a1.81 1.81 0 00-.826 2.412 1.783 1.783 0 002.396.822 1.81 1.81 0 00.826-2.411 1.783 1.783 0 00-2.396-.823zM77.434 134.984a1.244 1.244 0 00-.568 1.658 1.227 1.227 0 001.647.566 1.245 1.245 0 00.569-1.658 1.226 1.226 0 00-1.648-.566z" fill="#5C5C5C"/><path d="M121.864 21.503l-12.37 13.188c-3.841 4.043-9.978 4.515-13.806.974-3.829-3.541-3.89-9.606-.048-13.649l12.369-13.188c3.841-4.043 9.979-4.515 13.807-.974 3.867 3.433 3.889 9.605.048 13.649z" fill="url(#paint0_linear_3840_762)"/><path d="M119.186 19.972c2.351-2.566 2.156-6.615-.436-9.044-2.592-2.428-6.599-2.317-8.95.25-2.351 2.566-2.156 6.615.436 9.044 2.592 2.428 6.599 2.317 8.95-.25z" fill="#323232"/><g filter="url(#filter2_d_3840_762)"><path d="M79.275 39.692l-3.627 3.41c-.432.396-1.106.533-1.686.343l-4.31-1.343c-.58-.19-1.065-.708-1.13-1.29l-1.03-4.353c-.15-.587.04-1.242.471-1.638l3.627-3.41c.259-.237.275-.734.032-.992a.65.65 0 00-.412-.184 10.723 10.723 0 00-8.795 2.694 10.738 10.738 0 00-2.946 10.814L35.046 66.245c-2.073 1.983-2.262 5.207-.239 7.365 1.942 2.07 5.203 2.209 7.19.304l24.682-22.74c3.647 1.15 7.61.24 10.375-2.378a10.57 10.57 0 003.375-8.558c.013-.414-.397-.68-.73-.694-.168-.007-.254.072-.424.148zM37.747 70.75c-.729-.777-.688-2.018.088-2.731.777-.713 1.947-.664 2.676.113.728.777.688 2.018-.089 2.73-.777.714-1.947.664-2.675-.112-.003.082-.003.082 0 0z" fill="#4D4D4D"/></g><path opacity=".4" d="M59.336 48.37L41.303 64.725M62.223 51.393L44.107 67.744" stroke="#2A2E37" stroke-miterlimit="10"/><path d="M180.657 42.594c-.379-.242-.948-.231-1.485-.014a1.96 1.96 0 01-2.699-.509 1.451 1.451 0 01-.204-1.344c.159-.459-.03-.953-.519-1.172l-1.01-.644-1.614 2.145-19.183-12.245a1.03 1.03 0 01-.442-.656l-.078-.517c-.062-.413-.33-.678-.693-.816l-4.956-1.856-2.564 3.408 3.627 3.623c.268.264.631.402 1.073.311l.553-.114a1.01 1.01 0 01.821.15l19.184 12.246-1.614 2.146 1.009.644c.379.242.948.231 1.374.036a1.96 1.96 0 012.7.51c.284.367.362.884.203 1.343-.159.46.141.93.519 1.172l13.757 8.782c2.146 1.37 5.352.708 6.966-1.438 0 0 .095-.126.08-.23 1.488-2.225.955-4.994-1.191-6.364l-13.614-8.597z" fill="#4D4D4D"/><path opacity=".4" d="M177.527 48.944l10.35 6.87M179.294 46.42l10.349 6.869" stroke="#2A2E37" stroke-miterlimit="10"/><path d="M168.923 80.064l-24.259 15.052a2.781 2.781 0 00-.901 3.846l15.052 24.259a2.782 2.782 0 003.846.901l24.26-15.052a2.782 2.782 0 00.9-3.846l-15.052-24.26a2.781 2.781 0 00-3.846-.9z" fill="url(#paint1_linear_3840_762)"/><path d="M170.692 89.83l-7.317 4.541c-1.295-1.205-3.152-1.696-4.848-1.192-2.61.798-4.058 3.614-3.195 6.182l-2.307 1.432c-.593.368-.793 1.223-.425 1.816.409.659 1.263.859 1.857.491l2.307-1.431c1.214 1.073 2.882 1.407 4.38 1.025 2.625-.624 4.221-3.35 3.597-5.975l7.383-4.58c.593-.37.793-1.224.425-1.817-.343-.7-1.197-.9-1.857-.491zm-8.488 8.098c.032.346-.044.666-.225.962-.091.147-.116.254-.248.336-.157.189-.248.336-.446.459-1.014.72-2.361.461-3.082-.553l-.082-.132a2.169 2.169 0 01.701-2.99c.989-.614 2.271-.314 2.95.634.246.395.425.832.432 1.284zm15.728 3.571l-2.241 1.391c-1.296-1.205-3.153-1.696-4.848-1.192-2.61.797-3.992 3.572-3.195 6.182l-7.383 4.581a1.391 1.391 0 00-.45 1.923 1.39 1.39 0 001.922.45l7.384-4.581a4.68 4.68 0 004.38 1.026c2.625-.625 4.221-3.35 3.596-5.975l2.308-1.432a1.39 1.39 0 00.45-1.923 1.391 1.391 0 00-1.923-.45zm-3.787 3.902c.327.528.309 1.087.184 1.621-.025.107-.116.254-.141.361-.091.148-.116.255-.207.402-.091.148-.157.189-.248.337a1500968941724.744 1500968941724.744 0 00-.396.245c-.131.082-.263.164-.436.18-.066.041-.132.082-.239.057-.848.252-1.891-.105-2.423-.962-.49-.791-.438-1.737.058-2.41.157-.188.248-.336.445-.459l.198-.122a2.167 2.167 0 012.991.7c.239-.057.28.009.214.05z" fill="#323232"/><defs><filter id="filter0_d_3840_762" x="37.626" y="52.605" width="116.675" height="115.727" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_762"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_762" result="shape"/></filter><filter id="filter1_d_3840_762" x="91.151" y="29.293" width="83.964" height="84.796" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_762"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_762" result="shape"/></filter><filter id="filter2_d_3840_762" x="11.386" y="19.123" width="91.071" height="89.134" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_762"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_762" result="shape"/></filter><linearGradient id="paint0_linear_3840_762" x1="103.07" y1="10.157" x2="115.039" y2="32.696" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint1_linear_3840_762" x1="156.368" y1="86.925" x2="174.975" y2="116.914" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient></defs></svg>`));
 export const SettingsLight = (/*#__PURE__*/createFluentIllustration('SettingsLight', `<svg viewBox="0 0 231 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M114.169 135.922c30.727 0 55.636-25.006 55.636-55.852 0-30.847-24.909-55.852-55.636-55.852-30.727 0-55.636 25.005-55.636 55.852 0 30.846 24.909 55.852 55.636 55.852z" fill="#EBEBEB"/><path d="M134.026 20.076a1.24 1.24 0 00.078-1.747 1.228 1.228 0 00-1.74-.079 1.24 1.24 0 00-.079 1.747 1.23 1.23 0 001.741.08z" fill="#EDE5D9"/><g filter="url(#filter0_d_3840_744)"><path fill-rule="evenodd" clip-rule="evenodd" d="M108.437 135.358c1.712.176 3.131-1.297 4.354-2.51.826-.82 1.653-1.641 2.48-2.734 1.962-2.19 4.141-1.566 6.733-.824.322.092.65.186.984.277 3.873.904 5.11-2.302 6.979-7.148.392-1.016.812-2.105 1.29-3.244 1.674-3.321-.802-4.79-3.468-6.371a52.597 52.597 0 01-1.218-.738c-2.351-1.4-2.295-3.595-2.228-6.247.011-.457.023-.928.023-1.41-.258-2.814 1.836-4.072 3.856-5.287 2.299-1.381 4.502-2.705 3.035-6.198-.36-.856-.691-1.679-1.005-2.46-2.1-5.218-3.465-8.61-7.539-7.658-.361.097-.713.199-1.059.298-2.559.736-4.716 1.356-6.659-.572-.538-.6-1.094-1.12-1.617-1.61-1.619-1.516-2.933-2.746-2.517-5.226 1.713-7.284 1.253-7.454-6.396-10.272-.966-.356-2.046-.753-3.25-1.213-3.519-1.31-4.776.864-6.094 3.144-1.16 2.006-2.368 4.094-5.206 3.966-.645 0-1.27.031-1.868.061-2.473.124-4.518.226-5.85-1.975-3.935-6.345-4.358-6.157-12.043-2.739-.926.412-1.958.871-3.115 1.371-3.786 1.61-2.796 4.8-1.716 8.278l.062.2c.654 2.596-.762 3.994-2.474 5.684-.452.446-.924.912-1.384 1.425-1.16 1.316-2.63 2.825-2.37 4.56 3.75 25.002 24.01 44.575 49.28 47.172zm-25.688-13.995c7.993 3.008 17.088-.82 20.12-8.75 3.031-7.931-.827-16.955-8.82-19.963-7.993-3.008-17.088.82-20.12 8.751-3.032 7.93.827 16.954 8.82 19.962z" fill="url(#paint0_linear_3840_744)"/></g><g filter="url(#filter1_d_3840_744)"><path d="M152.608 61.608c-.217-3.267.216-5.445-1.948-5.663-1.515-.218-2.597 0-3.246-1.307s-1.515-2.396-.649-3.485c2.164-3.267 1.515-3.05-2.164-6.535-2.381-2.178-3.463 2.396-5.843 1.307-1.298-.653-2.597-.653-3.03-2.178-.865-3.703-.865-3.267-6.059-3.05-1.948 0-1.948 1.743-1.948 3.485-.216 1.525-1.298 1.743-2.813 2.397-2.38 1.306-3.895-3.05-6.059-.654a21.47 21.47 0 00-2.492 3.01c-.217.314-.416.652-.464 1.03-.161 1.27 1.045 1.774 1.815 3.048.039.065.059.057.059.133 0 .067.017.216.04.28.146.404-.04.808-.04 1.212 0 .595-.319 1.16-.515 1.721-.235.674-.496 1.172-1.433 1.329-3.679.653-3.246.871-3.029 6.099.216 3.267 4.111.871 5.193 3.267.352.709.99 1.705.984 2.521-.003.363-.148.715-.319 1.035-.236.444-.591.849-.844 1.287a4.919 4.919 0 00-.502 1.21c-.413 1.596 1.324 2.63 3.062 4.185 2.38 2.178 3.462-2.396 5.843-1.307 2.813 1.09 2.596.436 3.246 3.703.432 2.178 3.029 1.525 5.842 1.307 3.247-.218.866-4.139 3.247-5.228.432-.218 1.082-.435 1.514-.653 1.299-.654 2.164.218 3.463.871 1.947 1.307 3.029-.653 5.194-3.267 1.298-1.525 0-2.614-1.082-3.921-.866-1.09-.217-2.178.216-3.703.866-2.178 4.977-.218 4.761-3.485zm-14.932 4.14c-2.597 2.83-6.925 3.267-9.955.653-2.813-2.614-3.246-6.97-.649-10.02 2.597-2.832 6.925-3.268 9.955-.654 2.813 2.614 3.246 6.97.649 10.02z" fill="url(#paint1_linear_3840_744)"/></g><path d="M46.656 78.987a4.285 4.285 0 00-1.963 5.713 4.24 4.24 0 005.69 1.949 4.285 4.285 0 001.962-5.713 4.24 4.24 0 00-5.69-1.949z" fill="#EDE5D9"/><path d="M129.376 141.046a2.815 2.815 0 00-2.723 2.892 2.792 2.792 0 002.875 2.721 2.817 2.817 0 002.723-2.892 2.792 2.792 0 00-2.875-2.721zM120.927 148.125a1.915 1.915 0 00-1.852 1.967 1.899 1.899 0 001.955 1.85 1.916 1.916 0 001.852-1.967 1.899 1.899 0 00-1.955-1.85z" fill="#EAEEF9"/><path d="M40.994 97.117a1.804 1.804 0 00-.826 2.405 1.786 1.786 0 002.396.821 1.805 1.805 0 00.826-2.406 1.785 1.785 0 00-2.396-.82z" fill="#EDE5D9"/><path d="M76.934 135.024a1.241 1.241 0 00-.568 1.655 1.228 1.228 0 001.648.564 1.24 1.24 0 00.568-1.654 1.228 1.228 0 00-1.647-.565z" fill="#EAEEF9"/><path opacity=".2" d="M120.934 22.503l-12.985 13.62c-4.033 4.176-10.476 4.663-14.494 1.006-4.018-3.658-4.082-9.922-.05-14.097L106.392 9.41c4.032-4.176 10.475-4.663 14.493-1.005 4.059 3.547 4.082 9.921.05 14.097z" fill="url(#paint2_linear_3840_744)"/><path d="M117.82 20.918c2.545-2.733 2.439-6.948-.236-9.415-2.676-2.466-6.908-2.25-9.453.483-2.545 2.734-2.439 6.949.236 9.415 2.676 2.467 6.908 2.25 9.453-.483z" fill="#fff"/><path opacity=".2" d="M78.775 39.96l-3.627 3.401c-.431.395-1.105.532-1.685.342l-4.31-1.339c-.58-.19-1.066-.706-1.13-1.288l-1.031-4.342c-.149-.585.04-1.238.471-1.633l3.627-3.402c.26-.237.275-.733.032-.99a.65.65 0 00-.412-.184 10.74 10.74 0 00-8.795 2.688c-2.936 2.77-4.076 6.938-2.946 10.788L34.546 66.45c-2.073 1.98-2.262 5.195-.239 7.347 1.942 2.066 5.203 2.204 7.19.304L66.18 51.415a10.506 10.506 0 0010.375-2.373 10.531 10.531 0 003.376-8.538c.013-.413-.397-.678-.732-.692-.167-.007-.253.072-.423.147zM37.247 70.943c-.729-.774-.688-2.013.088-2.724.777-.711 1.947-.662 2.676.113.728.775.688 2.013-.089 2.724-.777.712-1.947.662-2.675-.113-.003.083-.003.083 0 0z" fill="url(#paint3_linear_3840_744)"/><path opacity=".4" d="M58.836 48.616L40.803 64.933M61.724 51.633L43.607 67.945" stroke="#BCB9B3" stroke-miterlimit="10"/><path opacity=".2" d="M180.157 42.856c-.378-.241-.947-.23-1.484-.014a1.963 1.963 0 01-2.7-.508 1.447 1.447 0 01-.203-1.34c.159-.458-.03-.951-.519-1.17l-1.01-.643-1.614 2.14-19.184-12.217a1.03 1.03 0 01-.441-.653l-.078-.516c-.062-.413-.33-.676-.693-.814l-4.956-1.852-2.564 3.4 3.627 3.614a1.14 1.14 0 001.073.31l.553-.113c.332-.069.568-.011.821.15l19.184 12.217-1.615 2.14 1.01.643c.379.241.947.23 1.374.037a1.962 1.962 0 012.699.507c.284.367.362.883.204 1.341-.159.458.14.928.519 1.17l13.757 8.76c2.145 1.367 5.352.707 6.966-1.433 0 0 .095-.126.079-.23 1.488-2.22.956-4.982-1.19-6.349l-13.615-8.577z" fill="url(#paint4_linear_3840_744)"/><path opacity=".4" d="M177.029 49.19l10.349 6.852M178.795 46.671l10.35 6.854" stroke="#D6DCE8" stroke-miterlimit="10"/><g filter="url(#filter2_d_3840_744)"><path d="M168.424 80.235L144.165 95.25a2.77 2.77 0 00-.901 3.837l15.053 24.202a2.785 2.785 0 003.845.898l24.26-15.016a2.772 2.772 0 00.901-3.837L172.27 81.133a2.785 2.785 0 00-3.846-.898z" fill="#EDE5DA"/></g><path d="M170.194 89.979l-7.318 4.53c-1.295-1.203-3.152-1.693-4.848-1.19-2.609.796-4.058 3.605-3.194 6.168l-2.308 1.428c-.593.367-.793 1.22-.425 1.812.409.657 1.264.857 1.857.49l2.307-1.428c1.214 1.07 2.882 1.404 4.38 1.023 2.626-.623 4.222-3.342 3.597-5.96l7.383-4.571c.594-.368.794-1.22.426-1.812-.343-.698-1.198-.898-1.857-.49zm-8.489 8.078c.032.345-.043.664-.225.96-.091.147-.116.253-.248.335-.157.188-.248.335-.445.458-1.014.718-2.362.46-3.083-.552l-.081-.131a2.16 2.16 0 01.7-2.984c.989-.612 2.271-.313 2.95.633.246.394.425.83.432 1.281zm15.729 3.563l-2.242 1.387c-1.295-1.202-3.152-1.692-4.848-1.189-2.609.795-3.992 3.564-3.194 6.167l-7.384 4.57a1.387 1.387 0 00-.45 1.919 1.393 1.393 0 001.923.449l7.383-4.57a4.688 4.688 0 004.38 1.023c2.625-.623 4.222-3.342 3.597-5.961l2.307-1.428c.659-.408.86-1.261.451-1.918a1.392 1.392 0 00-1.923-.449zm-3.788 3.892c.327.526.309 1.084.184 1.617-.025.107-.116.254-.141.361-.091.147-.116.254-.207.401-.091.147-.157.188-.248.336l-.197.122-.198.122c-.132.082-.264.164-.437.179-.065.041-.131.082-.238.057-.848.252-1.891-.105-2.423-.959-.491-.79-.439-1.733.057-2.404.157-.188.248-.336.446-.458l.197-.122a2.173 2.173 0 012.992.698c.238-.056.279.01.213.05z" fill="#fff"/><defs><linearGradient id="paint0_linear_3840_744" x1="95.264" y1="62.156" x2="95.264" y2="136.363" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_3840_744" x1="132.621" y1="39.617" x2="132.621" y2="81.697" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint2_linear_3840_744" x1="112.295" y1="29.577" x2="104.522" y2="14.28" gradientUnits="userSpaceOnUse"><stop offset=".003" stop-color="#877357"/><stop offset="1" stop-color="#938065"/></linearGradient><linearGradient id="paint3_linear_3840_744" x1="71.715" y1="51.848" x2="40.831" y2="50.318" gradientUnits="userSpaceOnUse"><stop offset=".003" stop-color="#938065"/><stop offset="1" stop-color="#938065"/></linearGradient><linearGradient id="paint4_linear_3840_744" x1="174.306" y1="56.082" x2="168.934" y2="27.734" gradientUnits="userSpaceOnUse"><stop offset=".003" stop-color="#938065"/><stop offset="1" stop-color="#938065"/></linearGradient><filter id="filter0_d_3840_744" x="47.127" y="58.816" width="96.675" height="95.557" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="7"/><feGaussianBlur stdDeviation="6"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_744"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_744" result="shape"/></filter><filter id="filter1_d_3840_744" x="90.652" y="25.559" width="83.964" height="84.699" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="7"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_744"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_744" result="shape"/></filter><filter id="filter2_d_3840_744" x="120.84" y="68.812" width="88.907" height="88.8" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0.397708 0 0 0 0 0.47749 0 0 0 0 0.575 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3840_744"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3840_744" result="shape"/></filter></defs></svg>`));
 export const SuccessDark = (/*#__PURE__*/createFluentIllustration('SuccessDark', `<svg viewBox="0 0 231 179" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M198.474 77.631l2.085-.927a.725.725 0 000-1.391l-2.085-.928a11.752 11.752 0 01-5.793-5.796l-.927-2.087a.724.724 0 00-1.39 0l-.927 2.087a11.746 11.746 0 01-5.792 5.796l-2.085.928a.725.725 0 000 1.39l2.085.928a11.746 11.746 0 015.792 5.797l.927 2.087a.724.724 0 001.39 0l.927-2.087c.927-2.55 3.012-4.638 5.793-5.797zM192.285 111.078l1.39-.696c.463-.232.463-.696 0-.927l-1.39-.464a6.623 6.623 0 01-3.939-3.942l-.695-1.391c-.232-.464-.695-.464-.927 0l-.463 1.391c-.927 1.855-2.317 3.246-3.939 3.942l-1.622.695c-.231 0-.231.696 0 .696l1.39.696c1.854.695 3.244 2.086 4.171 3.941l.695 1.391c.232.464.695.464.927 0l.463-1.391a6.619 6.619 0 013.939-3.941zM48.387 135.959l2.085-.927a.725.725 0 000-1.391l-2.085-.928a11.75 11.75 0 01-5.793-5.796l-.927-2.087a.724.724 0 00-1.39 0l-.927 2.087a11.75 11.75 0 01-5.792 5.796l-2.085.928a.725.725 0 000 1.391l2.085.927a11.752 11.752 0 015.792 5.797l.927 2.087a.724.724 0 001.39 0l.927-2.087c.927-2.551 3.012-4.637 5.793-5.797zM49.227 32.334l1.39-.695c.463-.232.463-.696 0-.928l-1.39-.463a6.623 6.623 0 01-3.94-3.942l-.694-1.391c-.232-.464-.695-.464-.927 0l-.463 1.391c-.927 1.855-2.317 3.246-3.94 3.942l-1.621.695c-.232 0-.232.696 0 .696l1.39.695c1.854.696 3.244 2.087 4.17 3.942l.696 1.391c.231.464.695.464.926 0l.464-1.39a6.623 6.623 0 013.939-3.943zM67.066 15.97l1.159-.463c.232-.232.232-.464 0-.696l-1.159-.464c-1.39-.463-2.317-1.623-3.012-3.014l-.463-1.16c-.232-.231-.463-.231-.695 0l-.464 1.16c-.463 1.391-1.621 2.319-3.012 3.014l-.695.464c-.463.232-.463.696 0 .696l1.159.463c1.39.464 2.317 1.623 3.012 3.015l.463 1.159c.232.232.464.232.695 0l.464-1.16c.231-1.39 1.39-2.55 2.548-3.014z" fill="#595959"/><g filter="url(#filter0_d_3871_354047)"><path d="M116.072 146.8c30.856 0 55.87-25.014 55.87-55.87 0-30.856-25.014-55.87-55.87-55.87-30.856 0-55.87 25.014-55.87 55.87 0 30.856 25.014 55.87 55.87 55.87z" fill="#fff"/></g><path d="M116.072 146.8c30.856 0 55.87-25.014 55.87-55.87 0-30.856-25.014-55.87-55.87-55.87-30.856 0-55.87 25.014-55.87 55.87 0 30.856 25.014 55.87 55.87 55.87z" fill="url(#paint0_linear_3871_354047)"/><path d="M106.258 116.517c-1.446 0-2.807-.532-3.912-1.683L83.547 95.259c-2.126-2.214-2.126-5.934 0-8.148 2.127-2.215 5.7-2.215 7.826 0l14.885 15.5 35.045-36.404c2.126-2.214 5.699-2.214 7.825 0 2.127 2.215 2.127 5.935 0 8.15l-38.957 40.477c-1.021 1.151-2.467 1.683-3.913 1.683z" fill="#656565"/><defs><linearGradient id="paint0_linear_3871_354047" x1="116.072" y1="35.06" x2="116.072" y2="146.8" gradientUnits="userSpaceOnUse"><stop stop-color="#FAFAFA"/><stop offset=".828" stop-color="#F5F2EC"/></linearGradient><filter id="filter0_d_3871_354047" x="38.879" y="24.398" width="154.386" height="154.386" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="10.662"/><feGaussianBlur stdDeviation="10.662"/><feColorMatrix values="0 0 0 0 0.482733 0 0 0 0 0.482733 0 0 0 0 0.482733 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3871_354047"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3871_354047" result="shape"/></filter></defs></svg>`));
@@ -43,7 +49,3 @@ export const UserDark = (/*#__PURE__*/createFluentIllustration('UserDark', `<svg
 export const UserGroupDark = (/*#__PURE__*/createFluentIllustration('UserGroupDark', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M118.982 132.7c28 0 50.7-22.7 50.7-50.8 0-28.1-22.8-50.8-50.7-50.8-28 0-50.7 22.7-50.7 50.8 0 28.1 22.7 50.8 50.7 50.8z" fill="#242424"/><path d="M166.382 48.6a4.1 4.1 0 100-8.2 4.1 4.1 0 000 8.2zM172.382 32.6a2.8 2.8 0 100-5.6 2.8 2.8 0 000 5.6zM70.982 48.5a2.8 2.8 0 100-5.6 2.8 2.8 0 000 5.6zM59.2 109.028a5.2 5.2 0 100-10.4 5.2 5.2 0 000 10.4z" fill="#5C5C5C"/><path d="M172.94 104.529h-15.381c-.727 0-1.351-.468-1.559-1.091-1.507-4.469-4.832-5.56-12.887-8.73-.467-.208-.519-.831-.052-1.091 2.339-1.143 1.923-2.13 1.923-8.99a1.581 1.581 0 01-.26-.364c-.519-.779-.883-1.714-.987-2.65l-.52-.26c-1.559.364-1.455-1.299-1.87-4.52-.156-1.351.052-1.611.623-1.82l.52-.675c-3.066-6.859-1.507-11.172 1.507-12.73-.988-2.183-1.039-2.962-1.039-2.962s6.391 1.09 8.573.623c2.806-.572 7.067.104 8.678 3.793 2.65 1.04 3.69 2.754 3.95 4.521.363 2.91-1.248 6.08-1.559 7.327.052.052.052.104.104.156.415.208.571.572.415 1.715-.363 3.066-.312 4.988-1.87 4.52l-1.403 2.443c-.052.312-.104.415-.104.571-.052.468-.052 1.351-.052 6.444 0 .623.156 1.195.467 1.663.312.571.832.987 1.455 1.247h.052c3.638 1.507 6.236 2.494 8.159 3.273 4.261 1.715 4.988 2.183 5.352 5.041.104 1.403-.884 2.546-2.235 2.546z" fill="#A1A1A1"/><path d="M162.494 74.857c.156-.78-.207-1.715-.519-2.39-.052-.104-.104-.208-.156-.364-1.195-2.442-3.949-3.481-6.547-3.585-6.652-.364-7.275.935-9.354-.987.624.987.728 2.494-.312 4.416-.675 1.351-2.026 1.923-3.221 2.339-3.014-6.86-1.507-11.172 1.507-12.731-.988-2.183-1.04-2.962-1.04-2.962s6.392 1.091 8.574.623c2.806-.571 7.067.104 8.678 3.794 2.65 1.039 3.69 2.754 3.949 4.52.416 2.91-1.247 6.08-1.559 7.327z" fill="#3A3A3A"/><path d="M160.052 92.473v1.195h-15.433v-1.559c.364-.831.26-.78.26-7.483a1.579 1.579 0 01-.26-.363v-.364c.312.572.779 1.091 1.247 1.507l4.833 3.377c1.143.936 2.754.988 3.897.052l4.521-4.053c.156-.156.312-.26.467-.415-.052.467-.052 1.35-.052 6.443.052.572.208 1.143.52 1.663z" fill="url(#paint0_linear_4784_322)"/><path d="M93.435 92.94c-2.494-.051-2.754-.519-2.754-.519v-4.157c2.65-2.183 4.885-4.729 5.508-8.054 1.663-.364 2.702-1.871 2.546-3.534-.052-.727-.52-1.403-.571-2.182v-1.092c0-.26 0-.467-.052-.727 0-.26-.052-.52-.052-.78-.156-1.247-.468-2.546-1.3-4.053-2.078-3.897-6.183-6.495-10.808-6.495-.935 0-1.87.104-2.754.312-1.662.363-3.17 1.09-4.52 2.13l-.78.624c-.052.052-.104.104-.208.156-1.455 1.35-2.598 3.066-3.221 4.988a13.26 13.26 0 00-.52 5.872v.312c.052.364.156.311-.104.623-.052.104-.156.208-.208.364-.312.572-.571 1.351-.208 2.286.624 1.767 1.871 1.507 3.014 2.495a.394.394 0 01-.208.104h-.052l-1.299.415c-7.794 2.443-8.678 10.705-.675 12.212-.572.052-1.091.363-1.455.623-2.546 1.975-4.157 4.989-4.573 7.951-.156.883.572 1.714 1.507 1.714h12.367c.26 0 .52-.207.624-.467 1.143-4.157 3.741-5.612 9.145-7.795l3.118-1.247c.52-.208 1.091-.416 1.715-.675.572-.208.52-1.04-.104-1.195-.624-.156-1.247-.208-1.87-.26-.52.052-.936.052-1.248.052zm-15.849-5.3c0-.52.364-1.09 1.144-1.714.311.415.727.883 1.195 1.299.364.311.727.623 1.143.935.26.156.52.364.78.572.051.052.103.103.155.103.052.053.156.104.208.156v2.547c-.052-.052-.104-.104-.156-.104-.26-.208-.467-.416-.675-.572-.364-.26-.728-.468-1.091-.675-.156-.104-.26-.156-.364-.26-1.195-.78-2.287-1.507-2.339-2.287z" fill="#A1A1A1"/><path d="M93.435 92.94c-.416 1.508-.883 2.443-.883 2.443l-13.407-2.13.78-3.326c.104.104.26.156.364.208.363.207.779.467 1.09.727.209.156.468.312.676.572l.104.104V88.99c-.104-.052-.156-.104-.208-.156 2.183 1.092 4.885 1.663 8.782-.623-.052 0-.052.052-.104.052v4.157s.312.468 2.806.52z" fill="url(#paint1_linear_4784_322)"/><path d="M98.06 72.675a16.867 16.867 0 01-7.015.883c-3.742-.311-7.379-1.818-10.185-4.313-.935 2.703-3.274 4.729-5.924 5.924-.364.156-.727.312-1.091.416v-.104a13.26 13.26 0 01.52-5.872c.623-1.922 1.766-3.637 3.221-4.988.052-.052.104-.104.208-.156l.78-.624a13.378 13.378 0 014.52-2.13 12.44 12.44 0 012.755-.312c4.624 0 8.73 2.598 10.808 6.495.78 1.507 1.143 2.806 1.299 4.001.104.26.104.52.104.78z" fill="#3A3A3A"/><path d="M78.885 86.081c.312.416.676.78.988 1.091-.364-.311-.728-.675-.988-1.09z" fill="#3973F4"/><path d="M83.77 99.228c-.987 1.247-2.858 1.299-4.417 1.247 1.663-1.61.728-6.859-5.092-6.235h-.052c-8.003-1.507-7.12-9.77.623-12.212l1.3-.416a.406.406 0 01.207.208 15.108 15.108 0 002.339 4.054c-3.17 2.65 1.247 3.481 3.377 5.508 1.87 1.09 3.43 5.664 1.715 7.846z" fill="#3A3A3A"/><path d="M78.834 86.081c-.053-.052-.104-.103-.104-.156 0 0 .051.105.103.156z" fill="#3973F4"/><path d="M153.037 106.919c.26 1.922-1.247 3.689-3.222 3.689h-60.9c-1.975 0-3.482-1.767-3.222-3.689.571-4.002 1.455-4.833 7.379-7.223l1.454-.572c2.755-1.091 6.34-2.442 11.173-4.469 1.091-.467 1.974-1.299 2.442-2.39.26-.572.416-1.247.416-1.923V81.25c-.156-.208-.312-.364-.416-.572a8.645 8.645 0 01-1.455-3.949l-.78-.364c-2.338.572-2.13-1.922-2.754-6.703-.259-1.975.052-2.39.936-2.702l.727-.987c-4.469-10.133-2.234-16.473 2.183-18.707-1.351-3.222-1.455-4.365-1.455-4.365s9.405 1.559 12.627.935c4.105-.831 10.444.156 12.783 5.612 3.897 1.56 5.404 4.001 5.768 6.703.571 4.261-1.871 8.938-2.339 10.757.052.104.052.156.104.208.624.312.832.831.624 2.546-.572 4.52-.416 7.327-2.754 6.703l-2.079 3.638c-.104.415-.104.571-.156.883-.104.676-.052 2.027-.052 9.457 0 .884.26 1.767.728 2.443.519.78 1.247 1.455 2.13 1.818.052 0 .052 0 .104.052 5.352 2.235 9.198 3.69 12.004 4.78 6.391 2.495 7.379 3.275 8.002 7.484z" fill="#A1A1A1"/><path d="M134.434 66.907c.208-1.143-.312-2.546-.727-3.534-.052-.155-.156-.311-.208-.52-1.819-3.637-5.82-5.092-9.665-5.3-9.821-.52-10.705 1.352-13.771-1.454.936 1.455 1.04 3.689-.415 6.495-1.04 1.975-3.014 2.858-4.729 3.43-4.469-10.133-2.234-16.473 2.183-18.707-1.455-3.274-1.559-4.417-1.559-4.417s9.405 1.559 12.627.935c4.105-.831 10.444.156 12.783 5.612 3.897 1.56 5.404 4.001 5.768 6.703.623 4.21-1.819 8.886-2.287 10.757z" fill="#3A3A3A"/><path d="M130.849 92.784v1.767h-22.656v-2.286c.26-.572.415-1.247.415-1.923V81.25c-.155-.208-.311-.364-.415-.572v-.52a7.663 7.663 0 001.87 2.235l7.119 4.988a4.544 4.544 0 005.768.104l6.652-5.975c.208-.208.467-.416.675-.624-.104.675-.052 2.026-.052 9.457-.052.884.208 1.715.624 2.442z" fill="url(#paint2_linear_4784_322)"/><defs><linearGradient id="paint0_linear_4784_322" x1="152.306" y1="93.364" x2="152.306" y2="88.247" gradientUnits="userSpaceOnUse"><stop stop-color="#A1A1A1"/><stop offset=".991" stop-color="#8A8A8A"/></linearGradient><linearGradient id="paint1_linear_4784_322" x1="86.263" y1="95.159" x2="86.263" y2="91.403" gradientUnits="userSpaceOnUse"><stop stop-color="#A1A1A1"/><stop offset=".991" stop-color="#8A8A8A"/></linearGradient><linearGradient id="paint2_linear_4784_322" x1="119.477" y1="94.103" x2="119.477" y2="86.564" gradientUnits="userSpaceOnUse"><stop stop-color="#A1A1A1"/><stop offset=".991" stop-color="#8A8A8A"/></linearGradient></defs></svg>`));
 export const UserGroupLight = (/*#__PURE__*/createFluentIllustration('UserGroupLight', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M118.982 132.7c28 0 50.7-22.7 50.7-50.8 0-28.1-22.8-50.8-50.7-50.8-28 0-50.7 22.7-50.7 50.8 0 28.1 22.7 50.8 50.7 50.8z" fill="#EBEBEB"/><path d="M166.382 48.6a4.1 4.1 0 100-8.2 4.1 4.1 0 000 8.2zM172.382 32.6a2.8 2.8 0 100-5.6 2.8 2.8 0 000 5.6zM70.982 48.5a2.8 2.8 0 100-5.6 2.8 2.8 0 000 5.6zM59.2 109.028a5.2 5.2 0 100-10.4 5.2 5.2 0 000 10.4z" fill="#E6DED1"/><g filter="url(#filter0_d_4784_289)"><path d="M172.94 104.529h-15.381c-.727 0-1.351-.468-1.559-1.091-1.507-4.469-4.832-5.56-12.887-8.73-.467-.208-.519-.831-.052-1.091 2.286-1.118 1.94-2.086 1.924-8.538-.001-.287-.132-.559-.261-.816a6.186 6.186 0 01-.937-2.304.691.691 0 00-.362-.502.782.782 0 00-.437-.063c-1.32.147-1.247-1.507-1.641-4.561-.13-1.126-.008-1.494.368-1.706.159-.09.314-.19.426-.334a.984.984 0 00.12-.982c-2.383-5.667-1.542-9.523.628-11.445.565-.5.514-2.119.285-2.838-.278-.875.653-.738 1.562-.611 2.059.288 5.35.662 6.795.352 2.294-.467 5.56-.101 7.556 2.064.899.974 1.874 1.923 2.94 2.712 1.339.99 1.942 2.25 2.132 3.538.35 2.806-1.135 5.854-1.521 7.186a.321.321 0 00.037.262c.016.023.041.041.066.054.387.212.529.587.378 1.696-.157 1.328-.237 2.442-.403 3.24-.275 1.316-1.941 2.105-2.611 3.27-.171.3-.33.62-.361.964l-.002.06c-.052.468-.052 1.351-.052 6.444 0 .623.156 1.195.467 1.663.308.564.818.976 1.431 1.237.016.006.033.01.05.01s.035.003.05.01c3.625 1.501 6.216 2.486 8.135 3.263 4.261 1.715 4.988 2.183 5.352 5.041.104 1.403-.884 2.546-2.235 2.546z" fill="url(#paint0_linear_4784_289)"/></g><path d="M162.494 74.857c.156-.78-.207-1.715-.519-2.39-.052-.104-.104-.208-.156-.364-1.195-2.442-3.949-3.481-6.548-3.585-6.651-.364-7.274.935-9.353-.987.624.987.728 2.494-.312 4.416-.675 1.351-2.026 1.923-3.221 2.339-3.014-6.86-1.507-11.172 1.507-12.731-.988-2.183-1.04-2.962-1.04-2.962s6.392 1.091 8.574.623c2.806-.571 7.067.104 8.678 3.794 2.65 1.039 3.689 2.754 3.949 4.52.416 2.91-1.247 6.08-1.559 7.327z" fill="#676767"/><path opacity=".1" d="M171.328 104.528c1.611-2.494.987-4.832-1.611-7.586 4.261 1.715 4.989 2.182 5.353 5.04.207 1.299-.832 2.494-2.183 2.494h-1.559v.052z" fill="#B7B4AF"/><path d="M160.052 92.473v1.195h-15.433v-1.559c.364-.831.26-.78.26-7.483a1.581 1.581 0 01-.26-.363v-.364c.312.572.779 1.091 1.247 1.507l4.833 3.377c1.143.936 2.754.988 3.897.052l4.521-4.053c.156-.156.312-.26.467-.415-.052.467-.052 1.35-.052 6.443.052.572.208 1.143.52 1.663z" fill="url(#paint1_linear_4784_289)"/><g filter="url(#filter1_d_4784_289)"><path d="M93.435 92.94a12.782 12.782 0 01-1.546-.111c-.84-.122-1.208-.854-1.208-1.704v-.501c0-1.495.694-2.884 1.755-3.936 1.038-1.029 1.946-2.144 2.627-3.379.969-1.758 2.842-3.299 3.502-5.195a3.4 3.4 0 00.17-1.438c-.052-.728-.52-1.403-.571-2.182v-1.092c0-.082 0-.16-.002-.235-.007-.33-.056-.658-.086-.986a3.077 3.077 0 01-.016-.286c-.156-1.247-.468-2.546-1.3-4.053-2.078-3.897-6.183-6.495-10.808-6.495a12.1 12.1 0 00-2.754.312c-1.662.363-3.17 1.09-4.52 2.13l-.78.624c-.122.122-.273.215-.397.335-1.365 1.322-2.436 2.97-3.032 4.81a13.259 13.259 0 00-.52 5.87v.157c0 .104.005.208.029.309.032.136.058.184.008.279-.075.14-.175.244-.26.377-.066.102-.12.23-.173.34-.27.553-.454 1.275-.124 2.124.59 1.673 1.74 1.53 2.83 2.346.102.077.104.252-.024.252a.333.333 0 00-.101.016l-1.25.4c-7.249 2.272-8.52 9.577-2.225 11.8.244.087.305.885.095 1.035-2.546 1.975-4.157 4.989-4.573 7.951-.156.883.572 1.714 1.507 1.714h12.367c.26 0 .52-.208.624-.467 1.143-4.157 3.741-5.612 9.145-7.795l3.118-1.247c.52-.208 1.091-.416 1.715-.675.572-.208.52-1.04-.104-1.196a9.363 9.363 0 00-1.414-.22 6.217 6.217 0 00-.914-.004c-.312.016-.575.016-.79.016zm-15.849-5.3c0-.115.018-.233.055-.353.28-.926 1.56-.705 2.284-.062.364.311.727.623 1.143.935.26.156.52.364.78.572l.043.043c.062.056.147.092.216.138.07.047.104.125.104.208v2.25c0 .06-.076.08-.136.065a.088.088 0 01-.039-.018c-.25-.202-.454-.404-.656-.556-.364-.26-.728-.468-1.091-.675-.248-.166-.49-.342-.739-.506-1.04-.688-1.917-1.344-1.964-2.04z" fill="url(#paint2_linear_4784_289)"/></g><path d="M93.435 92.94c-.416 1.508-.883 2.443-.883 2.443l-13.407-2.13.78-3.326c.104.104.26.156.364.208.363.207.779.467 1.09.727.209.156.468.312.676.572l.104.104V88.99c-.104-.052-.156-.104-.208-.156 2.183 1.092 4.885 1.663 8.782-.623-.052 0-.052.052-.104.052v4.157s.312.468 2.806.52z" fill="url(#paint3_linear_4784_289)"/><path d="M98.06 72.675a16.865 16.865 0 01-7.015.883c-3.742-.312-7.379-1.819-10.185-4.313-.935 2.702-3.274 4.729-5.924 5.924-.364.156-.727.312-1.091.416v-.104a13.26 13.26 0 01.52-5.872c.623-1.923 1.766-3.637 3.221-4.988.052-.052.104-.104.208-.156l.78-.624a13.378 13.378 0 014.52-2.13 12.434 12.434 0 012.755-.312c4.624 0 8.73 2.598 10.808 6.495.78 1.507 1.143 2.806 1.299 4.001.104.26.104.52.104.78z" fill="#676767"/><path d="M78.885 86.081c.312.416.676.78.988 1.092-.364-.312-.728-.676-.988-1.092z" fill="#3973F4"/><path d="M83.77 99.228c-.987 1.247-2.858 1.299-4.417 1.247 1.663-1.61.728-6.859-5.092-6.235h-.052c-8.003-1.507-7.12-9.77.623-12.212l1.3-.416a.406.406 0 01.207.208 15.108 15.108 0 002.339 4.054c-3.17 2.65 1.247 3.481 3.377 5.508 1.87 1.09 3.43 5.664 1.715 7.846z" fill="#676767"/><path d="M78.834 86.081c-.053-.052-.104-.103-.104-.156 0 0 .051.105.103.156z" fill="#3973F4"/><g filter="url(#filter2_d_4784_289)"><path d="M153.037 106.919c.26 1.922-1.247 3.689-3.222 3.689h-60.9c-1.975 0-3.482-1.767-3.222-3.689.571-4.002 1.455-4.833 7.379-7.223l1.454-.572c2.755-1.091 6.34-2.442 11.173-4.469 1.091-.467 1.974-1.299 2.442-2.39.26-.572.416-1.247.416-1.923v-8.385c0-.46-.211-.869-.416-1.28a8.668 8.668 0 01-1.383-3.424 1.002 1.002 0 00-.552-.75 1.134 1.134 0 00-.626-.077c-1.994.26-1.835-2.219-2.428-6.765-.217-1.65-.035-2.211.546-2.533.242-.135.478-.288.642-.511.294-.4.349-.927.158-1.385-3.504-8.414-2.291-14.108.903-16.893.819-.715.792-3.057.472-4.095-.4-1.297.976-1.134 2.32-.95 3.032.414 7.851.953 9.977.541 3.605-.73 8.935-.057 11.763 3.814.782 1.071 1.741 2.027 2.878 2.71 2.538 1.523 3.611 3.57 3.91 5.791.556 4.145-1.739 8.682-2.296 10.6a.44.44 0 00.025.316c.022.042.058.06.099.082.575.316.761.857.561 2.513-.237 1.876-.349 3.457-.566 4.62-.384 2.059-2.912 3.35-3.951 5.168-.21.367-.377.76-.434 1.178-.01.08-.022.161-.038.258-.104.676-.052 2.027-.052 9.457 0 .884.26 1.767.728 2.443.508.763 1.216 1.426 2.074 1.795.037.016.072.012.108.03.036.018.072.053.109.069 5.323 2.221 9.151 3.67 11.947 4.757 6.391 2.494 7.379 3.274 8.002 7.483z" fill="url(#paint4_linear_4784_289)"/></g><path d="M134.434 66.907c.208-1.143-.312-2.546-.727-3.534-.052-.155-.156-.311-.208-.52-1.819-3.637-5.82-5.092-9.665-5.3-9.821-.52-10.705 1.352-13.771-1.454.936 1.455 1.04 3.689-.415 6.495-1.04 1.975-3.014 2.858-4.729 3.43-4.469-10.133-2.234-16.473 2.183-18.707-1.455-3.274-1.559-4.417-1.559-4.417s9.405 1.559 12.627.935c4.105-.831 10.444.156 12.783 5.612 3.897 1.56 5.404 4.001 5.768 6.703.623 4.21-1.819 8.886-2.287 10.757z" fill="#676767"/><path opacity=".1" d="M147.529 110.556c2.39-3.637 1.455-7.067-2.39-11.12 6.287 2.494 7.326 3.274 7.898 7.431.26 1.922-1.247 3.689-3.222 3.689h-2.286z" fill="#B7B4AF"/><path d="M130.849 92.785v1.766h-22.656v-2.286c.26-.572.415-1.247.415-1.923V81.25c-.155-.208-.311-.364-.415-.572v-.52a7.664 7.664 0 001.87 2.235l7.119 4.988a4.544 4.544 0 005.768.104l6.652-5.975c.208-.208.467-.416.675-.624-.104.675-.052 2.027-.052 9.457-.052.884.208 1.715.624 2.443z" fill="url(#paint5_linear_4784_289)"/><defs><linearGradient id="paint0_linear_4784_289" x1="157.992" y1="57.584" x2="157.992" y2="105.024" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_4784_289" x1="152.306" y1="93.364" x2="152.306" y2="88.247" gradientUnits="userSpaceOnUse"><stop stop-color="#fff" stop-opacity="0"/><stop offset=".991" stop-color="#DAD6CF"/></linearGradient><linearGradient id="paint2_linear_4784_289" x1="83.445" y1="60.348" x2="83.445" y2="104.994" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint3_linear_4784_289" x1="86.263" y1="95.159" x2="86.263" y2="91.403" gradientUnits="userSpaceOnUse"><stop stop-color="#fff" stop-opacity="0"/><stop offset=".991" stop-color="#DAD6CF"/></linearGradient><linearGradient id="paint4_linear_4784_289" x1="119.343" y1="41.334" x2="119.343" y2="111.338" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint5_linear_4784_289" x1="119.477" y1="94.103" x2="119.477" y2="86.564" gradientUnits="userSpaceOnUse"><stop stop-color="#fff" stop-opacity="0"/><stop offset=".991" stop-color="#DAD6CF"/></linearGradient><filter id="filter0_d_4784_289" x="118.824" y="47.823" width="78.358" height="89.706" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4784_289"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4784_289" result="shape"/></filter><filter id="filter1_d_4784_289" x="46.16" y="50.347" width="74.591" height="87.181" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4784_289"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4784_289" result="shape"/></filter><filter id="filter2_d_4784_289" x="63.663" y="32.17" width="111.403" height="111.438" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4784_289"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4784_289" result="shape"/></filter></defs></svg>`));
 export const UserLight = (/*#__PURE__*/createFluentIllustration('UserLight', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M169.779 70.89c0 13.418-4.782 25.624-12.777 35.189-10.064 12.062-25.196 19.7-42.184 19.7-16.274 0-30.835-7.067-40.9-18.273A54.955 54.955 0 0160 70.889C60 40.554 84.554 16 114.889 16c30.336 0 54.89 24.554 54.89 54.89z" fill="#EBEBEB"/><g filter="url(#filter0_d_4784_240)"><path d="M151.863 103.723c2.844.914 3.46 4.474 1.318 6.556-9.881 9.608-23.37 15.5-38.292 15.5-14.357 0-27.38-5.5-37.16-14.49-2.123-1.952-1.81-5.307.79-6.555a34.017 34.017 0 011.966-.868l1.999-.785c3.783-1.499 8.708-3.355 15.346-6.139 1.5-.642 2.712-1.784 3.355-3.283a6.488 6.488 0 00.571-2.641V79.642c0-.631-.289-1.193-.571-1.757a11.887 11.887 0 01-1.899-4.704 1.378 1.378 0 00-.76-1.029c-.265-.124-.567-.144-.858-.107-2.746.357-2.589-3.048-3.335-9.293-.299-2.266-.049-3.037.749-3.48.333-.184.656-.395.882-.701.404-.548.479-1.273.217-1.902-4.796-11.516-3.159-19.323 1.193-23.164 1.138-1.004 1.038-4.27.566-5.71-.577-1.765 1.32-1.515 3.16-1.264 4.169.57 10.818 1.313 13.789.747 5.145-1.042 12.845-.005 16.594 5.877.713 1.118 1.675 2.076 2.845 2.7 3.988 2.127 5.669 5.097 6.043 8.34.774 5.765-2.471 12.085-3.181 14.66a.353.353 0 00.046.25.717.717 0 00.242.247c.719.442.946 1.214.681 3.4-.401 3.167-.541 5.722-1.009 7.351-.626 2.181-3.559 3.23-4.684 5.2l-.603 1.055c-.227.397-.412.822-.479 1.274-.025.173-.043.33-.078.538-.143.928-.071 2.784-.071 12.99 0 1.214.357 2.356.999 3.427.7 1.05 1.675 1.963 2.924 2.469.049.02.067.021.117.037.052.017.127.075.178.096 7.311 3.051 12.569 5.041 16.41 6.534z" fill="url(#paint0_linear_4784_240)"/></g><path d="M137.302 58.898c.285-1.57-.428-3.498-.999-4.854-.072-.214-.215-.428-.286-.713-2.498-4.997-7.994-6.995-13.276-7.281-13.491-.714-14.704 1.856-18.915-1.998 1.285 1.998 1.427 5.067-.571 8.922-1.428 2.712-4.14 3.925-6.496 4.71-6.138-13.918-3.069-22.626 2.998-25.695-1.998-4.426-2.141-5.996-2.141-5.996s12.919 2.141 17.345 1.285c5.639-1.142 14.347.214 17.559 7.709 5.353 2.14 7.423 5.496 7.923 9.207.785 5.782-2.499 12.134-3.141 14.704z" fill="#676767"/><path d="M132.377 94.444v2.427h-31.192v-3.14a6.488 6.488 0 00.571-2.642V78.67c-.214-.285-.428-.499-.571-.784v-.714c.642 1.142 1.499 2.212 2.569 3.069l9.779 6.78a6.241 6.241 0 007.923.144l9.136-8.209c.286-.285.643-.57.928-.856-.142.927-.071 2.783-.071 12.99a5.426 5.426 0 00.928 3.355z" fill="url(#paint1_linear_4784_240)"/><defs><linearGradient id="paint0_linear_4784_240" x1="115.505" y1="23.685" x2="115.505" y2="126.855" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_4784_240" x1="116.721" y1="96.257" x2="116.721" y2="85.938" gradientUnits="userSpaceOnUse"><stop stop-color="#fff" stop-opacity="0"/><stop offset=".991" stop-color="#DAD6CF"/></linearGradient><filter id="filter0_d_4784_240" x="54.329" y="15.354" width="122.133" height="143.424" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4784_240"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4784_240" result="shape"/></filter></defs></svg>`));
-export const UserProfileDark = (/*#__PURE__*/createFluentIllustration('UserProfileDark', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M119.711 134.141c30.037 0 54.389-24.298 54.389-54.376S149.641 25.39 119.711 25.39c-30.038 0-54.39 24.298-54.39 54.376s24.352 54.376 54.39 54.376z" fill="#242424"/><path d="M180 23.997a3 3 0 01-3.004 2.997 3 3 0 01-3.003-2.997A3 3 0 01176.996 21 3 3 0 01180 23.997zM174.958 39.732a4.393 4.393 0 01-4.398 4.389 4.394 4.394 0 01-4.399-4.389 4.394 4.394 0 014.399-4.389 4.393 4.393 0 014.398 4.389zM68.217 44.014a3 3 0 003.004-2.998 3 3 0 00-3.004-2.997 3 3 0 00-3.003 2.998 3 3 0 003.003 2.997zM61.157 103.237a5.572 5.572 0 01-5.579 5.566A5.572 5.572 0 0150 103.237a5.572 5.572 0 015.578-5.566 5.572 5.572 0 015.579 5.566z" fill="#3A3A3A"/><path d="M163.117 115.245H76.303c-3.991 0-7.234-3.236-7.234-7.218V52.519c0-3.983 3.243-7.219 7.234-7.219h86.814c3.992 0 7.235 3.236 7.235 7.219v55.757c-.25 3.733-3.493 6.969-7.235 6.969z" fill="url(#paint0_linear_4784_278)"/><path d="M99.905 47.858H84.871v5.516h15.034v-5.516z" fill="url(#paint1_linear_4784_278)"/><path d="M153.795 47.858h-15.033v5.516h15.033v-5.516z" fill="url(#paint2_linear_4784_278)"/><path d="M87.03 35.343h10.727V50.4H87.03V35.343z" fill="url(#paint3_linear_4784_278)"/><path d="M140.92 35.343h10.727V50.4H140.92V35.343z" fill="url(#paint4_linear_4784_278)"/><path fill-rule="evenodd" clip-rule="evenodd" d="M100.501 64.815c-5.731 0-10.377 4.636-10.377 10.355 0 3.161 1.42 5.992 3.658 7.891-5.1 2.51-8.648 7.764-8.648 13.768 0 1.182.96 2.14 2.145 2.14a2.143 2.143 0 002.146-2.14c0-6.036 5.027-11.052 11.076-11.052 6.05 0 11.076 5.016 11.076 11.052 0 1.182.961 2.14 2.146 2.14a2.143 2.143 0 002.145-2.14c0-6.004-3.548-11.258-8.647-13.767a10.324 10.324 0 003.658-7.892c0-5.719-4.646-10.355-10.378-10.355zM94.415 75.17c0-3.354 2.725-6.073 6.086-6.073 3.362 0 6.087 2.719 6.087 6.073 0 3.354-2.725 6.073-6.087 6.073-3.361 0-6.086-2.719-6.086-6.073z" fill="#CBCBCB"/><path d="M125.298 73.676c0-1.182.96-2.14 2.145-2.14h26.693c1.185 0 2.146.958 2.146 2.14 0 1.182-.961 2.14-2.146 2.14h-26.693a2.143 2.143 0 01-2.145-2.14zM127.443 84.977a2.143 2.143 0 00-2.145 2.14c0 1.183.96 2.141 2.145 2.141h10.478a2.143 2.143 0 002.145-2.14c0-1.183-.96-2.141-2.145-2.141h-10.478z" fill="#CBCBCB"/><defs><linearGradient id="paint0_linear_4784_278" x1="119.677" y1="43.682" x2="119.677" y2="115.999" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint1_linear_4784_278" x1="146.279" y1="47.858" x2="146.279" y2="53.374" gradientUnits="userSpaceOnUse"><stop stop-color="#363636"/><stop offset="1" stop-color="#1D1D1D"/></linearGradient><linearGradient id="paint2_linear_4784_278" x1="146.279" y1="47.858" x2="146.279" y2="53.374" gradientUnits="userSpaceOnUse"><stop stop-color="#363636"/><stop offset="1" stop-color="#1D1D1D"/></linearGradient><linearGradient id="paint3_linear_4784_278" x1="146.28" y1="34.995" x2="146.28" y2="50.561" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint4_linear_4784_278" x1="146.28" y1="34.995" x2="146.28" y2="50.561" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient></defs></svg>`));
-export const UserProfileLight = (/*#__PURE__*/createFluentIllustration('UserProfileLight', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M119.711 134.391c30.037 0 54.389-24.351 54.389-54.496 0-30.145-24.459-54.497-54.389-54.497-30.038 0-54.39 24.352-54.39 54.497s24.352 54.496 54.39 54.496z" fill="#EBEBEB"/><path d="M170.56 44.172a4.398 4.398 0 100-8.797 4.398 4.398 0 000 8.797zM176.997 27.008a3.004 3.004 0 10-.001-6.008 3.004 3.004 0 00.001 6.008zM68.218 44.064a3.004 3.004 0 100-6.007 3.004 3.004 0 000 6.007zM55.579 108.997a5.579 5.579 0 10-.001-11.158 5.579 5.579 0 000 11.158z" fill="#E8DFD3"/><g filter="url(#filter0_d_4784_267)"><path d="M163.117 115.453H76.304a7.24 7.24 0 01-7.235-7.234v-55.63a7.24 7.24 0 017.235-7.235h86.813a7.24 7.24 0 017.235 7.234v55.88c-.25 3.742-3.493 6.985-7.235 6.985z" fill="url(#paint0_linear_4784_267)"/></g><path d="M84.872 48.99c0-.593.48-1.073 1.073-1.073h12.888c.593 0 1.073.48 1.073 1.073v3.382c0 .593-.48 1.073-1.073 1.073H85.945c-.592 0-1.073-.48-1.073-1.073V48.99zM138.762 48.99c0-.593.48-1.073 1.072-1.073h12.889c.592 0 1.073.48 1.073 1.073v3.382c0 .593-.481 1.073-1.073 1.073h-12.889c-.592 0-1.072-.48-1.072-1.073V48.99z" fill="#D4CDC0"/><g filter="url(#filter1_d_4784_267)"><path d="M96.685 35.375h-8.581c-.593 0-1.073.48-1.073 1.073V49.39c0 .592.48 1.073 1.073 1.073h8.581c.593 0 1.073-.48 1.073-1.073V36.448c0-.593-.48-1.073-1.073-1.073z" fill="url(#paint1_linear_4784_267)"/><path d="M150.575 35.375h-8.582c-.592 0-1.072.48-1.072 1.073V49.39c0 .592.48 1.073 1.072 1.073h8.582c.592 0 1.073-.48 1.073-1.073V36.448c0-.593-.481-1.073-1.073-1.073z" fill="url(#paint2_linear_4784_267)"/></g><path fill-rule="evenodd" clip-rule="evenodd" d="M100.501 64.912c-5.731 0-10.378 4.646-10.378 10.377 0 3.169 1.42 6.006 3.659 7.91-5.099 2.514-8.648 7.78-8.648 13.797a2.146 2.146 0 004.291 0c0-6.05 5.027-11.076 11.077-11.076 6.049 0 11.076 5.027 11.076 11.076a2.145 2.145 0 104.291 0c0-6.017-3.549-11.283-8.648-13.798a10.356 10.356 0 003.658-7.909c0-5.731-4.646-10.377-10.378-10.377zm-6.087 10.377a6.087 6.087 0 1112.174 0 6.087 6.087 0 01-12.174 0z" fill="#BCB9B3"/><path d="M125.299 73.793c0-1.185.96-2.146 2.145-2.146h26.693a2.145 2.145 0 010 4.291h-26.693a2.145 2.145 0 01-2.145-2.145zM127.444 85.118a2.145 2.145 0 000 4.291h10.478a2.145 2.145 0 000-4.29h-10.478z" fill="#BCB9B3"/><defs><linearGradient id="paint0_linear_4784_267" x1="119.678" y1="43.732" x2="119.678" y2="116.209" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_4784_267" x1="146.281" y1="35.026" x2="146.281" y2="50.626" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint2_linear_4784_267" x1="146.281" y1="35.026" x2="146.281" y2="50.626" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><filter id="filter0_d_4784_267" x="57.069" y="41.354" width="125.283" height="94.1" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="8"/><feGaussianBlur stdDeviation="6"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.17 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4784_267"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4784_267" result="shape"/></filter><filter id="filter1_d_4784_267" x="65.031" y="24.375" width="108.617" height="59.089" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4784_267"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4784_267" result="shape"/></filter></defs></svg>`));
-export const WelcomePlaneDark = (/*#__PURE__*/createFluentIllustration('WelcomePlaneDark', `<svg viewBox="0 0 437 176" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M311.498 1.684c1.374-.383 2.663-.48 4.037-.48 1.288 0 2.663.097 3.951.288 1.289.192 2.663.576 3.951 1.248 1.289.671 2.577 1.63 3.35 3.262l-1.46-.384 1.546-1.343c.601-.384 1.117-.864 1.632-1.248 1.117-.767 2.319-1.439 3.608-1.918 1.288-.48 2.577-.864 3.865-1.056 1.374-.096 2.663-.096 4.037.288a.44.44 0 01.258.576c-.086.288-.258.384-.515.288h-.086c-1.117-.288-2.405-.192-3.608 0-1.203.191-2.405.671-3.522 1.247-1.116.575-2.147 1.247-3.178 2.015-.515.383-1.031.863-1.46 1.247l-1.374 1.343c-.43.384-.945.288-1.289-.096l-.086.096-.086-.192c-.343-1.055-1.288-1.919-2.319-2.494a13.183 13.183 0 00-3.522-1.344 24.053 24.053 0 00-3.779-.575c-1.288-.096-2.577-.096-3.693.192h-.086c-.258.096-.43-.096-.516-.384 0-.288.086-.576.344-.576zM280.803 27.464c1.031-.269 1.997-.336 3.028-.336.966 0 1.997.067 2.963.201.967.135 1.997.403 2.964.873.966.47 1.932 1.142 2.512 2.284l-1.095-.269 1.16-.94c.451-.269.837-.605 1.224-.873a12.827 12.827 0 012.705-1.343c.967-.336 1.933-.605 2.899-.739 1.031-.067 1.997-.067 3.028.201.193.068.258.27.193.403-.064.202-.193.269-.386.202h-.065c-.837-.202-1.803-.134-2.705 0-.902.134-1.804.47-2.642.873-.837.403-1.61.873-2.383 1.41-.387.269-.773.605-1.095.873l-1.031.94c-.322.27-.709.202-.966-.067l-.065-.067-.064-.134c-.258-.739-.967-1.343-1.74-1.746a10.333 10.333 0 00-2.641-.94 19.257 19.257 0 00-2.835-.403c-.966-.068-1.932-.068-2.77.134h-.064c-.193.067-.322-.067-.387-.269 0-.067.129-.268.258-.268zM98.698 1.684c1.374-.383 2.663-.48 4.037-.48 1.289 0 2.663.097 3.951.288 1.289.192 2.663.576 3.952 1.248 1.288.671 2.576 1.63 3.349 3.262l-1.46-.384 1.546-1.343c.602-.384 1.117-.864 1.632-1.248 1.117-.767 2.32-1.439 3.608-1.918 1.288-.48 2.577-.864 3.865-1.056 1.375-.096 2.663-.096 4.037.288.258.096.344.384.258.576-.086.288-.258.384-.515.288h-.086c-1.117-.288-2.405-.192-3.608 0-1.202.191-2.405.671-3.522 1.247-1.116.575-2.147 1.247-3.178 2.015-.515.383-1.03.863-1.46 1.247l-1.374 1.343c-.43.384-.945.288-1.289-.096l-.086-.096-.085-.192c-.344-1.055-1.289-1.919-2.32-2.494a13.175 13.175 0 00-3.521-1.343 24.084 24.084 0 00-3.78-.576c-1.288-.096-2.577-.096-3.693.192h-.086c-.258.096-.43-.096-.516-.384 0-.096.172-.384.344-.384zM320.243 24.15c1.375-.346 2.663-.432 4.038-.432 1.288 0 2.662.086 3.951.259 1.288.173 2.663.518 3.951 1.122 1.288.605 2.577 1.468 3.35 2.936l-1.46-.345 1.546-1.209c.601-.345 1.116-.777 1.632-1.123 1.116-.69 2.319-1.295 3.607-1.726 1.289-.432 2.577-.777 3.866-.95 1.374-.087 2.662-.087 4.037.259.258.086.343.345.257.518-.085.259-.257.345-.515.259h-.086c-1.116-.26-2.405-.173-3.607 0-1.203.173-2.405.604-3.522 1.122a23.204 23.204 0 00-3.178 1.814c-.516.345-1.031.777-1.46 1.122l-1.375 1.21c-.429.345-.945.258-1.288-.087l-.086-.087-.086-.172c-.344-.95-1.288-1.727-2.319-2.245a14.107 14.107 0 00-3.522-1.21 26.578 26.578 0 00-3.779-.517c-1.289-.087-2.577-.087-3.694.172h-.086c-.257.087-.429-.086-.515-.345 0-.086.086-.345.343-.345z" fill="#5C5C5C"/><path d="M239.176 42.5l58.277 45.434h12.931M238.243 131.772l59.21-43.838" stroke="#8C8C8C" stroke-width="5" stroke-miterlimit="10"/><g filter="url(#filter0_d_3869_2076)"><path d="M315.711 79.531l-.316-1.431s1.216-2.215 1.826-3.36c.689-1.143 1.327-1.344 4.242-1.582 2.919-.316 5.138-.747 4.149 1.636-1.068 2.38-4.053 8.35-4.053 8.35l-5.848-3.613z" fill="#484848"/></g><g filter="url(#filter1_d_3869_2076)"><path d="M377.15 82.104l20.344-28.541a3.247 3.247 0 011.792-1.24c1.35-.321 3.656-.906 5.871-1.259 3.323-.53 7.115-1.035 5.594 2.5-1.526 3.611-11.283 27.767-11.283 27.767l-22.318.773z" fill="#484848"/></g><g filter="url(#filter2_d_3869_2076)"><path d="M314.482 76.086c.235.013 10.566 9.287 10.566 9.287l34.783-5.058 19.658.578s16.579-.53 21.664-.258c5.084.272 11.847 1.419 12.438 5.065.59 3.647 2.035 4.51-.157 5.886-2.112 1.38-10.502 7.22-22.07 6.444-11.569-.774-25.267 1.638-25.267 1.638s-.077-1.498-3.832-1.698c-3.754-.2-50.143-7-50.143-7s-7.674-.252-6.801-7.829c.952-7.572 3.525-7.277 5.172-7.268 1.729-.065 3.989.213 3.989.213z" fill="url(#paint0_linear_3869_2076)"/></g><g filter="url(#filter3_d_3869_2076)"><path d="M312.97 83.864l7.353.393s-4.339 6.37-5.905 7.779c-1.565 1.41-4.593 3.763-7.548 3.29-2.956-.472-2.164-2.08-1.282-3.919.882-1.839 1.888-4.536 1.888-4.536l5.494-3.007z" fill="#6C6C6C"/></g><g filter="url(#filter4_d_3869_2076)"><path d="M391.929 96.253c-5.948-1.732-9.921-2.259-16.912-.589-6.992 1.67-9.73 3.017-9.73 3.017L351.207 123c-.866 1.525.212 3.39 1.933 3.482 3.445.106 8.489-.332 11.385-3.164 4.639-4.625 27.404-27.064 27.404-27.064z" fill="#6C6C6C"/></g><path opacity=".4" d="M422.124 87.492l-.071 1.332c-.866 8.913-5.622 15.732-11.019 15.444-3.989-.213-7.3-4.32-8.638-10.128a27.495 27.495 0 01-.625-7.42l.016-.314c.309-5.799 2.222-10.805 4.97-13.802 1.751-1.95 3.93-3.09 6.198-2.97l.391.021c4.833.573 8.501 6.82 8.945 14.702-.133 1.015-.109 2.038-.167 3.135z" fill="#6C6C6C"/><path fill-rule="evenodd" clip-rule="evenodd" d="M390.064 85.784c.434-2.079 1.627-4.553 3.836-5.558l.385.85c-1.809.824-2.9 2.94-3.309 4.903l-.001.007c-.792 3.516-.01 7.22 1.585 10.463.272.54.612 1.071.946 1.444v.096c-.333.126-.751.085-1.051.085a8.684 8.684 0 01-.729-1.206l-.002-.004c-1.659-3.373-2.51-7.298-1.66-11.08z" fill="#8A8A8A"/><path d="M359.041 80.43l-.52 2.408 5.979 4.092 11.548-.327c1.021-.024 1.849-.844 1.899-1.785l.063-1.175 1.48-2.75-1.811-1.354a4.141 4.141 0 00-2.309-.83l-9.963.017a2.502 2.502 0 00-1.108.177l-5.258 1.527z" fill="#3A3A3A"/><path fill-rule="evenodd" clip-rule="evenodd" d="M373.784 80.74l4.453 2.495-.456.816-4.231-2.37-9.167.204-5.752 1.407-.222-.908 5.851-1.43 9.524-.213z" fill="#6C6C6C"/><path fill-rule="evenodd" clip-rule="evenodd" d="M374.265 78.757l-.572 4.929.537 2.867-.917.172-.563-3.006.588-5.07.927.108z" fill="#6C6C6C"/><path d="M364.499 86.93l-.508-3.721.212-2.504.256-1.872" fill="#fff"/><path fill-rule="evenodd" clip-rule="evenodd" d="M364.668 80.757l.255-1.86-.925-.128-.258 1.885-.217 2.567.515 3.772.925-.127-.502-3.67.207-2.439z" fill="#6C6C6C"/><path opacity=".6" d="M413.194 69.569l-.701 13.165-1.173-.063-.849 1.212-3.792-11.283c1.751-1.95 3.929-3.09 6.198-2.97.082-.074.238-.065.317-.061zM422.124 87.492l-.071 1.332-5.014-.11.062-1.176-1.21-.85 6.321-2.335c-.054 1.019-.03 2.042-.088 3.139zM410.583 89.155l-8.266 4.98a27.495 27.495 0 01-.625-7.42l7.822.418-.063 1.175 1.132.847z" fill="#777"/><path opacity=".3" d="M364.079 100.739c2.363-.188 5.122-.434 5.122-.434l-11.96 19.873-4.471-.003" fill="#D1D1D1"/><path opacity=".4" d="M382.07 75.137l4.381.234 12.895-19.744-3.207-.172" fill="#A3A3A3"/><path d="M391.818 89.489a.548.548 0 10.057-1.095.548.548 0 00-.057 1.095zM392.195 92.73a.548.548 0 10.057-1.094.548.548 0 00-.057 1.095zM393.284 95.854a.548.548 0 10.057-1.095.548.548 0 00-.057 1.095zM392.31 86.136a.549.549 0 10.057-1.097.549.549 0 00-.057 1.097zM393.958 83.159a.548.548 0 10.059-1.095.548.548 0 00-.059 1.095z" fill="#8A8A8A"/><path opacity=".6" d="M410.878 104.259l.31-13.186 1.173.063.849-1.212 4.109 11.221c-1.755 2.028-3.855 3.174-6.049 3.135l-.392-.021z" fill="#777"/><path d="M411.321 82.671c-.824.742-3.902 5.529-.159 10.365 4.023-1.907 6.844-3.328 6.731-5.613-.032-2.36-6.572-4.752-6.572-4.752z" fill="#8A8A8A"/><path d="M235.325 151.867H62.436c-3.092 0-5.622-2.535-5.622-5.633V29.727c0-3.098 2.53-5.633 5.622-5.633h172.889c3.092 0 5.623 2.535 5.623 5.633V146.14c.093 3.192-2.437 5.727-5.623 5.727z" fill="url(#paint1_linear_3869_2076)"/><g filter="url(#filter5_d_3869_2076)"><path d="M65.905 72.57c0 5.269-4.253 9.529-9.557 9.529H24.43C18.076 82.236 13 77.151 13 71.013c0-6.185 5.121-11.316 11.569-10.995C30.1 42.702 55.433 45.13 57.583 63.042c4.755.595 8.322 4.626 8.322 9.528z" fill="#424242"/></g><path d="M40.755 82.19c9.374 0 17.01-7.604 17.01-16.995 0-9.392-7.636-16.996-17.01-16.996-9.373 0-17.01 7.604-17.01 16.996 0 9.39 7.591 16.996 17.01 16.996z" fill="url(#paint2_linear_3869_2076)"/><path d="M56.348 82.099c5.259 0 9.557-4.26 9.557-9.529 0-5.268-4.298-9.528-9.557-9.528-5.258 0-9.556 4.26-9.556 9.528 0 5.269 4.298 9.529 9.556 9.529z" fill="url(#paint3_linear_3869_2076)"/><g filter="url(#filter6_d_3869_2076)"><path d="M337.629 121.374c0 4.644-3.749 8.4-8.425 8.4h-28.139c-5.603.122-10.078-4.362-10.078-9.774 0-5.452 4.515-9.976 10.199-9.693 4.878-15.267 27.212-13.126 29.106 2.666 4.193.525 7.337 4.079 7.337 8.401z" fill="#424242"/></g><path d="M315.457 129.855c8.264 0 14.997-6.704 14.997-14.984 0-8.28-6.733-14.984-14.997-14.984s-14.997 6.704-14.997 14.984c0 8.28 6.692 14.984 14.997 14.984z" fill="url(#paint4_linear_3869_2076)"/><path d="M329.205 129.774c4.636 0 8.425-3.756 8.425-8.401 0-4.644-3.789-8.4-8.425-8.4-4.636 0-8.426 3.756-8.426 8.4 0 4.645 3.79 8.401 8.426 8.401z" fill="url(#paint5_linear_3869_2076)"/><g filter="url(#filter7_d_3869_2076)"><path d="M397.333 145.893c0 3.345-2.7 6.049-6.067 6.049H371.006c-4.034.087-7.256-3.141-7.256-7.037 0-3.926 3.251-7.183 7.343-6.979 3.513-10.993 19.593-9.451 20.957 1.919 3.019.378 5.283 2.937 5.283 6.048z" fill="#424242"/></g><path d="M381.369 152c5.951 0 10.798-4.827 10.798-10.789 0-5.961-4.847-10.788-10.798-10.788-5.95 0-10.797 4.827-10.797 10.788A10.78 10.78 0 00381.369 152z" fill="url(#paint6_linear_3869_2076)"/><path d="M391.267 151.942c3.338 0 6.066-2.705 6.066-6.049 0-3.344-2.728-6.048-6.066-6.048-3.338 0-6.066 2.704-6.066 6.048s2.728 6.049 6.066 6.049z" fill="url(#paint7_linear_3869_2076)"/><path d="M240.977 39.594H56.844V29.89c0-3.295 2.53-5.99 5.622-5.99h172.889c3.092 0 5.622 2.695 5.622 5.99v9.704z" fill="#3A3A3A"/><path d="M65.72 34.617a2.72 2.72 0 002.717-2.722 2.72 2.72 0 00-2.717-2.723 2.72 2.72 0 00-2.718 2.723 2.72 2.72 0 002.718 2.723z" fill="#CCC"/><path d="M73.966 34.617a2.72 2.72 0 002.718-2.722 2.72 2.72 0 00-2.718-2.723 2.72 2.72 0 00-2.717 2.723 2.72 2.72 0 002.717 2.723z" fill="#505050"/><path d="M82.213 34.617a2.72 2.72 0 002.717-2.722 2.72 2.72 0 00-2.717-2.723 2.72 2.72 0 00-2.718 2.723 2.72 2.72 0 002.718 2.723z" fill="#484848"/><defs><linearGradient id="paint0_linear_3869_2076" x1="360.375" y1="77.974" x2="359.225" y2="99.526" gradientUnits="userSpaceOnUse"><stop stop-color="#6C6C6C"/><stop offset=".996" stop-color="#3B3B3B"/></linearGradient><linearGradient id="paint1_linear_3869_2076" x1="148.822" y1="21.139" x2="148.822" y2="153.244" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint2_linear_3869_2076" x1="27.891" y1="55.259" x2="33.595" y2="59.66" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint3_linear_3869_2076" x1="49.121" y1="67" x2="52.32" y2="69.474" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint4_linear_3869_2076" x1="304.115" y1="106.111" x2="309.144" y2="109.991" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint5_linear_3869_2076" x1="322.833" y1="116.462" x2="325.654" y2="118.643" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint6_linear_3869_2076" x1="373.203" y1="134.905" x2="376.824" y2="137.698" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint7_linear_3869_2076" x1="386.679" y1="142.357" x2="388.71" y2="143.928" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><filter id="filter0_d_3869_2076" x="300.396" y="62.909" width="30.455" height="30.235" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dx="-5"/><feGaussianBlur stdDeviation="5"/><feColorMatrix values="0 0 0 0 0.397708 0 0 0 0 0.47749 0 0 0 0 0.575 0 0 0 0.2 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter1_d_3869_2076" x="355.15" y="39.689" width="77.947" height="75.415" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter2_d_3869_2076" x="283.252" y="64.863" width="153.354" height="67.804" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter3_d_3869_2076" x="289.784" y="73.864" width="35.539" height="31.524" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dx="-5"/><feGaussianBlur stdDeviation="5"/><feColorMatrix values="0 0 0 0 0.397708 0 0 0 0 0.47749 0 0 0 0 0.575 0 0 0 0.2 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter4_d_3869_2076" x="328.901" y="83.649" width="85.027" height="75.846" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter5_d_3869_2076" x="0" y="46.24" width="78.905" height="59.861" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="6.5"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter6_d_3869_2076" x="277.987" y="97.923" width="72.643" height="55.853" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="6.5"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter7_d_3869_2076" x="350.75" y="128.449" width="59.583" height="47.494" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="6.5"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter></defs></svg>`));
-export const WelcomePlaneLight = (/*#__PURE__*/createFluentIllustration('WelcomePlaneLight', `<svg viewBox="0 0 424 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M306.24 1.685c1.378-.384 2.669-.48 4.047-.48 1.291 0 2.668.096 3.96.288 1.291.192 2.669.576 3.96 1.248s2.583 1.631 3.357 3.263l-1.463-.384 1.55-1.344c.602-.383 1.119-.863 1.635-1.247 1.119-.768 2.325-1.44 3.616-1.92s2.583-.864 3.874-1.056c1.378-.096 2.669-.096 4.046.288a.44.44 0 01.259.576c-.086.288-.259.384-.517.288h-.086c-1.119-.288-2.411-.192-3.616 0s-2.41.672-3.529 1.248c-1.12.576-2.153 1.248-3.186 2.015-.516.384-1.033.864-1.463 1.248l-1.378 1.344c-.43.384-.947.288-1.291-.096l-.086.096-.086-.192c-.345-1.056-1.292-1.92-2.325-2.496a13.222 13.222 0 00-3.53-1.343 24.163 24.163 0 00-3.787-.576c-1.292-.096-2.583-.096-3.702.192h-.086c-.259.096-.431-.096-.517-.384 0-.288.086-.576.344-.576zM275.477 27.475c1.033-.269 2.001-.336 3.034-.336.969 0 2.002.067 2.97.201.969.135 2.002.404 2.97.874.969.47 1.937 1.142 2.519 2.284l-1.098-.269 1.162-.94c.452-.269.84-.605 1.227-.874a12.855 12.855 0 012.712-1.343c.968-.336 1.937-.605 2.905-.74 1.033-.067 2.002-.067 3.035.202.194.067.258.269.194.403-.065.202-.194.269-.388.202h-.064c-.84-.202-1.808-.134-2.712 0-.904.134-1.808.47-2.647.873-.84.403-1.615.874-2.389 1.411-.388.269-.775.605-1.098.874l-1.033.94c-.323.269-.71.202-.969-.067l-.064-.067-.065-.135c-.258-.739-.968-1.343-1.743-1.746a10.374 10.374 0 00-2.647-.941 19.344 19.344 0 00-2.841-.403c-.969-.067-1.937-.067-2.777.134h-.064c-.194.067-.323-.067-.388-.268 0-.068.129-.27.259-.27zM92.957 1.685c1.377-.384 2.669-.48 4.046-.48 1.291 0 2.669.096 3.96.288 1.291.192 2.669.576 3.96 1.248 1.292.672 2.583 1.631 3.358 3.263l-1.464-.384 1.55-1.344c.602-.383 1.119-.863 1.636-1.247 1.119-.768 2.324-1.44 3.615-1.92 1.292-.48 2.583-.864 3.874-1.056 1.378-.096 2.669-.096 4.047.288a.441.441 0 01.258.576c-.086.288-.258.384-.517.288h-.086c-1.119-.288-2.41-.192-3.615 0-1.206.192-2.411.672-3.53 1.248-1.119.576-2.152 1.248-3.186 2.015-.516.384-1.033.864-1.463 1.248l-1.378 1.344c-.43.384-.947.288-1.291-.096l-.086-.096-.086-.192c-.344-1.056-1.291-1.92-2.324-2.496a13.248 13.248 0 00-3.53-1.343 24.18 24.18 0 00-3.788-.576c-1.291-.096-2.583-.096-3.702.192h-.086c-.258.096-.43-.096-.517-.384 0-.096.172-.384.345-.384zM315.005 24.16c1.378-.346 2.669-.432 4.047-.432 1.291 0 2.669.086 3.96.259 1.291.173 2.669.518 3.96 1.123 1.291.604 2.583 1.468 3.358 2.937l-1.464-.346 1.55-1.21c.602-.345 1.119-.777 1.635-1.122 1.12-.691 2.325-1.296 3.616-1.728 1.292-.432 2.583-.777 3.874-.95 1.378-.086 2.669-.086 4.047.26.258.086.344.345.258.517-.086.26-.258.346-.517.26h-.086c-1.119-.26-2.41-.173-3.616 0-1.205.172-2.41.604-3.529 1.123a23.236 23.236 0 00-3.186 1.814c-.516.345-1.033.777-1.463 1.123l-1.378 1.209c-.43.345-.947.26-1.291-.087l-.086-.086-.086-.173c-.345-.95-1.292-1.727-2.325-2.245a14.158 14.158 0 00-3.529-1.21 26.71 26.71 0 00-3.788-.518c-1.292-.087-2.583-.087-3.702.173h-.086c-.259.086-.431-.087-.517-.346 0-.086.086-.345.344-.345z" fill="#E8DFD3"/><path d="M233.753 42.516l58.41 45.453h12.961M232.818 131.825l59.345-43.856" stroke="#E4E0DA" stroke-width="6" stroke-miterlimit="10"/><path d="M310.463 79.563l-.316-1.432s1.218-2.215 1.83-3.362c.69-1.143 1.329-1.344 4.251-1.582 2.926-.316 5.15-.747 4.158 1.637-1.07 2.38-4.062 8.353-4.062 8.353l-5.861-3.614z" fill="url(#paint0_linear_3869_2050)"/><path d="M372.043 82.137l20.39-28.553a3.25 3.25 0 011.796-1.24c1.353-.322 3.664-.906 5.884-1.26 3.331-.53 7.131-1.034 5.606 2.5-1.529 3.614-11.307 27.78-11.307 27.78l-22.369.773z" fill="url(#paint1_linear_3869_2050)"/><path d="M309.232 76.117c.236.012 10.591 9.29 10.591 9.29l34.862-5.06 19.702.579s16.617-.53 21.713-.258c5.095.271 11.874 1.419 12.466 5.067.592 3.648 2.039 4.51-.157 5.888-2.117 1.38-10.526 7.222-22.121 6.447-11.594-.775-25.324 1.638-25.324 1.638s-.077-1.498-3.84-1.699c-3.763-.2-50.257-7.002-50.257-7.002s-7.692-.253-6.816-7.832c.954-7.576 3.533-7.28 5.183-7.272 1.733-.064 3.998.213 3.998.213z" fill="url(#paint2_linear_3869_2050)"/><path d="M307.717 83.898l7.369.392s-4.349 6.372-5.918 7.783c-1.569 1.41-4.603 3.764-7.566 3.292-2.962-.473-2.169-2.081-1.285-3.921.885-1.84 1.893-4.538 1.893-4.538l5.507-3.008z" fill="url(#paint3_linear_3869_2050)"/><path d="M386.854 96.291c-5.961-1.732-9.942-2.259-16.95-.588-7.007 1.67-9.752 3.018-9.752 3.018l-14.112 24.327c-.867 1.526.213 3.392 1.937 3.484 3.454.105 8.509-.333 11.412-3.166 4.649-4.626 27.465-27.075 27.465-27.075z" fill="url(#paint4_linear_3869_2050)"/><path opacity=".4" d="M417.118 87.527l-.071 1.333c-.868 8.916-5.634 15.738-11.044 15.449-3.998-.213-7.317-4.32-8.658-10.131a27.476 27.476 0 01-.626-7.424l.016-.313c.31-5.802 2.227-10.81 4.982-13.808 1.755-1.95 3.938-3.092 6.212-2.97l.392.02c4.843.573 8.52 6.822 8.964 14.708-.132 1.015-.108 2.038-.167 3.136z" fill="#EDE5DA"/><path fill-rule="evenodd" clip-rule="evenodd" d="M384.985 85.818c.435-2.08 1.631-4.555 3.844-5.56l.387.851c-1.814.824-2.907 2.94-3.317 4.904l-.001.007c-.794 3.518-.01 7.224 1.589 10.468.272.54.613 1.071.948 1.444v.096c-.334.127-.752.085-1.054.085a8.792 8.792 0 01-.73-1.207l-.002-.003c-1.663-3.375-2.516-7.302-1.664-11.085z" fill="#B7B4AF"/><path d="M353.892 80.462l-.522 2.41 5.993 4.093 11.574-.327c1.024-.024 1.854-.845 1.904-1.785l.062-1.176 1.484-2.752-1.815-1.354a4.155 4.155 0 00-2.314-.831l-9.986.018a2.507 2.507 0 00-1.11.177l-5.27 1.527z" fill="#777"/><path fill-rule="evenodd" clip-rule="evenodd" d="M368.669 80.773l4.463 2.496-.456.816-4.241-2.371-9.188.204-5.766 1.408-.221-.909 5.864-1.431 9.545-.213z" fill="#B7B4AF"/><path fill-rule="evenodd" clip-rule="evenodd" d="M369.15 78.789l-.573 4.93.538 2.869-.919.172-.564-3.008.589-5.071.929.108z" fill="#B7B4AF"/><path d="M359.363 86.965l-.509-3.723.212-2.504.257-1.873" fill="#fff"/><path fill-rule="evenodd" clip-rule="evenodd" d="M359.531 80.79l.255-1.862-.926-.127-.259 1.885-.217 2.568.516 3.774.926-.127-.502-3.67.207-2.442z" fill="#B7B4AF"/><path opacity=".6" d="M408.168 69.597l-.702 13.17-1.176-.063-.851 1.213-3.801-11.288c1.755-1.95 3.939-3.092 6.212-2.97.083-.075.24-.067.318-.063zM417.118 87.527l-.071 1.333-5.025-.11.062-1.177-1.212-.85 6.335-2.336c-.054 1.02-.03 2.042-.089 3.14zM405.551 89.19l-8.284 4.983c-.506-2.306-.769-4.758-.627-7.423l7.84.418-.063 1.176 1.134.846z" fill="#fff"/><path opacity=".3" d="M358.941 100.779c2.369-.188 5.134-.434 5.134-.434l-11.988 19.881-4.481-.003M376.974 75.167l4.39.234 12.924-19.752-3.214-.172" fill="#D0C6B8"/><path d="M386.743 89.525a.55.55 0 10.058-1.098.55.55 0 00-.058 1.097zM387.121 92.768a.55.55 0 10.059-1.097.55.55 0 00-.059 1.097zM388.212 95.892a.549.549 0 10.06-1.096.549.549 0 00-.06 1.096zM387.237 86.17a.55.55 0 10.059-1.097.55.55 0 00-.059 1.097zM388.889 83.192a.55.55 0 10.058-1.098.55.55 0 00-.058 1.098z" fill="#777"/><path opacity=".6" d="M405.846 104.301l.31-13.192 1.176.063.851-1.213 4.119 11.227c-1.759 2.029-3.864 3.174-6.064 3.136l-.392-.021z" fill="#fff"/><path d="M406.29 82.704c-.826.743-3.912 5.531-.16 10.37 4.033-1.908 6.86-3.33 6.746-5.616-.031-2.36-6.586-4.754-6.586-4.754z" fill="#777"/><g filter="url(#filter0_d_3869_2050)"><path d="M229.895 151.928H56.614c-3.1 0-5.636-2.536-5.636-5.635V29.739c0-3.1 2.536-5.635 5.636-5.635h173.281c3.099 0 5.635 2.536 5.635 5.635v116.46c.094 3.193-2.442 5.729-5.635 5.729z" fill="url(#paint5_linear_3869_2050)"/></g><g filter="url(#filter1_d_3869_2050)"><path d="M60.358 72.293c0 5.204-4.208 9.413-9.458 9.413H19.313C13.023 81.842 8 76.82 8 70.755c0-6.11 5.068-11.178 11.45-10.861 5.475-17.106 30.545-14.707 32.672 2.987 4.706.588 8.236 4.57 8.236 9.412z" fill="url(#paint6_linear_3869_2050)"/></g><path d="M50.9 81.706c5.204 0 9.458-4.208 9.458-9.413 0-5.204-4.254-9.412-9.458-9.412-5.204 0-9.458 4.208-9.458 9.412 0 5.205 4.254 9.413 9.458 9.413z" fill="url(#paint7_linear_3869_2050)"/><path d="M35.469 81.797c9.277 0 16.834-7.512 16.834-16.79 0-9.276-7.557-16.788-16.834-16.788S18.635 55.73 18.635 65.008c0 9.277 7.512 16.789 16.834 16.789z" fill="url(#paint8_linear_3869_2050)"/><g filter="url(#filter2_d_3869_2050)"><path d="M333.368 121.066c0 4.646-3.757 8.404-8.444 8.404H296.721c-5.616.121-10.101-4.364-10.101-9.778 0-5.455 4.525-9.98 10.223-9.697 4.889-15.273 27.273-13.132 29.172 2.666 4.202.526 7.353 4.081 7.353 8.405z" fill="url(#paint9_linear_3869_2050)"/></g><path d="M324.924 129.47c4.646 0 8.444-3.758 8.444-8.404 0-4.647-3.798-8.404-8.444-8.404-4.647 0-8.445 3.757-8.445 8.404 0 4.646 3.798 8.404 8.445 8.404z" fill="url(#paint10_linear_3869_2050)"/><path d="M311.146 129.551c8.283 0 15.03-6.707 15.03-14.99 0-8.283-6.747-14.99-15.03-14.99s-15.031 6.707-15.031 14.99c0 8.283 6.707 14.99 15.031 14.99z" fill="url(#paint11_linear_3869_2050)"/><g filter="url(#filter3_d_3869_2050)"><path d="M393.079 145.914c0 3.333-2.695 6.028-6.057 6.028h-20.229c-4.029.087-7.246-3.13-7.246-7.014a6.959 6.959 0 017.333-6.955c3.506-10.955 19.562-9.419 20.924 1.913 3.014.376 5.275 2.927 5.275 6.028z" fill="url(#paint12_linear_3869_2050)"/></g><path d="M387.021 151.942c3.333 0 6.058-2.695 6.058-6.028 0-3.333-2.725-6.028-6.058-6.028-3.332 0-6.057 2.695-6.057 6.028 0 3.333 2.725 6.028 6.057 6.028z" fill="url(#paint13_linear_3869_2050)"/><path d="M377.139 152c5.941 0 10.781-4.811 10.781-10.752 0-5.941-4.84-10.752-10.781-10.752-5.941 0-10.781 4.811-10.781 10.752 0 5.941 4.811 10.752 10.781 10.752z" fill="url(#paint14_linear_3869_2050)"/><path d="M235.56 39.61H51.008V29.9c0-3.295 2.536-5.992 5.636-5.992h173.28c3.1 0 5.636 2.697 5.636 5.992v9.708z" fill="url(#paint15_linear_3869_2050)"/><path d="M59.905 34.631a2.724 2.724 0 100-5.447 2.724 2.724 0 000 5.447z" fill="#F4F1EA"/><path d="M68.17 34.631a2.724 2.724 0 100-5.447 2.724 2.724 0 000 5.447z" fill="#BAB7B4"/><path d="M76.435 34.631a2.724 2.724 0 100-5.447 2.724 2.724 0 000 5.447z" fill="#949390"/><defs><linearGradient id="paint0_linear_3869_2050" x1="313.948" y1="74.903" x2="320.14" y2="76.646" gradientUnits="userSpaceOnUse"><stop stop-color="#EFE5D8"/><stop offset="1" stop-color="#D0BDA6"/></linearGradient><linearGradient id="paint1_linear_3869_2050" x1="384.869" y1="56.859" x2="404.471" y2="62.341" gradientUnits="userSpaceOnUse"><stop stop-color="#EFE5D8"/><stop offset="1" stop-color="#D0BDA6"/></linearGradient><linearGradient id="paint2_linear_3869_2050" x1="338.99" y1="82.144" x2="366.144" y2="118.418" gradientUnits="userSpaceOnUse"><stop stop-color="#F0E8DD"/><stop offset="1" stop-color="#D8CABA"/></linearGradient><linearGradient id="paint3_linear_3869_2050" x1="305.241" y1="86.32" x2="314.028" y2="89.389" gradientUnits="userSpaceOnUse"><stop stop-color="#F6F0E6"/><stop offset="1" stop-color="#E2D7C9"/></linearGradient><linearGradient id="paint4_linear_3869_2050" x1="360.996" y1="101.004" x2="384.379" y2="108.672" gradientUnits="userSpaceOnUse"><stop stop-color="#F6F0E6"/><stop offset="1" stop-color="#E2D7C9"/></linearGradient><linearGradient id="paint5_linear_3869_2050" x1="143.195" y1="21.147" x2="143.195" y2="153.306" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint6_linear_3869_2050" x1="34.179" y1="62.082" x2="34.162" y2="82.069" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#F0E8DD"/></linearGradient><linearGradient id="paint7_linear_3869_2050" x1="40.435" y1="60.854" x2="48.82" y2="70.024" gradientUnits="userSpaceOnUse"><stop stop-color="#D8CABA"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint8_linear_3869_2050" x1="17.058" y1="57.304" x2="30.273" y2="62.475" gradientUnits="userSpaceOnUse"><stop stop-color="#F0E8DD"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint9_linear_3869_2050" x1="309.994" y1="111.948" x2="309.979" y2="129.794" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#F0E8DD"/></linearGradient><linearGradient id="paint10_linear_3869_2050" x1="315.58" y1="110.852" x2="323.066" y2="119.04" gradientUnits="userSpaceOnUse"><stop stop-color="#D8CABA"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint11_linear_3869_2050" x1="294.708" y1="107.682" x2="306.506" y2="112.299" gradientUnits="userSpaceOnUse"><stop stop-color="#F0E8DD"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint12_linear_3869_2050" x1="376.313" y1="139.374" x2="376.302" y2="152.175" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#F0E8DD"/></linearGradient><linearGradient id="paint13_linear_3869_2050" x1="380.32" y1="138.588" x2="385.689" y2="144.461" gradientUnits="userSpaceOnUse"><stop stop-color="#D8CABA"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint14_linear_3869_2050" x1="365.348" y1="136.314" x2="373.811" y2="139.626" gradientUnits="userSpaceOnUse"><stop stop-color="#F0E8DD"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint15_linear_3869_2050" x1="228.239" y1="23.91" x2="56.921" y2="30.921" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><filter id="filter0_d_3869_2050" x="44.978" y="20.104" width="196.554" height="139.824" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="2"/><feGaussianBlur stdDeviation="3"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2050"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2050" result="shape"/></filter><filter id="filter1_d_3869_2050" x="0" y="43.26" width="68.358" height="49.449" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="3"/><feGaussianBlur stdDeviation="4"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2050"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2050" result="shape"/></filter><filter id="filter2_d_3869_2050" x="278.62" y="94.607" width="62.748" height="45.865" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="3"/><feGaussianBlur stdDeviation="4"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2050"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2050" result="shape"/></filter><filter id="filter3_d_3869_2050" x="353.547" y="125.522" width="45.532" height="33.422" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="1"/><feGaussianBlur stdDeviation="3"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2050"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2050" result="shape"/></filter></defs></svg>`));
