+++ illustrations/src/illustrations/chunk-2.tsx
@@ -1,0 +1,5 @@
+import { createFluentIllustration } from "../utils/createFluentIllustration";
+export const UserProfileDark = (/*#__PURE__*/createFluentIllustration('UserProfileDark', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M119.711 134.141c30.037 0 54.389-24.298 54.389-54.376S149.641 25.39 119.711 25.39c-30.038 0-54.39 24.298-54.39 54.376s24.352 54.376 54.39 54.376z" fill="#242424"/><path d="M180 23.997a3 3 0 01-3.004 2.997 3 3 0 01-3.003-2.997A3 3 0 01176.996 21 3 3 0 01180 23.997zM174.958 39.732a4.393 4.393 0 01-4.398 4.389 4.394 4.394 0 01-4.399-4.389 4.394 4.394 0 014.399-4.389 4.393 4.393 0 014.398 4.389zM68.217 44.014a3 3 0 003.004-2.998 3 3 0 00-3.004-2.997 3 3 0 00-3.003 2.998 3 3 0 003.003 2.997zM61.157 103.237a5.572 5.572 0 01-5.579 5.566A5.572 5.572 0 0150 103.237a5.572 5.572 0 015.578-5.566 5.572 5.572 0 015.579 5.566z" fill="#3A3A3A"/><path d="M163.117 115.245H76.303c-3.991 0-7.234-3.236-7.234-7.218V52.519c0-3.983 3.243-7.219 7.234-7.219h86.814c3.992 0 7.235 3.236 7.235 7.219v55.757c-.25 3.733-3.493 6.969-7.235 6.969z" fill="url(#paint0_linear_4784_278)"/><path d="M99.905 47.858H84.871v5.516h15.034v-5.516z" fill="url(#paint1_linear_4784_278)"/><path d="M153.795 47.858h-15.033v5.516h15.033v-5.516z" fill="url(#paint2_linear_4784_278)"/><path d="M87.03 35.343h10.727V50.4H87.03V35.343z" fill="url(#paint3_linear_4784_278)"/><path d="M140.92 35.343h10.727V50.4H140.92V35.343z" fill="url(#paint4_linear_4784_278)"/><path fill-rule="evenodd" clip-rule="evenodd" d="M100.501 64.815c-5.731 0-10.377 4.636-10.377 10.355 0 3.161 1.42 5.992 3.658 7.891-5.1 2.51-8.648 7.764-8.648 13.768 0 1.182.96 2.14 2.145 2.14a2.143 2.143 0 002.146-2.14c0-6.036 5.027-11.052 11.076-11.052 6.05 0 11.076 5.016 11.076 11.052 0 1.182.961 2.14 2.146 2.14a2.143 2.143 0 002.145-2.14c0-6.004-3.548-11.258-8.647-13.767a10.324 10.324 0 003.658-7.892c0-5.719-4.646-10.355-10.378-10.355zM94.415 75.17c0-3.354 2.725-6.073 6.086-6.073 3.362 0 6.087 2.719 6.087 6.073 0 3.354-2.725 6.073-6.087 6.073-3.361 0-6.086-2.719-6.086-6.073z" fill="#CBCBCB"/><path d="M125.298 73.676c0-1.182.96-2.14 2.145-2.14h26.693c1.185 0 2.146.958 2.146 2.14 0 1.182-.961 2.14-2.146 2.14h-26.693a2.143 2.143 0 01-2.145-2.14zM127.443 84.977a2.143 2.143 0 00-2.145 2.14c0 1.183.96 2.141 2.145 2.141h10.478a2.143 2.143 0 002.145-2.14c0-1.183-.96-2.141-2.145-2.141h-10.478z" fill="#CBCBCB"/><defs><linearGradient id="paint0_linear_4784_278" x1="119.677" y1="43.682" x2="119.677" y2="115.999" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint1_linear_4784_278" x1="146.279" y1="47.858" x2="146.279" y2="53.374" gradientUnits="userSpaceOnUse"><stop stop-color="#363636"/><stop offset="1" stop-color="#1D1D1D"/></linearGradient><linearGradient id="paint2_linear_4784_278" x1="146.279" y1="47.858" x2="146.279" y2="53.374" gradientUnits="userSpaceOnUse"><stop stop-color="#363636"/><stop offset="1" stop-color="#1D1D1D"/></linearGradient><linearGradient id="paint3_linear_4784_278" x1="146.28" y1="34.995" x2="146.28" y2="50.561" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint4_linear_4784_278" x1="146.28" y1="34.995" x2="146.28" y2="50.561" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient></defs></svg>`));
+export const UserProfileLight = (/*#__PURE__*/createFluentIllustration('UserProfileLight', `<svg viewBox="0 0 230 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M119.711 134.391c30.037 0 54.389-24.351 54.389-54.496 0-30.145-24.459-54.497-54.389-54.497-30.038 0-54.39 24.352-54.39 54.497s24.352 54.496 54.39 54.496z" fill="#EBEBEB"/><path d="M170.56 44.172a4.398 4.398 0 100-8.797 4.398 4.398 0 000 8.797zM176.997 27.008a3.004 3.004 0 10-.001-6.008 3.004 3.004 0 00.001 6.008zM68.218 44.064a3.004 3.004 0 100-6.007 3.004 3.004 0 000 6.007zM55.579 108.997a5.579 5.579 0 10-.001-11.158 5.579 5.579 0 000 11.158z" fill="#E8DFD3"/><g filter="url(#filter0_d_4784_267)"><path d="M163.117 115.453H76.304a7.24 7.24 0 01-7.235-7.234v-55.63a7.24 7.24 0 017.235-7.235h86.813a7.24 7.24 0 017.235 7.234v55.88c-.25 3.742-3.493 6.985-7.235 6.985z" fill="url(#paint0_linear_4784_267)"/></g><path d="M84.872 48.99c0-.593.48-1.073 1.073-1.073h12.888c.593 0 1.073.48 1.073 1.073v3.382c0 .593-.48 1.073-1.073 1.073H85.945c-.592 0-1.073-.48-1.073-1.073V48.99zM138.762 48.99c0-.593.48-1.073 1.072-1.073h12.889c.592 0 1.073.48 1.073 1.073v3.382c0 .593-.481 1.073-1.073 1.073h-12.889c-.592 0-1.072-.48-1.072-1.073V48.99z" fill="#D4CDC0"/><g filter="url(#filter1_d_4784_267)"><path d="M96.685 35.375h-8.581c-.593 0-1.073.48-1.073 1.073V49.39c0 .592.48 1.073 1.073 1.073h8.581c.593 0 1.073-.48 1.073-1.073V36.448c0-.593-.48-1.073-1.073-1.073z" fill="url(#paint1_linear_4784_267)"/><path d="M150.575 35.375h-8.582c-.592 0-1.072.48-1.072 1.073V49.39c0 .592.48 1.073 1.072 1.073h8.582c.592 0 1.073-.48 1.073-1.073V36.448c0-.593-.481-1.073-1.073-1.073z" fill="url(#paint2_linear_4784_267)"/></g><path fill-rule="evenodd" clip-rule="evenodd" d="M100.501 64.912c-5.731 0-10.378 4.646-10.378 10.377 0 3.169 1.42 6.006 3.659 7.91-5.099 2.514-8.648 7.78-8.648 13.797a2.146 2.146 0 004.291 0c0-6.05 5.027-11.076 11.077-11.076 6.049 0 11.076 5.027 11.076 11.076a2.145 2.145 0 104.291 0c0-6.017-3.549-11.283-8.648-13.798a10.356 10.356 0 003.658-7.909c0-5.731-4.646-10.377-10.378-10.377zm-6.087 10.377a6.087 6.087 0 1112.174 0 6.087 6.087 0 01-12.174 0z" fill="#BCB9B3"/><path d="M125.299 73.793c0-1.185.96-2.146 2.145-2.146h26.693a2.145 2.145 0 010 4.291h-26.693a2.145 2.145 0 01-2.145-2.145zM127.444 85.118a2.145 2.145 0 000 4.291h10.478a2.145 2.145 0 000-4.29h-10.478z" fill="#BCB9B3"/><defs><linearGradient id="paint0_linear_4784_267" x1="119.678" y1="43.732" x2="119.678" y2="116.209" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint1_linear_4784_267" x1="146.281" y1="35.026" x2="146.281" y2="50.626" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint2_linear_4784_267" x1="146.281" y1="35.026" x2="146.281" y2="50.626" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><filter id="filter0_d_4784_267" x="57.069" y="41.354" width="125.283" height="94.1" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="8"/><feGaussianBlur stdDeviation="6"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.17 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4784_267"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4784_267" result="shape"/></filter><filter id="filter1_d_4784_267" x="65.031" y="24.375" width="108.617" height="59.089" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_4784_267"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_4784_267" result="shape"/></filter></defs></svg>`));
+export const WelcomePlaneDark = (/*#__PURE__*/createFluentIllustration('WelcomePlaneDark', `<svg viewBox="0 0 437 176" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M311.498 1.684c1.374-.383 2.663-.48 4.037-.48 1.288 0 2.663.097 3.951.288 1.289.192 2.663.576 3.951 1.248 1.289.671 2.577 1.63 3.35 3.262l-1.46-.384 1.546-1.343c.601-.384 1.117-.864 1.632-1.248 1.117-.767 2.319-1.439 3.608-1.918 1.288-.48 2.577-.864 3.865-1.056 1.374-.096 2.663-.096 4.037.288a.44.44 0 01.258.576c-.086.288-.258.384-.515.288h-.086c-1.117-.288-2.405-.192-3.608 0-1.203.191-2.405.671-3.522 1.247-1.116.575-2.147 1.247-3.178 2.015-.515.383-1.031.863-1.46 1.247l-1.374 1.343c-.43.384-.945.288-1.289-.096l-.086.096-.086-.192c-.343-1.055-1.288-1.919-2.319-2.494a13.183 13.183 0 00-3.522-1.344 24.053 24.053 0 00-3.779-.575c-1.288-.096-2.577-.096-3.693.192h-.086c-.258.096-.43-.096-.516-.384 0-.288.086-.576.344-.576zM280.803 27.464c1.031-.269 1.997-.336 3.028-.336.966 0 1.997.067 2.963.201.967.135 1.997.403 2.964.873.966.47 1.932 1.142 2.512 2.284l-1.095-.269 1.16-.94c.451-.269.837-.605 1.224-.873a12.827 12.827 0 012.705-1.343c.967-.336 1.933-.605 2.899-.739 1.031-.067 1.997-.067 3.028.201.193.068.258.27.193.403-.064.202-.193.269-.386.202h-.065c-.837-.202-1.803-.134-2.705 0-.902.134-1.804.47-2.642.873-.837.403-1.61.873-2.383 1.41-.387.269-.773.605-1.095.873l-1.031.94c-.322.27-.709.202-.966-.067l-.065-.067-.064-.134c-.258-.739-.967-1.343-1.74-1.746a10.333 10.333 0 00-2.641-.94 19.257 19.257 0 00-2.835-.403c-.966-.068-1.932-.068-2.77.134h-.064c-.193.067-.322-.067-.387-.269 0-.067.129-.268.258-.268zM98.698 1.684c1.374-.383 2.663-.48 4.037-.48 1.289 0 2.663.097 3.951.288 1.289.192 2.663.576 3.952 1.248 1.288.671 2.576 1.63 3.349 3.262l-1.46-.384 1.546-1.343c.602-.384 1.117-.864 1.632-1.248 1.117-.767 2.32-1.439 3.608-1.918 1.288-.48 2.577-.864 3.865-1.056 1.375-.096 2.663-.096 4.037.288.258.096.344.384.258.576-.086.288-.258.384-.515.288h-.086c-1.117-.288-2.405-.192-3.608 0-1.202.191-2.405.671-3.522 1.247-1.116.575-2.147 1.247-3.178 2.015-.515.383-1.03.863-1.46 1.247l-1.374 1.343c-.43.384-.945.288-1.289-.096l-.086-.096-.085-.192c-.344-1.055-1.289-1.919-2.32-2.494a13.175 13.175 0 00-3.521-1.343 24.084 24.084 0 00-3.78-.576c-1.288-.096-2.577-.096-3.693.192h-.086c-.258.096-.43-.096-.516-.384 0-.096.172-.384.344-.384zM320.243 24.15c1.375-.346 2.663-.432 4.038-.432 1.288 0 2.662.086 3.951.259 1.288.173 2.663.518 3.951 1.122 1.288.605 2.577 1.468 3.35 2.936l-1.46-.345 1.546-1.209c.601-.345 1.116-.777 1.632-1.123 1.116-.69 2.319-1.295 3.607-1.726 1.289-.432 2.577-.777 3.866-.95 1.374-.087 2.662-.087 4.037.259.258.086.343.345.257.518-.085.259-.257.345-.515.259h-.086c-1.116-.26-2.405-.173-3.607 0-1.203.173-2.405.604-3.522 1.122a23.204 23.204 0 00-3.178 1.814c-.516.345-1.031.777-1.46 1.122l-1.375 1.21c-.429.345-.945.258-1.288-.087l-.086-.087-.086-.172c-.344-.95-1.288-1.727-2.319-2.245a14.107 14.107 0 00-3.522-1.21 26.578 26.578 0 00-3.779-.517c-1.289-.087-2.577-.087-3.694.172h-.086c-.257.087-.429-.086-.515-.345 0-.086.086-.345.343-.345z" fill="#5C5C5C"/><path d="M239.176 42.5l58.277 45.434h12.931M238.243 131.772l59.21-43.838" stroke="#8C8C8C" stroke-width="5" stroke-miterlimit="10"/><g filter="url(#filter0_d_3869_2076)"><path d="M315.711 79.531l-.316-1.431s1.216-2.215 1.826-3.36c.689-1.143 1.327-1.344 4.242-1.582 2.919-.316 5.138-.747 4.149 1.636-1.068 2.38-4.053 8.35-4.053 8.35l-5.848-3.613z" fill="#484848"/></g><g filter="url(#filter1_d_3869_2076)"><path d="M377.15 82.104l20.344-28.541a3.247 3.247 0 011.792-1.24c1.35-.321 3.656-.906 5.871-1.259 3.323-.53 7.115-1.035 5.594 2.5-1.526 3.611-11.283 27.767-11.283 27.767l-22.318.773z" fill="#484848"/></g><g filter="url(#filter2_d_3869_2076)"><path d="M314.482 76.086c.235.013 10.566 9.287 10.566 9.287l34.783-5.058 19.658.578s16.579-.53 21.664-.258c5.084.272 11.847 1.419 12.438 5.065.59 3.647 2.035 4.51-.157 5.886-2.112 1.38-10.502 7.22-22.07 6.444-11.569-.774-25.267 1.638-25.267 1.638s-.077-1.498-3.832-1.698c-3.754-.2-50.143-7-50.143-7s-7.674-.252-6.801-7.829c.952-7.572 3.525-7.277 5.172-7.268 1.729-.065 3.989.213 3.989.213z" fill="url(#paint0_linear_3869_2076)"/></g><g filter="url(#filter3_d_3869_2076)"><path d="M312.97 83.864l7.353.393s-4.339 6.37-5.905 7.779c-1.565 1.41-4.593 3.763-7.548 3.29-2.956-.472-2.164-2.08-1.282-3.919.882-1.839 1.888-4.536 1.888-4.536l5.494-3.007z" fill="#6C6C6C"/></g><g filter="url(#filter4_d_3869_2076)"><path d="M391.929 96.253c-5.948-1.732-9.921-2.259-16.912-.589-6.992 1.67-9.73 3.017-9.73 3.017L351.207 123c-.866 1.525.212 3.39 1.933 3.482 3.445.106 8.489-.332 11.385-3.164 4.639-4.625 27.404-27.064 27.404-27.064z" fill="#6C6C6C"/></g><path opacity=".4" d="M422.124 87.492l-.071 1.332c-.866 8.913-5.622 15.732-11.019 15.444-3.989-.213-7.3-4.32-8.638-10.128a27.495 27.495 0 01-.625-7.42l.016-.314c.309-5.799 2.222-10.805 4.97-13.802 1.751-1.95 3.93-3.09 6.198-2.97l.391.021c4.833.573 8.501 6.82 8.945 14.702-.133 1.015-.109 2.038-.167 3.135z" fill="#6C6C6C"/><path fill-rule="evenodd" clip-rule="evenodd" d="M390.064 85.784c.434-2.079 1.627-4.553 3.836-5.558l.385.85c-1.809.824-2.9 2.94-3.309 4.903l-.001.007c-.792 3.516-.01 7.22 1.585 10.463.272.54.612 1.071.946 1.444v.096c-.333.126-.751.085-1.051.085a8.684 8.684 0 01-.729-1.206l-.002-.004c-1.659-3.373-2.51-7.298-1.66-11.08z" fill="#8A8A8A"/><path d="M359.041 80.43l-.52 2.408 5.979 4.092 11.548-.327c1.021-.024 1.849-.844 1.899-1.785l.063-1.175 1.48-2.75-1.811-1.354a4.141 4.141 0 00-2.309-.83l-9.963.017a2.502 2.502 0 00-1.108.177l-5.258 1.527z" fill="#3A3A3A"/><path fill-rule="evenodd" clip-rule="evenodd" d="M373.784 80.74l4.453 2.495-.456.816-4.231-2.37-9.167.204-5.752 1.407-.222-.908 5.851-1.43 9.524-.213z" fill="#6C6C6C"/><path fill-rule="evenodd" clip-rule="evenodd" d="M374.265 78.757l-.572 4.929.537 2.867-.917.172-.563-3.006.588-5.07.927.108z" fill="#6C6C6C"/><path d="M364.499 86.93l-.508-3.721.212-2.504.256-1.872" fill="#fff"/><path fill-rule="evenodd" clip-rule="evenodd" d="M364.668 80.757l.255-1.86-.925-.128-.258 1.885-.217 2.567.515 3.772.925-.127-.502-3.67.207-2.439z" fill="#6C6C6C"/><path opacity=".6" d="M413.194 69.569l-.701 13.165-1.173-.063-.849 1.212-3.792-11.283c1.751-1.95 3.929-3.09 6.198-2.97.082-.074.238-.065.317-.061zM422.124 87.492l-.071 1.332-5.014-.11.062-1.176-1.21-.85 6.321-2.335c-.054 1.019-.03 2.042-.088 3.139zM410.583 89.155l-8.266 4.98a27.495 27.495 0 01-.625-7.42l7.822.418-.063 1.175 1.132.847z" fill="#777"/><path opacity=".3" d="M364.079 100.739c2.363-.188 5.122-.434 5.122-.434l-11.96 19.873-4.471-.003" fill="#D1D1D1"/><path opacity=".4" d="M382.07 75.137l4.381.234 12.895-19.744-3.207-.172" fill="#A3A3A3"/><path d="M391.818 89.489a.548.548 0 10.057-1.095.548.548 0 00-.057 1.095zM392.195 92.73a.548.548 0 10.057-1.094.548.548 0 00-.057 1.095zM393.284 95.854a.548.548 0 10.057-1.095.548.548 0 00-.057 1.095zM392.31 86.136a.549.549 0 10.057-1.097.549.549 0 00-.057 1.097zM393.958 83.159a.548.548 0 10.059-1.095.548.548 0 00-.059 1.095z" fill="#8A8A8A"/><path opacity=".6" d="M410.878 104.259l.31-13.186 1.173.063.849-1.212 4.109 11.221c-1.755 2.028-3.855 3.174-6.049 3.135l-.392-.021z" fill="#777"/><path d="M411.321 82.671c-.824.742-3.902 5.529-.159 10.365 4.023-1.907 6.844-3.328 6.731-5.613-.032-2.36-6.572-4.752-6.572-4.752z" fill="#8A8A8A"/><path d="M235.325 151.867H62.436c-3.092 0-5.622-2.535-5.622-5.633V29.727c0-3.098 2.53-5.633 5.622-5.633h172.889c3.092 0 5.623 2.535 5.623 5.633V146.14c.093 3.192-2.437 5.727-5.623 5.727z" fill="url(#paint1_linear_3869_2076)"/><g filter="url(#filter5_d_3869_2076)"><path d="M65.905 72.57c0 5.269-4.253 9.529-9.557 9.529H24.43C18.076 82.236 13 77.151 13 71.013c0-6.185 5.121-11.316 11.569-10.995C30.1 42.702 55.433 45.13 57.583 63.042c4.755.595 8.322 4.626 8.322 9.528z" fill="#424242"/></g><path d="M40.755 82.19c9.374 0 17.01-7.604 17.01-16.995 0-9.392-7.636-16.996-17.01-16.996-9.373 0-17.01 7.604-17.01 16.996 0 9.39 7.591 16.996 17.01 16.996z" fill="url(#paint2_linear_3869_2076)"/><path d="M56.348 82.099c5.259 0 9.557-4.26 9.557-9.529 0-5.268-4.298-9.528-9.557-9.528-5.258 0-9.556 4.26-9.556 9.528 0 5.269 4.298 9.529 9.556 9.529z" fill="url(#paint3_linear_3869_2076)"/><g filter="url(#filter6_d_3869_2076)"><path d="M337.629 121.374c0 4.644-3.749 8.4-8.425 8.4h-28.139c-5.603.122-10.078-4.362-10.078-9.774 0-5.452 4.515-9.976 10.199-9.693 4.878-15.267 27.212-13.126 29.106 2.666 4.193.525 7.337 4.079 7.337 8.401z" fill="#424242"/></g><path d="M315.457 129.855c8.264 0 14.997-6.704 14.997-14.984 0-8.28-6.733-14.984-14.997-14.984s-14.997 6.704-14.997 14.984c0 8.28 6.692 14.984 14.997 14.984z" fill="url(#paint4_linear_3869_2076)"/><path d="M329.205 129.774c4.636 0 8.425-3.756 8.425-8.401 0-4.644-3.789-8.4-8.425-8.4-4.636 0-8.426 3.756-8.426 8.4 0 4.645 3.79 8.401 8.426 8.401z" fill="url(#paint5_linear_3869_2076)"/><g filter="url(#filter7_d_3869_2076)"><path d="M397.333 145.893c0 3.345-2.7 6.049-6.067 6.049H371.006c-4.034.087-7.256-3.141-7.256-7.037 0-3.926 3.251-7.183 7.343-6.979 3.513-10.993 19.593-9.451 20.957 1.919 3.019.378 5.283 2.937 5.283 6.048z" fill="#424242"/></g><path d="M381.369 152c5.951 0 10.798-4.827 10.798-10.789 0-5.961-4.847-10.788-10.798-10.788-5.95 0-10.797 4.827-10.797 10.788A10.78 10.78 0 00381.369 152z" fill="url(#paint6_linear_3869_2076)"/><path d="M391.267 151.942c3.338 0 6.066-2.705 6.066-6.049 0-3.344-2.728-6.048-6.066-6.048-3.338 0-6.066 2.704-6.066 6.048s2.728 6.049 6.066 6.049z" fill="url(#paint7_linear_3869_2076)"/><path d="M240.977 39.594H56.844V29.89c0-3.295 2.53-5.99 5.622-5.99h172.889c3.092 0 5.622 2.695 5.622 5.99v9.704z" fill="#3A3A3A"/><path d="M65.72 34.617a2.72 2.72 0 002.717-2.722 2.72 2.72 0 00-2.717-2.723 2.72 2.72 0 00-2.718 2.723 2.72 2.72 0 002.718 2.723z" fill="#CCC"/><path d="M73.966 34.617a2.72 2.72 0 002.718-2.722 2.72 2.72 0 00-2.718-2.723 2.72 2.72 0 00-2.717 2.723 2.72 2.72 0 002.717 2.723z" fill="#505050"/><path d="M82.213 34.617a2.72 2.72 0 002.717-2.722 2.72 2.72 0 00-2.717-2.723 2.72 2.72 0 00-2.718 2.723 2.72 2.72 0 002.718 2.723z" fill="#484848"/><defs><linearGradient id="paint0_linear_3869_2076" x1="360.375" y1="77.974" x2="359.225" y2="99.526" gradientUnits="userSpaceOnUse"><stop stop-color="#6C6C6C"/><stop offset=".996" stop-color="#3B3B3B"/></linearGradient><linearGradient id="paint1_linear_3869_2076" x1="148.822" y1="21.139" x2="148.822" y2="153.244" gradientUnits="userSpaceOnUse"><stop stop-color="#5C5C5C"/><stop offset=".996" stop-color="#4D4D4D"/></linearGradient><linearGradient id="paint2_linear_3869_2076" x1="27.891" y1="55.259" x2="33.595" y2="59.66" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint3_linear_3869_2076" x1="49.121" y1="67" x2="52.32" y2="69.474" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint4_linear_3869_2076" x1="304.115" y1="106.111" x2="309.144" y2="109.991" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint5_linear_3869_2076" x1="322.833" y1="116.462" x2="325.654" y2="118.643" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint6_linear_3869_2076" x1="373.203" y1="134.905" x2="376.824" y2="137.698" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><linearGradient id="paint7_linear_3869_2076" x1="386.679" y1="142.357" x2="388.71" y2="143.928" gradientUnits="userSpaceOnUse"><stop stop-color="#686868"/><stop offset=".994" stop-color="#424242"/></linearGradient><filter id="filter0_d_3869_2076" x="300.396" y="62.909" width="30.455" height="30.235" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dx="-5"/><feGaussianBlur stdDeviation="5"/><feColorMatrix values="0 0 0 0 0.397708 0 0 0 0 0.47749 0 0 0 0 0.575 0 0 0 0.2 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter1_d_3869_2076" x="355.15" y="39.689" width="77.947" height="75.415" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter2_d_3869_2076" x="283.252" y="64.863" width="153.354" height="67.804" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter3_d_3869_2076" x="289.784" y="73.864" width="35.539" height="31.524" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dx="-5"/><feGaussianBlur stdDeviation="5"/><feColorMatrix values="0 0 0 0 0.397708 0 0 0 0 0.47749 0 0 0 0 0.575 0 0 0 0.2 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter4_d_3869_2076" x="328.901" y="83.649" width="85.027" height="75.846" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="11"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter5_d_3869_2076" x="0" y="46.24" width="78.905" height="59.861" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="6.5"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter6_d_3869_2076" x="277.987" y="97.923" width="72.643" height="55.853" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="6.5"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter><filter id="filter7_d_3869_2076" x="350.75" y="128.449" width="59.583" height="47.494" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="11"/><feGaussianBlur stdDeviation="6.5"/><feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.22 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2076"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2076" result="shape"/></filter></defs></svg>`));
+export const WelcomePlaneLight = (/*#__PURE__*/createFluentIllustration('WelcomePlaneLight', `<svg viewBox="0 0 424 160" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M306.24 1.685c1.378-.384 2.669-.48 4.047-.48 1.291 0 2.668.096 3.96.288 1.291.192 2.669.576 3.96 1.248s2.583 1.631 3.357 3.263l-1.463-.384 1.55-1.344c.602-.383 1.119-.863 1.635-1.247 1.119-.768 2.325-1.44 3.616-1.92s2.583-.864 3.874-1.056c1.378-.096 2.669-.096 4.046.288a.44.44 0 01.259.576c-.086.288-.259.384-.517.288h-.086c-1.119-.288-2.411-.192-3.616 0s-2.41.672-3.529 1.248c-1.12.576-2.153 1.248-3.186 2.015-.516.384-1.033.864-1.463 1.248l-1.378 1.344c-.43.384-.947.288-1.291-.096l-.086.096-.086-.192c-.345-1.056-1.292-1.92-2.325-2.496a13.222 13.222 0 00-3.53-1.343 24.163 24.163 0 00-3.787-.576c-1.292-.096-2.583-.096-3.702.192h-.086c-.259.096-.431-.096-.517-.384 0-.288.086-.576.344-.576zM275.477 27.475c1.033-.269 2.001-.336 3.034-.336.969 0 2.002.067 2.97.201.969.135 2.002.404 2.97.874.969.47 1.937 1.142 2.519 2.284l-1.098-.269 1.162-.94c.452-.269.84-.605 1.227-.874a12.855 12.855 0 012.712-1.343c.968-.336 1.937-.605 2.905-.74 1.033-.067 2.002-.067 3.035.202.194.067.258.269.194.403-.065.202-.194.269-.388.202h-.064c-.84-.202-1.808-.134-2.712 0-.904.134-1.808.47-2.647.873-.84.403-1.615.874-2.389 1.411-.388.269-.775.605-1.098.874l-1.033.94c-.323.269-.71.202-.969-.067l-.064-.067-.065-.135c-.258-.739-.968-1.343-1.743-1.746a10.374 10.374 0 00-2.647-.941 19.344 19.344 0 00-2.841-.403c-.969-.067-1.937-.067-2.777.134h-.064c-.194.067-.323-.067-.388-.268 0-.068.129-.27.259-.27zM92.957 1.685c1.377-.384 2.669-.48 4.046-.48 1.291 0 2.669.096 3.96.288 1.291.192 2.669.576 3.96 1.248 1.292.672 2.583 1.631 3.358 3.263l-1.464-.384 1.55-1.344c.602-.383 1.119-.863 1.636-1.247 1.119-.768 2.324-1.44 3.615-1.92 1.292-.48 2.583-.864 3.874-1.056 1.378-.096 2.669-.096 4.047.288a.441.441 0 01.258.576c-.086.288-.258.384-.517.288h-.086c-1.119-.288-2.41-.192-3.615 0-1.206.192-2.411.672-3.53 1.248-1.119.576-2.152 1.248-3.186 2.015-.516.384-1.033.864-1.463 1.248l-1.378 1.344c-.43.384-.947.288-1.291-.096l-.086-.096-.086-.192c-.344-1.056-1.291-1.92-2.324-2.496a13.248 13.248 0 00-3.53-1.343 24.18 24.18 0 00-3.788-.576c-1.291-.096-2.583-.096-3.702.192h-.086c-.258.096-.43-.096-.517-.384 0-.096.172-.384.345-.384zM315.005 24.16c1.378-.346 2.669-.432 4.047-.432 1.291 0 2.669.086 3.96.259 1.291.173 2.669.518 3.96 1.123 1.291.604 2.583 1.468 3.358 2.937l-1.464-.346 1.55-1.21c.602-.345 1.119-.777 1.635-1.122 1.12-.691 2.325-1.296 3.616-1.728 1.292-.432 2.583-.777 3.874-.95 1.378-.086 2.669-.086 4.047.26.258.086.344.345.258.517-.086.26-.258.346-.517.26h-.086c-1.119-.26-2.41-.173-3.616 0-1.205.172-2.41.604-3.529 1.123a23.236 23.236 0 00-3.186 1.814c-.516.345-1.033.777-1.463 1.123l-1.378 1.209c-.43.345-.947.26-1.291-.087l-.086-.086-.086-.173c-.345-.95-1.292-1.727-2.325-2.245a14.158 14.158 0 00-3.529-1.21 26.71 26.71 0 00-3.788-.518c-1.292-.087-2.583-.087-3.702.173h-.086c-.259.086-.431-.087-.517-.346 0-.086.086-.345.344-.345z" fill="#E8DFD3"/><path d="M233.753 42.516l58.41 45.453h12.961M232.818 131.825l59.345-43.856" stroke="#E4E0DA" stroke-width="6" stroke-miterlimit="10"/><path d="M310.463 79.563l-.316-1.432s1.218-2.215 1.83-3.362c.69-1.143 1.329-1.344 4.251-1.582 2.926-.316 5.15-.747 4.158 1.637-1.07 2.38-4.062 8.353-4.062 8.353l-5.861-3.614z" fill="url(#paint0_linear_3869_2050)"/><path d="M372.043 82.137l20.39-28.553a3.25 3.25 0 011.796-1.24c1.353-.322 3.664-.906 5.884-1.26 3.331-.53 7.131-1.034 5.606 2.5-1.529 3.614-11.307 27.78-11.307 27.78l-22.369.773z" fill="url(#paint1_linear_3869_2050)"/><path d="M309.232 76.117c.236.012 10.591 9.29 10.591 9.29l34.862-5.06 19.702.579s16.617-.53 21.713-.258c5.095.271 11.874 1.419 12.466 5.067.592 3.648 2.039 4.51-.157 5.888-2.117 1.38-10.526 7.222-22.121 6.447-11.594-.775-25.324 1.638-25.324 1.638s-.077-1.498-3.84-1.699c-3.763-.2-50.257-7.002-50.257-7.002s-7.692-.253-6.816-7.832c.954-7.576 3.533-7.28 5.183-7.272 1.733-.064 3.998.213 3.998.213z" fill="url(#paint2_linear_3869_2050)"/><path d="M307.717 83.898l7.369.392s-4.349 6.372-5.918 7.783c-1.569 1.41-4.603 3.764-7.566 3.292-2.962-.473-2.169-2.081-1.285-3.921.885-1.84 1.893-4.538 1.893-4.538l5.507-3.008z" fill="url(#paint3_linear_3869_2050)"/><path d="M386.854 96.291c-5.961-1.732-9.942-2.259-16.95-.588-7.007 1.67-9.752 3.018-9.752 3.018l-14.112 24.327c-.867 1.526.213 3.392 1.937 3.484 3.454.105 8.509-.333 11.412-3.166 4.649-4.626 27.465-27.075 27.465-27.075z" fill="url(#paint4_linear_3869_2050)"/><path opacity=".4" d="M417.118 87.527l-.071 1.333c-.868 8.916-5.634 15.738-11.044 15.449-3.998-.213-7.317-4.32-8.658-10.131a27.476 27.476 0 01-.626-7.424l.016-.313c.31-5.802 2.227-10.81 4.982-13.808 1.755-1.95 3.938-3.092 6.212-2.97l.392.02c4.843.573 8.52 6.822 8.964 14.708-.132 1.015-.108 2.038-.167 3.136z" fill="#EDE5DA"/><path fill-rule="evenodd" clip-rule="evenodd" d="M384.985 85.818c.435-2.08 1.631-4.555 3.844-5.56l.387.851c-1.814.824-2.907 2.94-3.317 4.904l-.001.007c-.794 3.518-.01 7.224 1.589 10.468.272.54.613 1.071.948 1.444v.096c-.334.127-.752.085-1.054.085a8.792 8.792 0 01-.73-1.207l-.002-.003c-1.663-3.375-2.516-7.302-1.664-11.085z" fill="#B7B4AF"/><path d="M353.892 80.462l-.522 2.41 5.993 4.093 11.574-.327c1.024-.024 1.854-.845 1.904-1.785l.062-1.176 1.484-2.752-1.815-1.354a4.155 4.155 0 00-2.314-.831l-9.986.018a2.507 2.507 0 00-1.11.177l-5.27 1.527z" fill="#777"/><path fill-rule="evenodd" clip-rule="evenodd" d="M368.669 80.773l4.463 2.496-.456.816-4.241-2.371-9.188.204-5.766 1.408-.221-.909 5.864-1.431 9.545-.213z" fill="#B7B4AF"/><path fill-rule="evenodd" clip-rule="evenodd" d="M369.15 78.789l-.573 4.93.538 2.869-.919.172-.564-3.008.589-5.071.929.108z" fill="#B7B4AF"/><path d="M359.363 86.965l-.509-3.723.212-2.504.257-1.873" fill="#fff"/><path fill-rule="evenodd" clip-rule="evenodd" d="M359.531 80.79l.255-1.862-.926-.127-.259 1.885-.217 2.568.516 3.774.926-.127-.502-3.67.207-2.442z" fill="#B7B4AF"/><path opacity=".6" d="M408.168 69.597l-.702 13.17-1.176-.063-.851 1.213-3.801-11.288c1.755-1.95 3.939-3.092 6.212-2.97.083-.075.24-.067.318-.063zM417.118 87.527l-.071 1.333-5.025-.11.062-1.177-1.212-.85 6.335-2.336c-.054 1.02-.03 2.042-.089 3.14zM405.551 89.19l-8.284 4.983c-.506-2.306-.769-4.758-.627-7.423l7.84.418-.063 1.176 1.134.846z" fill="#fff"/><path opacity=".3" d="M358.941 100.779c2.369-.188 5.134-.434 5.134-.434l-11.988 19.881-4.481-.003M376.974 75.167l4.39.234 12.924-19.752-3.214-.172" fill="#D0C6B8"/><path d="M386.743 89.525a.55.55 0 10.058-1.098.55.55 0 00-.058 1.097zM387.121 92.768a.55.55 0 10.059-1.097.55.55 0 00-.059 1.097zM388.212 95.892a.549.549 0 10.06-1.096.549.549 0 00-.06 1.096zM387.237 86.17a.55.55 0 10.059-1.097.55.55 0 00-.059 1.097zM388.889 83.192a.55.55 0 10.058-1.098.55.55 0 00-.058 1.098z" fill="#777"/><path opacity=".6" d="M405.846 104.301l.31-13.192 1.176.063.851-1.213 4.119 11.227c-1.759 2.029-3.864 3.174-6.064 3.136l-.392-.021z" fill="#fff"/><path d="M406.29 82.704c-.826.743-3.912 5.531-.16 10.37 4.033-1.908 6.86-3.33 6.746-5.616-.031-2.36-6.586-4.754-6.586-4.754z" fill="#777"/><g filter="url(#filter0_d_3869_2050)"><path d="M229.895 151.928H56.614c-3.1 0-5.636-2.536-5.636-5.635V29.739c0-3.1 2.536-5.635 5.636-5.635h173.281c3.099 0 5.635 2.536 5.635 5.635v116.46c.094 3.193-2.442 5.729-5.635 5.729z" fill="url(#paint5_linear_3869_2050)"/></g><g filter="url(#filter1_d_3869_2050)"><path d="M60.358 72.293c0 5.204-4.208 9.413-9.458 9.413H19.313C13.023 81.842 8 76.82 8 70.755c0-6.11 5.068-11.178 11.45-10.861 5.475-17.106 30.545-14.707 32.672 2.987 4.706.588 8.236 4.57 8.236 9.412z" fill="url(#paint6_linear_3869_2050)"/></g><path d="M50.9 81.706c5.204 0 9.458-4.208 9.458-9.413 0-5.204-4.254-9.412-9.458-9.412-5.204 0-9.458 4.208-9.458 9.412 0 5.205 4.254 9.413 9.458 9.413z" fill="url(#paint7_linear_3869_2050)"/><path d="M35.469 81.797c9.277 0 16.834-7.512 16.834-16.79 0-9.276-7.557-16.788-16.834-16.788S18.635 55.73 18.635 65.008c0 9.277 7.512 16.789 16.834 16.789z" fill="url(#paint8_linear_3869_2050)"/><g filter="url(#filter2_d_3869_2050)"><path d="M333.368 121.066c0 4.646-3.757 8.404-8.444 8.404H296.721c-5.616.121-10.101-4.364-10.101-9.778 0-5.455 4.525-9.98 10.223-9.697 4.889-15.273 27.273-13.132 29.172 2.666 4.202.526 7.353 4.081 7.353 8.405z" fill="url(#paint9_linear_3869_2050)"/></g><path d="M324.924 129.47c4.646 0 8.444-3.758 8.444-8.404 0-4.647-3.798-8.404-8.444-8.404-4.647 0-8.445 3.757-8.445 8.404 0 4.646 3.798 8.404 8.445 8.404z" fill="url(#paint10_linear_3869_2050)"/><path d="M311.146 129.551c8.283 0 15.03-6.707 15.03-14.99 0-8.283-6.747-14.99-15.03-14.99s-15.031 6.707-15.031 14.99c0 8.283 6.707 14.99 15.031 14.99z" fill="url(#paint11_linear_3869_2050)"/><g filter="url(#filter3_d_3869_2050)"><path d="M393.079 145.914c0 3.333-2.695 6.028-6.057 6.028h-20.229c-4.029.087-7.246-3.13-7.246-7.014a6.959 6.959 0 017.333-6.955c3.506-10.955 19.562-9.419 20.924 1.913 3.014.376 5.275 2.927 5.275 6.028z" fill="url(#paint12_linear_3869_2050)"/></g><path d="M387.021 151.942c3.333 0 6.058-2.695 6.058-6.028 0-3.333-2.725-6.028-6.058-6.028-3.332 0-6.057 2.695-6.057 6.028 0 3.333 2.725 6.028 6.057 6.028z" fill="url(#paint13_linear_3869_2050)"/><path d="M377.139 152c5.941 0 10.781-4.811 10.781-10.752 0-5.941-4.84-10.752-10.781-10.752-5.941 0-10.781 4.811-10.781 10.752 0 5.941 4.811 10.752 10.781 10.752z" fill="url(#paint14_linear_3869_2050)"/><path d="M235.56 39.61H51.008V29.9c0-3.295 2.536-5.992 5.636-5.992h173.28c3.1 0 5.636 2.697 5.636 5.992v9.708z" fill="url(#paint15_linear_3869_2050)"/><path d="M59.905 34.631a2.724 2.724 0 100-5.447 2.724 2.724 0 000 5.447z" fill="#F4F1EA"/><path d="M68.17 34.631a2.724 2.724 0 100-5.447 2.724 2.724 0 000 5.447z" fill="#BAB7B4"/><path d="M76.435 34.631a2.724 2.724 0 100-5.447 2.724 2.724 0 000 5.447z" fill="#949390"/><defs><linearGradient id="paint0_linear_3869_2050" x1="313.948" y1="74.903" x2="320.14" y2="76.646" gradientUnits="userSpaceOnUse"><stop stop-color="#EFE5D8"/><stop offset="1" stop-color="#D0BDA6"/></linearGradient><linearGradient id="paint1_linear_3869_2050" x1="384.869" y1="56.859" x2="404.471" y2="62.341" gradientUnits="userSpaceOnUse"><stop stop-color="#EFE5D8"/><stop offset="1" stop-color="#D0BDA6"/></linearGradient><linearGradient id="paint2_linear_3869_2050" x1="338.99" y1="82.144" x2="366.144" y2="118.418" gradientUnits="userSpaceOnUse"><stop stop-color="#F0E8DD"/><stop offset="1" stop-color="#D8CABA"/></linearGradient><linearGradient id="paint3_linear_3869_2050" x1="305.241" y1="86.32" x2="314.028" y2="89.389" gradientUnits="userSpaceOnUse"><stop stop-color="#F6F0E6"/><stop offset="1" stop-color="#E2D7C9"/></linearGradient><linearGradient id="paint4_linear_3869_2050" x1="360.996" y1="101.004" x2="384.379" y2="108.672" gradientUnits="userSpaceOnUse"><stop stop-color="#F6F0E6"/><stop offset="1" stop-color="#E2D7C9"/></linearGradient><linearGradient id="paint5_linear_3869_2050" x1="143.195" y1="21.147" x2="143.195" y2="153.306" gradientUnits="userSpaceOnUse"><stop stop-color="#FDFEFF"/><stop offset=".996" stop-color="#F5F2EC"/></linearGradient><linearGradient id="paint6_linear_3869_2050" x1="34.179" y1="62.082" x2="34.162" y2="82.069" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#F0E8DD"/></linearGradient><linearGradient id="paint7_linear_3869_2050" x1="40.435" y1="60.854" x2="48.82" y2="70.024" gradientUnits="userSpaceOnUse"><stop stop-color="#D8CABA"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint8_linear_3869_2050" x1="17.058" y1="57.304" x2="30.273" y2="62.475" gradientUnits="userSpaceOnUse"><stop stop-color="#F0E8DD"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint9_linear_3869_2050" x1="309.994" y1="111.948" x2="309.979" y2="129.794" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#F0E8DD"/></linearGradient><linearGradient id="paint10_linear_3869_2050" x1="315.58" y1="110.852" x2="323.066" y2="119.04" gradientUnits="userSpaceOnUse"><stop stop-color="#D8CABA"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint11_linear_3869_2050" x1="294.708" y1="107.682" x2="306.506" y2="112.299" gradientUnits="userSpaceOnUse"><stop stop-color="#F0E8DD"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint12_linear_3869_2050" x1="376.313" y1="139.374" x2="376.302" y2="152.175" gradientUnits="userSpaceOnUse"><stop stop-color="#fff"/><stop offset=".996" stop-color="#F0E8DD"/></linearGradient><linearGradient id="paint13_linear_3869_2050" x1="380.32" y1="138.588" x2="385.689" y2="144.461" gradientUnits="userSpaceOnUse"><stop stop-color="#D8CABA"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint14_linear_3869_2050" x1="365.348" y1="136.314" x2="373.811" y2="139.626" gradientUnits="userSpaceOnUse"><stop stop-color="#F0E8DD"/><stop offset=".994" stop-color="#fff" stop-opacity="0"/></linearGradient><linearGradient id="paint15_linear_3869_2050" x1="228.239" y1="23.91" x2="56.921" y2="30.921" gradientUnits="userSpaceOnUse"><stop stop-color="#606060"/><stop offset="1" stop-color="#7B7B7B"/></linearGradient><filter id="filter0_d_3869_2050" x="44.978" y="20.104" width="196.554" height="139.824" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="2"/><feGaussianBlur stdDeviation="3"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.37 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2050"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2050" result="shape"/></filter><filter id="filter1_d_3869_2050" x="0" y="43.26" width="68.358" height="49.449" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="3"/><feGaussianBlur stdDeviation="4"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2050"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2050" result="shape"/></filter><filter id="filter2_d_3869_2050" x="278.62" y="94.607" width="62.748" height="45.865" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="3"/><feGaussianBlur stdDeviation="4"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2050"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2050" result="shape"/></filter><filter id="filter3_d_3869_2050" x="353.547" y="125.522" width="45.532" height="33.422" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB"><feFlood flood-opacity="0" result="BackgroundImageFix"/><feColorMatrix in="SourceAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/><feOffset dy="1"/><feGaussianBlur stdDeviation="3"/><feColorMatrix values="0 0 0 0 0.576471 0 0 0 0 0.501961 0 0 0 0 0.396078 0 0 0 0.27 0"/><feBlend in2="BackgroundImageFix" result="effect1_dropShadow_3869_2050"/><feBlend in="SourceGraphic" in2="effect1_dropShadow_3869_2050" result="shape"/></filter></defs></svg>`));
