+++ illustrations/src/index.ts
@@ -1,5 +1,6 @@
-export { AddUserDark, AddUserGroupDark, AddUserGroupLight, AddUserLight, AddUserProfileDark, AddUserProfileLight, AddUserToGroupDark, AddUserToGroupLight, AppointmentDark, AppointmentLight, BluetoothPermissionsDark, BluetoothPermissionsLight, BluetoothPermissonsErrorDark, BluetoothPermissonsErrorLight, CloudLocationDark, CloudLocationLight, CodeErrorDark, CodeErrorLight, CreateDataDark, CreateDataLight, DataDark, DataLight, DesktopDark, DesktopLight, DevicesDark, DevicesLight, DoorErrorDark, DoorErrorLight, EmailSentDark, EmailSentLight, EmptyFolderDark, EmptyFolderLight, EmptyGeneralDark, EmptyGeneralLight, EmptySpaceDark, EmptySpaceLight, FileMissingDark, FileMissingLight, FolderDark, FolderLight, GridLocationDark, GridLocationLight, LaunchDark, LaunchLight, LocationDark, LocationLight, LoginDark, LoginLight, MediaDark, MediaErrorDark } from './illustrations/chunk-0'
-export { MediaErrorLight, MediaLight, MilPermissionsDark, MilPermissionsLight, NoAccessDark, NoAccessLight, NoConnectionDark, NoConnectionLight, NoContentDark, NoContentLight, NoEmailsDark, NoEmailsLight, NoEventsDark, NoEventsLight, NoFriendsDark, NoFriendsLight, NoMatchDark, NoMatchGeneralDark, NoMatchGeneralLight, NoMatchLight, NoReportsDark, NoReportsLight, NoSitesDark, NoSitesLight, NotFoundDark, NotFoundLight, NotificationsDark, NotificationsLight, OrganizationDark, OrganizationLight, SettingsDark, SettingsLight, SuccessDark, SuccessLight, TeamDark, TeamLight, UnderConstructionDark, UnderConstructionLight, UnpackDark, UnpackLight, UserDark, UserGroupDark, UserGroupLight, UserLight, UserProfileDark, UserProfileLight, WelcomePlaneDark, WelcomePlaneLight } from './illustrations/chunk-1'
+export { AddUserDark, AddUserGroupDark, AddUserGroupLight, AddUserLight, AddUserProfileDark, AddUserProfileLight, AddUserToGroupDark, AddUserToGroupLight, AppointmentDark, AppointmentLight, BluetoothPermissionsDark, BluetoothPermissionsLight, BluetoothPermissonsErrorDark, BluetoothPermissonsErrorLight, CloudConnectDark, CloudConnectErrorDark, CloudConnectErrorLight, CloudConnectLight, CloudLocationDark, CloudLocationLight, CodeErrorDark, CodeErrorLight, CreateDataDark, CreateDataLight, DataDark, DataLight, DesktopDark, DesktopLight, DevicesDark, DevicesLight, DoorErrorDark, DoorErrorLight, EmailSentDark, EmailSentLight, EmptyFolderDark, EmptyFolderLight, EmptyGeneralDark, EmptyGeneralLight, EmptySpaceDark, EmptySpaceLight, FileMissingDark, FileMissingLight, FolderDark, FolderLight, GridLocationDark, GridLocationLight, LaunchDark, LaunchLight, LocationDark, LocationLight } from './illustrations/chunk-0'
+export { LoginDark, LoginLight, MediaDark, MediaErrorDark, MediaErrorLight, MediaLight, MilPermissionsDark, MilPermissionsLight, NoAccessDark, NoAccessLight, NoConnectionDark, NoConnectionLight, NoContentDark, NoContentLight, NoEmailsDark, NoEmailsLight, NoEventsDark, NoEventsLight, NoFriendsDark, NoFriendsLight, NoMatchDark, NoMatchGeneralDark, NoMatchGeneralLight, NoMatchLight, NoReportsDark, NoReportsLight, NoSitesDark, NoSitesLight, NotFoundDark, NotFoundLight, NotificationsDark, NotificationsLight, OrganizationDark, OrganizationLight, ProhibitedDark, ProhibitedLight, SettingsDark, SettingsLight, SuccessDark, SuccessLight, TeamDark, TeamLight, UnderConstructionDark, UnderConstructionLight, UnpackDark, UnpackLight, UserDark, UserGroupDark, UserGroupLight, UserLight } from './illustrations/chunk-1'
+export { UserProfileDark, UserProfileLight, WelcomePlaneDark, WelcomePlaneLight } from './illustrations/chunk-2'
 export type { AxisIllustrationProps } from './utils/types'
 export { bundleIllustration, bundleIllustrationSmart } from './utils/bundleIllustration'
 export type { TBundleIllustration, TBundleIllustrationVariant, TBundleIllustrationSmart} from './utils/bundleIllustration'
